### 2025-11-19 #9377 @kthyng
Thanks for the info @sampottinger. This has passed the scope query and I'll work on my other high level review items when I can.
