### 2025-11-19 dbfe6f2
Minor consistency fixes.
