### 2025-11-19 dc4a290
Minor consistency fixes.
