### 2025-12-01 62ab477
Fix truncation issue on save file button. (#670)
In the web-based IDE, fix an issue with truncation of the QTA file
downloaded through the save file button caused by incomplete encoding.
Thanks to Claude for identifying the problem.

Co-authored-by: Claude <noreply@anthropic.com>
