### 2025-12-01 a5427e6
Minor clarification in CONTRIBUTING.
