### 2025-12-01 42b6d60
Fix in tutorial 8.
Fix in tutorial 8 example code not being copied.
