### 2025-12-04 #9377 @sampottinger
Hello @kthyng! My apologies but, just in case it is helpful, I wanted to recommend some potential reviewers below from the JOSS reviewers list in case we are fortunate enough to move forward.

 - [Gregor Becker](https://reviewers.joss.theoj.org/reviewers/2752): `@GregorBecker`
 - [Andrew Bennett](https://reviewers.joss.theoj.org/reviewers/986): `@arbennett`
 - [Lionel Constantin](https://reviewers.joss.theoj.org/reviewers/4913): `@lionel42`
 - [Jack Leland](https://reviewers.joss.theoj.org/reviewers/17303): `@jackleland`
 - [Ali Ramadhan](https://reviewers.joss.theoj.org/reviewers/700): `@ali-ramadhan`
 - [Angela Rossi](https://reviewers.joss.theoj.org/reviewers/20208): `@rossiangela`
 - [Fred Shone](https://reviewers.joss.theoj.org/reviewers/3393): `@fredshone`

Thank you very much again for your consideration. Please let me know if there is anything else we can do!
