### 2025-12-05 fb47de8
Do not allow negative retirement.
