### 2025-12-05 9c26564
Do not allow negative retirement.
