### 2025-12-05 d85bb66
Merge pull request #674 from SchmidtDSE/bug/tollerance-exact-div
Ensure big decimal precision.
