### 2025-12-05 fe27625
Allow equations in UI input boxes.
We have gone back and forth on this a bit but allowing equations entered
from editor to appear in designer for now.

Co-authored-by: Claude <noreply@anthropic.com>
