### 2025-12-05 f885c03
Add operation fragment for power. (#676)
The operation fragment for power was accidentially left off on the
roadmap. This resolves that oversight. Our apologies! This impacted a
small number of users leveraging power in custom formulas like from UI but did not
impact engine internal calculations or most operations.
