### 2025-12-06 8bc9045
Correct #675.
Claude didn't quite get it right... Fixing the ability to use formulas
in the UI editor with interop to code editor. This is distinct to plain
old numbers.
