+++ editor/js/ui_editor_util.js
@@ -276,8 +276,10 @@ function setEngineNumberValue(valSelection, unitsSelection, source, defaultValue
     valSelection.value = valueOrDefault.getOriginalString();
   } else {
     const numericValue = valueOrDefault.getValue();
-    // Display empty string instead of NaN
-    valSelection.value = isNaN(numericValue) ? "" : numericValue;
+    const isStr = typeof numericValue === 'string' || numericValue instanceof String;
+    const isNan = isNaN(numericValue);
+    const useBlank = !isStr && isNan;
+    valSelection.value = useBlank ? "" : numericValue;
   }
   unitsSelection.value = valueOrDefault.getUnits();
 }
