### 2025-12-06 370a202
Lint fixes on #677.
