### 2025-12-07 1b24332
Slightly less noisy tests.
A few fixes to ensure tests are complete and also avoid noisy outputs,
cleaning up CI / CD.
