+++ engine/src/test/java/org/kigalisim/validate/RechargeLiveTests.java
@@ -50,6 +50,72 @@ public class RechargeLiveTests {
   }
 
   /**
+   * Ensure negative recharge clamped to zero.
+   */
+  @Test
+  public void testRechargeNegative() throws IOException {
+    String qtaPath = "../examples/recharge_negative.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    String scenarioName = "business as usual";
+    Stream<EngineResult> results = KigaliSimFacade.runScenario(program, scenarioName, progress -> {});
+    List<EngineResult> resultsList = results.collect(Collectors.toList());
+
+    // Check year 2 equipment (population) value
+    EngineResult resultYear2 = LiveTestsUtil.getResult(resultsList.stream(), 2, "test", "test");
+    assertNotNull(resultYear2, "Should have result for test/test in year 2");
+    assertEquals(200000.0, resultYear2.getPopulation().getValue().doubleValue(), 0.0001,
+        "Equipment should be 200000 units in year 2");
+    assertEquals("units", resultYear2.getPopulation().getUnits(),
+        "Equipment units should be units");
+  }
+
+  /**
+   * Ensure recharge of percent above 100 clamped to 100.
+   */
+  @Test
+  public void testRechargePositiveClamp() throws IOException {
+    String qtaPath = "../examples/recharge_positive_clamp.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    String scenarioName = "business as usual";
+    Stream<EngineResult> results = KigaliSimFacade.runScenario(program, scenarioName, progress -> {});
+    List<EngineResult> resultsList = results.collect(Collectors.toList());
+
+    // Check year 2 equipment (population) value
+    EngineResult resultYear2 = LiveTestsUtil.getResult(resultsList.stream(), 2, "test", "test");
+    assertNotNull(resultYear2, "Should have result for test/test in year 2");
+    assertEquals(100000.0, resultYear2.getPopulation().getValue().doubleValue(), 0.0001,
+        "Equipment should be 100000 units in year 2");
+    assertEquals("units", resultYear2.getPopulation().getUnits(),
+        "Equipment units should be units");
+  }
+
+  /**
+   * Ensure recharge of non-percent above 100 not clamped.
+   */
+  @Test
+  public void testRechargePositiveNoClamp() throws IOException {
+    String qtaPath = "../examples/recharge_positive_no_clamp.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    String scenarioName = "business as usual";
+    Stream<EngineResult> results = KigaliSimFacade.runScenario(program, scenarioName, progress -> {});
+    List<EngineResult> resultsList = results.collect(Collectors.toList());
+
+    // Check year 2 equipment (population) value
+    EngineResult resultYear2 = LiveTestsUtil.getResult(resultsList.stream(), 2, "test", "test");
+    assertNotNull(resultYear2, "Should have result for test/test in year 2");
+    assertEquals(190000.0, resultYear2.getPopulation().getValue().doubleValue(), 0.0001,
+        "Equipment should be 190000 units in year 2");
+    assertEquals("units", resultYear2.getPopulation().getUnits(),
+        "Equipment units should be units");
+  }
+
+  /**
    * Test for recharge import issue where recharge can go negative when sales exceed total demand.
    */
   @Test
