### 2025-12-07 4ecdcdc
Slightly less noisy tests. (#680)
* Minor intellij fix.

* Slightly less noisy tests.

A few fixes to ensure tests are complete and also avoid noisy outputs,
cleaning up CI / CD.
