+++ engine/src/main/java/org/kigalisim/engine/state/RechargeInformation.java
@@ -68,6 +68,21 @@ public class RechargeInformation {
    * @return A new RechargeInformation instance with accumulated values
    */
   public RechargeInformation add(EngineNumber newPopulation, EngineNumber newIntensity) {
+    boolean priorZero = population.getValue().equals(BigDecimal.ZERO);
+    if (priorZero) {
+      return new RechargeInformation(newPopulation, newIntensity);
+    }
+
+    boolean differentPopulationUnits = !population.getUnits().equals(newPopulation.getUnits());
+    if (differentPopulationUnits) {
+      throw new RuntimeException("Cannot mix units for recharge.");
+    }
+
+    boolean differentIntensityUnits = !intensity.getUnits().equals(newIntensity.getUnits());
+    if (differentIntensityUnits) {
+      throw new RuntimeException("Cannot mix units for recharge.");
+    }
+
     BigDecimal currentWeight = population.getValue().abs();
     BigDecimal addedWeight = newPopulation.getValue().abs();
 
@@ -81,8 +96,8 @@ public class RechargeInformation {
     BigDecimal accumulatedPopulation = population.getValue().add(newPopulation.getValue());
 
     return new RechargeInformation(
-        new EngineNumber(accumulatedPopulation, "%"),
-        new EngineNumber(weightedIntensity, "kg / unit")
+        new EngineNumber(accumulatedPopulation, newPopulation.getUnits()),
+        new EngineNumber(weightedIntensity, newIntensity.getUnits())
     );
   }
 
