### 2025-12-08 843f80c
Ensure last specified sales volume starts at 0.
