### 2025-12-08 9479e32
Style fix on #681.
