+++ engine/src/test/java/org/kigalisim/engine/state/StreamParameterizationTest.java
@@ -6,11 +6,15 @@
 
 package org.kigalisim.engine.state;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import java.math.BigDecimal;
 import org.junit.jupiter.api.Test;
 import org.kigalisim.engine.number.EngineNumber;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Tests for the StreamParameterization class.
