### 2025-12-08 c0b2812
Default no prior set sales values (#681)
* Minor intellij fix.

* Ensure last specified sales volume starts at 0.

* Style fix on #681.
