+++ engine/src/main/java/org/kigalisim/KigaliSimCommander.java
@@ -16,7 +16,7 @@ import picocli.CommandLine;
  *
  * @command kigalisim
  * @mixinStandardHelpOptions true
- * @version 0.1.1
+ * @version 0.1.3
  * @description "KigaliSim command line interface"
  * @subcommands { VersionCommand }
  */
