+++ engine/src/main/java/org/kigalisim/engine/state/OverridingConverterStateGetter.java
@@ -53,7 +53,7 @@ public class OverridingConverterStateGetter implements StateGetter {
    */
   public void setTotal(String streamName, EngineNumber value) {
     switch (streamName) {
-      case "sales", "domestic", "import" -> setVolume(value);
+      case "sales", "domestic", "import", "export" -> setVolume(value);
       case "equipment", "priorEquipment" -> setPopulation(value);
       case "consumption" -> setConsumption(value);
       default -> throw new IllegalArgumentException("Unrecognized stream name: " + streamName);
