### 2025-12-12 3f4c8ee
Add test for mixed units using assume keyword.
