### 2025-12-15 b3c79e1
Update tests to reflect new sales intent tracking behavior.
The StreamUpdateExecutor now tracks sales intent for all sales substreams
(domestic, import) regardless of whether they have equipment units. This
change allows combining domestic and import streams even when specified
in different unit types (kg vs units).

Test changes:
- testUpdateSalesCarryOverIgnoresVolumeBasedValues: Now expects kg values
  to create sales intent
- testUpdateSalesCarryOverWithMixedSpecifications: Now expects both unit-based
  and kg-based values to be combined in sales intent
- BasicLiveTests: Minor adjustment t
[…]
