### 2025-12-15 ca8dd4e
Add hypothesis tests for cap+change interaction issue.
Created two new test scenarios to investigate why testChangeCap fails:

1. testChangeCapDomestic - Tests capping domestic substream instead of sales
   - Result: PASSES ✅
   - QTA file: change_cap_test_domestic.qta

2. testChangeCapNonPercent - Tests absolute cap value instead of percentage
   - Result: PASSES ✅
   - QTA file: change_cap_test_nonpercent.qta

Findings:
- Capping a substream (domestic/import) works correctly with change statements
- Using absolute cap values works correctly with change statements
- Issue is specific to: perc
[…]
