### 2025-12-15 5d96455
Fix cap/floor compounding by updating component stream lastSpecified values.
Root cause: When capping/flooring "sales" with a percentage, the cap was applied
to lastSpecified["sales"], but the component streams (domestic, import) retained
their uncapped lastSpecified values. In subsequent years, change operations would
update the components, and updateUnitsTarget would recalculate sales from these
uncapped values, effectively erasing the previous year's cap.

Example flow (before fix):
- Year 2029: cap sales to 85% → sets lastSpecified["sales"] = 27.29 mt
- Year 2030: change domestic by 6% → u
[…]
