### 2025-12-15 ce7b717
Fix displacement not updating lastSpecified values in destination substance
When DisplaceExecutor displaces from one substance to another (e.g., HFC-134a
to R-600a), it was only updating the stream values but not the lastSpecified
values for domestic and import. This caused subsequent "change sales by X%"
commands to overwrite the displacement effects instead of compounding with them.

Root cause:
- LimitExecutor caps HFC-134a and updates its lastSpecified values (reduced)
- DisplaceExecutor adds mass to R-600a streams but leaves lastSpecified unchanged
- Next year: "change sales by 5%" applie
[…]
