### 2025-12-15 94d0741
Update test expectations and add policy order documentation
Test Changes:
- Rename testAccidentalDisplaceCheck to testPolicyOrderSensitivity
- Update test to verify that policy order DOES matter (not that it doesn't)
- Change expectation: Combined Reverse should be >10% higher than Combined
- Add detailed JavaDoc explaining why policy order affects outcomes

Documentation Changes:
- Tutorial 5: Add note explaining volume vs units-based displacement behavior
- Tutorial 6: Add brief note that policy order can matter, refer to Tutorial 7
- Tutorial 7: Add new "Understanding Policy Order" section 
[…]
