### 2025-12-15 9bbdf3e
Complete more comprehensive last specified value tracking.
