### 2025-12-15 08c3c69
Style fix on #685.
