### 2025-12-18 a005c80
Allow set to clear prior intent.
