### 2025-12-18 e690760
Allow set to clear prior intent.
