### 2025-12-18 1409456
Style fix on #687.
