### 2025-12-20 6036ea1
Improve flexibility for stream query.
Improve the flexiblity of ordering values in formulas so that get stream
can appear before or after. This works by adding an operation to remove
units (RemoveUnitsOperation) from something being used in a formula (along with
JointOperation to facilitate). This then allows the value to use the
units given / mandated by the user.

Claude fixed style issues but did not fundamentally author solution so
not listing as co-author.
