### 2025-12-23 7dafb04
Add failing test for sales vs substream cap discrepancy
When capping "sales" vs "domestic" separately with change statements,
the consumption values diverge significantly by year 10. This test
documents the bug where these should produce equivalent results.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
