+++ engine/src/main/java/org/kigalisim/engine/state/SimulationState.java
@@ -36,6 +36,7 @@ public class SimulationState {
 
   private final Map<String, StreamParameterization> substances;
   private final Map<String, EngineNumber> streams;
+  private final Map<String, EngineNumber> priorStreams;
   private final OverridingConverterStateGetter stateGetter;
   private final UnitConverter unitConverter;
   private int currentYear;
@@ -49,6 +50,7 @@ public class SimulationState {
   public SimulationState(OverridingConverterStateGetter stateGetter, UnitConverter unitConverter) {
     this.substances = new HashMap<>();
     this.streams = new HashMap<>();
+    this.priorStreams = new HashMap<>();
     this.stateGetter = stateGetter;
     this.unitConverter = unitConverter;
   }
@@ -351,21 +353,42 @@ public class SimulationState {
   }
 
   /**
-   * Get the value of a specific stream using key.
+   * Get the value of a specific stream using key from this year.
    *
    * @param useKey The key containing application and substance
    * @param name The stream name
    * @return The stream value
    */
   public EngineNumber getStream(UseKey useKey, String name) {
+    return getStream(useKey, name, false);
+  }
+
+  /**
+   * Get the value of a specific stream using key from this or prior year.
+   *
+   * @param useKey The key containing application and substance
+   * @param name The stream name
+   * @return The stream value
+   */
+  public EngineNumber getStream(UseKey useKey, String name, boolean priorYear) {
     ensureStreamKnown(name);
 
     return switch (name) {
-      case "sales" -> getStreamSales(useKey);
-      case "recycle" -> getStreamRecycle(useKey);
-      case "induction" -> getStreamInduction(useKey);
-      default -> getStreamDirect(useKey, name);
+      case "sales" -> getStreamSales(useKey, priorYear);
+      case "recycle" -> getStreamRecycle(useKey, priorYear);
+      case "induction" -> getStreamInduction(useKey, priorYear);
+      default -> getStreamDirect(useKey, name, priorYear);
     };
+  }
+
+  /**
+   * Get the sales stream value by summing domestic, import, and recycle streams from this year.
+   *
+   * @param useKey The key containing application and substance
+   * @return The total sales value in kg
+   */
+  private EngineNumber getStreamSales(UseKey useKey) {
+    return getStreamSales(useKey, false);
   }
 
   /**
@@ -374,10 +397,10 @@ public class SimulationState {
    * @param useKey The key containing application and substance
    * @return The total sales value in kg
    */
-  private EngineNumber getStreamSales(UseKey useKey) {
-    EngineNumber domesticAmountRaw = getStream(useKey, "domestic");
-    EngineNumber importAmountRaw = getStream(useKey, "import");
-    EngineNumber recycleAmountRaw = getStream(useKey, "recycle");
+  private EngineNumber getStreamSales(UseKey useKey, boolean priorYear) {
+    EngineNumber domesticAmountRaw = getStream(useKey, "domestic", priorYear);
+    EngineNumber importAmountRaw = getStream(useKey, "import", priorYear);
+    EngineNumber recycleAmountRaw = getStream(useKey, "recycle", priorYear);
 
     EngineNumber domesticAmount = unitConverter.convert(domesticAmountRaw, "kg");
     EngineNumber importAmount = unitConverter.convert(importAmountRaw, "kg");
@@ -393,14 +416,24 @@ public class SimulationState {
   }
 
   /**
-   * Get the recycle stream value by summing recycleRecharge and recycleEol streams.
+   * Get the recycle stream value by summing recycleRecharge and recycleEol streams in this year.
    *
    * @param useKey The key containing application and substance
    * @return The total recycle value in kg
[…]
