### 2025-12-23 51d4aa8
Clarify limit percentages relative to year prior.
