### 2025-12-23 98381ce
Clarify limit percentages relative to year prior.
