### 2025-12-26 e72013e
Add getPrior* methods to StateGetter for prior year values
