### 2025-12-26 872ed1f
Add percent variant support to UnitConverter with helper methods
