+++ engine/src/main/java/org/kigalisim/engine/number/UnitConverter.java
@@ -247,6 +247,93 @@ public class UnitConverter {
   }
 
   /**
+   * Check if a normalized unit string represents any percent variant.
+   *
+   * @param normalizedUnits The normalized unit string to check
+   * @return True if the unit is any percent variant, false otherwise
+   */
+  private boolean isPercentUnit(String normalizedUnits) {
+    return "%".equals(normalizedUnits) || "%prioryear".equals(normalizedUnits)
+        || "%currentyear".equals(normalizedUnits) || "%current".equals(normalizedUnits);
+  }
+
+  /**
+   * Check if a normalized unit string represents a prior year percentage.
+   *
+   * @param normalizedUnits The normalized unit string to check
+   * @return True if the unit is % prior year, false otherwise
+   */
+  private boolean getIsPriorPercent(String normalizedUnits) {
+    return "%prioryear".equals(normalizedUnits);
+  }
+
+  /**
+   * Get the target volume for percentage conversion based on percent unit type.
+   *
+   * @param normalizedUnits The normalized percent unit string
+   * @return The volume to use as base for percentage calculations
+   */
+  private EngineNumber getTargetVolume(String normalizedUnits) {
+    if (getIsPriorPercent(normalizedUnits)) {
+      return stateGetter.getPriorVolume();
+    } else {
+      return stateGetter.getVolume();
+    }
+  }
+
+  /**
+   * Get the target population for percentage conversion based on percent unit type.
+   *
+   * @param normalizedUnits The normalized percent unit string
+   * @return The population to use as base for percentage calculations
+   */
+  private EngineNumber getTargetPopulation(String normalizedUnits) {
+    if (getIsPriorPercent(normalizedUnits)) {
+      return stateGetter.getPriorPopulation();
+    } else {
+      return stateGetter.getPopulation();
+    }
+  }
+
+  /**
+   * Get the target GHG consumption for percentage conversion based on percent unit type.
+   *
+   * @param normalizedUnits The normalized percent unit string
+   * @return The GHG consumption to use as base for percentage calculations
+   */
+  private EngineNumber getTargetGhgConsumption(String normalizedUnits) {
+    if (getIsPriorPercent(normalizedUnits)) {
+      return stateGetter.getPriorGhgConsumption();
+    } else {
+      return stateGetter.getGhgConsumption();
+    }
+  }
+
+  /**
+   * Get the target energy consumption for percentage conversion based on percent unit type.
+   *
+   * @param normalizedUnits The normalized percent unit string
+   * @return The energy consumption to use as base for percentage calculations
+   */
+  private EngineNumber getTargetEnergyConsumption(String normalizedUnits) {
+    return stateGetter.getEnergyConsumption();
+  }
+
+  /**
+   * Get the target years elapsed for percentage conversion based on percent unit type.
+   *
+   * @param normalizedUnits The normalized percent unit string
+   * @return The years elapsed to use as base for percentage calculations
+   */
+  private EngineNumber getTargetYearsElapsed(String normalizedUnits) {
+    if (getIsPriorPercent(normalizedUnits)) {
+      return stateGetter.getPriorYearsElapsed();
+    } else {
+      return stateGetter.getYearsElapsed();
+    }
+  }
+
+  /**
    * Convert a number to new units.
    *
    * @param source The EngineNumber to convert
@@ -357,7 +444,7 @@ public class UnitConverter {
       case "kgCO2e" -> toKgCo2eConsumption(input);
       case "kwh" -> toEnergyConsumption(input);
[…]
