### 2025-12-26 679085e
Add calculatePercentageLimit helper to LimitExecutor for code deduplication
