### 2025-12-26 4efb22d
Enhance createUnitConverterWithTotal to set prior values for % prior year support
