+++ engine/src/main/java/org/kigalisim/engine/support/LimitExecutor.java
@@ -173,11 +173,6 @@ public class LimitExecutor {
    * <p>Example: Stream was 1000 mt in year 4. In year 5, "cap domestic to 85%"
    * will cap to 850 mt (85% of prior year's 1000 mt).</p>
    *
-   * <p><strong>Note on Architecture:</strong> This method retrieves prior year values
-   * directly via SimulationState rather than using UnitConverter. This is because
-   * UnitConverter's getPriorVolume() returns prior "sales" only, while limit operations
-   * may need prior values for other streams like "domestic" or "import".</p>
-   *
    * @param stream The stream name to cap
    * @param amount The percentage cap amount (e.g., 85 for 85%)
    * @param displaceTarget The target substance/stream for displacement, or null
@@ -188,15 +183,14 @@ public class LimitExecutor {
     EngineNumber currentValueRaw = engine.getStream(stream);
     EngineNumber currentValue = unitConverter.convert(currentValueRaw, "kg");
 
-    // Retrieve prior year value for this specific stream (e.g., domestic, import).
-    // UnitConverter's getPriorVolume() returns prior "sales" only, not prior of other streams.
     SimulationState simulationState = engine.getStreamKeeper();
     Scope scope = engine.getScope();
     EngineNumber lastSpecified = simulationState.getStream(scope, stream, true);
     boolean hasPrior = lastSpecified != null;
 
     if (hasPrior) {
-      EngineNumber newCappedValue = calculatePercentageLimit(lastSpecified, amount);
+      EngineNumber percentPriorYear = new EngineNumber(amount.getValue(), "% prior year");
+      EngineNumber newCappedValue = unitConverter.convert(percentPriorYear, lastSpecified.getUnits());
 
       EngineNumber currentInKg = unitConverter.convert(currentValueRaw, "kg");
       EngineNumber newCappedInKg = unitConverter.convert(newCappedValue, "kg");
@@ -305,11 +299,6 @@ public class LimitExecutor {
    * <p>Example: Stream was 1000 mt in year 4. In year 5, "floor domestic to 80%"
    * will floor to 800 mt (80% of prior year's 1000 mt).</p>
    *
-   * <p><strong>Note on Architecture:</strong> This method retrieves prior year values
-   * directly via SimulationState rather than using UnitConverter. This is because
-   * UnitConverter's getPriorVolume() returns prior "sales" only, while limit operations
-   * may need prior values for other streams like "domestic" or "import".</p>
-   *
    * @param stream The stream name to floor
    * @param amount The percentage floor amount (e.g., 80 for 80%)
    * @param displaceTarget The target substance/stream for displacement, or null
@@ -320,15 +309,14 @@ public class LimitExecutor {
     EngineNumber currentValueRaw = engine.getStream(stream);
     EngineNumber currentValue = unitConverter.convert(currentValueRaw, "kg");
 
-    // Retrieve prior year value for this specific stream (e.g., domestic, import).
-    // UnitConverter's getPriorVolume() returns prior "sales" only, not prior of other streams.
     SimulationState simulationState = engine.getStreamKeeper();
     Scope scope = engine.getScope();
     EngineNumber lastSpecified = simulationState.getStream(scope, stream, true);
     boolean hasPrior = lastSpecified != null;
 
     if (hasPrior) {
-      EngineNumber newFloorValue = calculatePercentageLimit(lastSpecified, amount);
+      EngineNumber percentPriorYear = new EngineNumber(amount.getValue(), "% prior year");
+      EngineNumber newFloorValue = unitConverter.convert(percentPriorYear, lastSpecified.getUnits());
 
       EngineNumber currentInKg = unitConverter.convert(currentValueRaw, "kg");
[…]
