### 2025-12-26 478c1d8
Add end-to-end TDD tests for % current unit across set, change, cap, and tutorial
