### 2025-12-26 9f41e5b
Add % prior year tests and fix ChangeExecutor for linear growth behavior
