+++ engine/src/test/java/org/kigalisim/validate/BasicLiveTests.java
@@ -1160,4 +1160,31 @@ public class BasicLiveTests {
     assertEquals("kg", year2Result.getDomestic().getUnits(),
         "Domestic units should be kg");
   }
+
+  /**
+   * Test set with % prior year.
+   * This validates that % prior year in set context calculates percentage relative to prior year values.
+   */
+  @Test
+  public void testSetPercentPriorYear() throws IOException {
+    String qtaPath = "../examples/set_percent_prior_year.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    String scenarioName = "business as usual";
+    Stream<EngineResult> results = KigaliSimFacade.runScenario(program, scenarioName, progress -> {});
+    List<EngineResult> resultsList = results.collect(Collectors.toList());
+
+    EngineResult year1Result = LiveTestsUtil.getResult(resultsList.stream(), 1, "test", "test");
+    assertNotNull(year1Result, "Should have result for test/test in year 1");
+    assertEquals(100.0, year1Result.getDomestic().getValue().doubleValue(), 0.0001,
+        "Year 1 domestic should be 100 kg");
+
+    EngineResult year2Result = LiveTestsUtil.getResult(resultsList.stream(), 2, "test", "test");
+    assertNotNull(year2Result, "Should have result for test/test in year 2");
+    assertEquals(110.0, year2Result.getDomestic().getValue().doubleValue(), 0.0001,
+        "Year 2 domestic should be 110 kg (110% of prior year with % prior year)");
+    assertEquals("kg", year2Result.getDomestic().getUnits(),
+        "Domestic units should be kg");
+  }
 }
