### 2025-12-26 97487e9
Bump version to 0.1.4, apply code quality improvements (JavaDoc, imports, char literals)
