### 2025-12-26 02aab2e
Add support for different forms of percent (#691)
Add support for different forms of percent such as % prior year and % current in addition to plain % which currently exists. Used some help from Claude given that much of this repeats existing patterns.

* Add % prior year and % current grammar support

* Add getPrior* methods to StateGetter for prior year values

* Add percent variant support to UnitConverter with helper methods

* Add calculatePercentageLimit helper to LimitExecutor for code deduplication

* Enhance createUnitConverterWithTotal to set prior values for % prior year support

* 
[…]
