+++ editor/js/ace-mode-qubectalk.js
@@ -11,9 +11,9 @@ ace.define("ace/mode/qubectalk", [
     const structureKeywords = "about|application|default|define|end|policy|simulations|start|" +
       "substance|uses|variables";
 
-    const commandKeywords = "across|as|assume|at|by|cap|change|charge|" +
+    const commandKeywords = "across|as|assume|at|by|cap|change|charge|current|" +
       "continued|during|enable|eol|floor|for|from|get|in|induction|initial|" +
-      "modify|no|of|only|recharge|recover|replace|replacement|retire|reuse|" +
+      "modify|no|of|only|prior|recharge|recover|replace|replacement|retire|reuse|" +
       "set|simulate|then|to|trials|using|with";
 
     const conditionalKeywords = "and|else|endif|if|or|xor";
