+++ engine/src/test/java/org/kigalisim/engine/number/UnitConverterTest.java
@@ -40,6 +40,10 @@ public class UnitConverterTest {
     lenient().when(mock.getPopulationChange(any(UnitConverter.class)))
         .thenReturn(new EngineNumber(0, "units"));
     lenient().when(mock.getEnergyConsumption()).thenReturn(new EngineNumber(0, "kwh"));
+    lenient().when(mock.getPriorVolume()).thenReturn(new EngineNumber(0, "kg"));
+    lenient().when(mock.getPriorPopulation()).thenReturn(new EngineNumber(0, "units"));
+    lenient().when(mock.getPriorGhgConsumption()).thenReturn(new EngineNumber(0, "tCO2e"));
+    lenient().when(mock.getPriorYearsElapsed()).thenReturn(new EngineNumber(0, "years"));
     return mock;
   }
 
@@ -651,5 +655,152 @@ public class UnitConverterTest {
 
     assertCloseTo(2, result.getValue(), 0.001);
     assertEquals("yrs", result.getUnits());
+  }
+
+  @Test
+  public void testPercentPriorYearToVolume() {
+    StateGetter mockStateGetter = createMockStateGetter();
+    lenient().when(mockStateGetter.getPriorVolume()).thenReturn(new EngineNumber(30, "kg"));
+
+    EngineNumber result = convertUnits(new EngineNumber(10, "% prior year"), "kg", mockStateGetter);
+
+    assertCloseTo(3, result.getValue(), 0.001);
+    assertEquals("kg", result.getUnits());
+  }
+
+  @Test
+  public void testPercentCurrentYearToVolume() {
+    StateGetter mockStateGetter = createMockStateGetter();
+    lenient().when(mockStateGetter.getVolume()).thenReturn(new EngineNumber(40, "kg"));
+
+    EngineNumber result = convertUnits(new EngineNumber(10, "% current year"), "kg", mockStateGetter);
+
+    assertCloseTo(4, result.getValue(), 0.001);
+    assertEquals("kg", result.getUnits());
+  }
+
+  @Test
+  public void testPercentCurrentToVolume() {
+    StateGetter mockStateGetter = createMockStateGetter();
+    lenient().when(mockStateGetter.getVolume()).thenReturn(new EngineNumber(40, "kg"));
+
+    EngineNumber result = convertUnits(new EngineNumber(10, "% current"), "kg", mockStateGetter);
+
+    assertCloseTo(4, result.getValue(), 0.001);
+    assertEquals("kg", result.getUnits());
+  }
+
+  @Test
+  public void testPercentPriorYearToPop() {
+    StateGetter mockStateGetter = createMockStateGetter();
+    lenient().when(mockStateGetter.getPriorPopulation()).thenReturn(new EngineNumber(50, "units"));
+
+    EngineNumber result = convertUnits(new EngineNumber(20, "% prior year"), "units", mockStateGetter);
+
+    assertCloseTo(10, result.getValue(), 0.001);
+    assertEquals("units", result.getUnits());
+  }
+
+  @Test
+  public void testPercentCurrentYearToPop() {
+    StateGetter mockStateGetter = createMockStateGetter();
+    lenient().when(mockStateGetter.getPopulation()).thenReturn(new EngineNumber(60, "units"));
+
+    EngineNumber result = convertUnits(new EngineNumber(20, "% current year"), "units", mockStateGetter);
+
+    assertCloseTo(12, result.getValue(), 0.001);
+    assertEquals("units", result.getUnits());
+  }
+
+  @Test
+  public void testPercentPriorYearToGhgConsumption() {
+    StateGetter mockStateGetter = createMockStateGetter();
+    lenient().when(mockStateGetter.getPriorGhgConsumption())
[…]

[…]
