### 2025-12-26 843a7a0
Revise beta changelog for post-beta updates
Updated the beta changelog to reflect the conclusion of the beta period and added new sections for major changes.
