### 2025-12-27 ab0311b
Fix lost spaces in ui translator for compound percents.
Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
