### 2025-12-27 01bf2d4
Add LimitDisplacementType enum and visitor methods for parsing
