+++ engine/src/main/java/org/kigalisim/lang/operation/FloorOperation.java
@@ -22,10 +22,23 @@ import org.kigalisim.lang.time.ParsedDuring;
  */
 public class FloorOperation implements Operation {
 
+  /**
+   * Enum representing the type of displacement.
+   */
+  public enum LimitDisplacementType {
+    /** Equivalent displacement - displaces based on context (volume if last specified as volume, units if units). */
+    EQUIVALENT,
+    /** By volume displacement - always convert to kg and displace by volume. */
+    BY_VOLUME,
+    /** By units displacement - always convert to units and displace by equipment count. */
+    BY_UNITS
+  }
+
   private final String stream;
   private final Operation valueOperation;
   private final Optional<String> displaceTarget;
   private final Optional<ParsedDuring> duringMaybe;
+  private final LimitDisplacementType displacementType;
 
   /**
    * Create a new FloorOperation that applies to all years.
@@ -38,6 +51,7 @@ public class FloorOperation implements Operation {
     this.valueOperation = valueOperation;
     this.displaceTarget = Optional.empty();
     duringMaybe = Optional.empty();
+    this.displacementType = LimitDisplacementType.EQUIVALENT;
   }
 
   /**
@@ -52,6 +66,24 @@ public class FloorOperation implements Operation {
     this.valueOperation = valueOperation;
     this.displaceTarget = Optional.ofNullable(displaceTarget);
     duringMaybe = Optional.empty();
+    this.displacementType = LimitDisplacementType.EQUIVALENT;
+  }
+
+  /**
+   * Create a new FloorOperation that applies to all years with displacement and displacement type.
+   *
+   * @param stream The name of the stream to floor.
+   * @param valueOperation The operation that calculates the minimum value.
+   * @param displaceTarget The name of the stream to displace excess to.
+   * @param displacementType The type of displacement to use.
+   */
+  public FloorOperation(String stream, Operation valueOperation, String displaceTarget,
+      LimitDisplacementType displacementType) {
+    this.stream = stream;
+    this.valueOperation = valueOperation;
+    this.displaceTarget = Optional.ofNullable(displaceTarget);
+    duringMaybe = Optional.empty();
+    this.displacementType = displacementType;
   }
 
   /**
@@ -66,6 +98,7 @@ public class FloorOperation implements Operation {
     this.valueOperation = valueOperation;
     this.displaceTarget = Optional.empty();
     duringMaybe = Optional.of(during);
+    this.displacementType = LimitDisplacementType.EQUIVALENT;
   }
 
   /**
@@ -82,6 +115,34 @@ public class FloorOperation implements Operation {
     this.valueOperation = valueOperation;
     this.displaceTarget = Optional.ofNullable(displaceTarget);
     duringMaybe = Optional.of(during);
+    this.displacementType = LimitDisplacementType.EQUIVALENT;
+  }
+
+  /**
+   * Create a new FloorOperation that applies to a specific time period with displacement and displacement type.
+   *
+   * @param stream The name of the stream to floor.
+   * @param valueOperation The operation that calculates the minimum value.
+   * @param displaceTarget The name of the stream to displace excess to.
+   * @param during The time period during which this operation applies.
+   * @param displacementType The type of displacement to use.
+   */
+  public FloorOperation(String stream, Operation valueOperation, String displaceTarget,
+      ParsedDuring during, LimitDisplacementType displacementType) {
+    this.stream = stream;
+    this.valueOperation = valueOperation;
+    this.displaceTarget = Optional.ofNullable(displaceTarget);
+    duringMaybe = Optional.of(during);
+    this.displacementType = displacementType;
+  }
+
+  /**
+   * Get the displacement type.
+   *
+   * @return The displacement type.
+   */
+  public LimitDisplacementType getDisplacementType() {
+    return displacementType;
   }
 
   /** {@inheritDoc} */
