### 2025-12-27 b62728c
Implement execution logic for explicit displacement types
