### 2025-12-27 077fd7f
Add documentation for displacement type options
