+++ engine/src/main/java/org/kigalisim/engine/support/LimitExecutor.java
@@ -37,6 +37,8 @@ import org.kigalisim.engine.recalc.StreamUpdateBuilder;
 import org.kigalisim.engine.state.Scope;
 import org.kigalisim.engine.state.SimulationState;
 import org.kigalisim.engine.state.YearMatcher;
+import org.kigalisim.lang.operation.CapOperation;
+import org.kigalisim.lang.operation.FloorOperation;
 
 /**
  * Executor for cap and floor limit operations on streams.
@@ -98,9 +100,10 @@ public class LimitExecutor {
    * @param amount The maximum value (percentage or absolute)
    * @param yearMatcher Matcher to determine if the cap applies to current year
    * @param displaceTarget Optional target for displacement (stream or substance), or null
+   * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
   public void executeCap(String stream, EngineNumber amount, YearMatcher yearMatcher,
-      String displaceTarget) {
+      String displaceTarget, CapOperation.LimitDisplacementType displacementType) {
     if (!EngineSupportUtils.getIsInRange(yearMatcher, engine.getYear())) {
       return;
     }
@@ -108,9 +111,9 @@ public class LimitExecutor {
     checkIsNotEquipment(stream);
 
     if ("%".equals(amount.getUnits())) {
-      capWithPercent(stream, amount, displaceTarget);
+      capWithPercent(stream, amount, displaceTarget, displacementType);
     } else {
-      capWithValue(stream, amount, displaceTarget);
+      capWithValue(stream, amount, displaceTarget, displacementType);
     }
   }
 
@@ -138,9 +141,10 @@ public class LimitExecutor {
    * @param amount The minimum value (percentage or absolute)
    * @param yearMatcher Matcher to determine if the floor applies to current year
    * @param displaceTarget Optional target for displacement (stream or substance), or null
+   * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
   public void executeFloor(String stream, EngineNumber amount, YearMatcher yearMatcher,
-      String displaceTarget) {
+      String displaceTarget, FloorOperation.LimitDisplacementType displacementType) {
     if (!EngineSupportUtils.getIsInRange(yearMatcher, engine.getYear())) {
       return;
     }
@@ -148,9 +152,9 @@ public class LimitExecutor {
     checkIsNotEquipment(stream);
 
     if ("%".equals(amount.getUnits())) {
-      floorWithPercent(stream, amount, displaceTarget);
+      floorWithPercent(stream, amount, displaceTarget, displacementType);
     } else {
-      floorWithValue(stream, amount, displaceTarget);
+      floorWithValue(stream, amount, displaceTarget, displacementType);
     }
   }
 
@@ -176,8 +180,10 @@ public class LimitExecutor {
    * @param stream The stream name to cap
    * @param amount The percentage cap amount (e.g., 85 for 85%)
    * @param displaceTarget The target substance/stream for displacement, or null
+   * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
-  private void capWithPercent(String stream, EngineNumber amount, String displaceTarget) {
+  private void capWithPercent(String stream, EngineNumber amount, String displaceTarget,
+      CapOperation.LimitDisplacementType displacementType) {
     UnitConverter unitConverter = EngineSupportUtils.createUnitConverterWithTotal(
         engine, stream);
     EngineNumber currentValueRaw = engine.getStream(stream);
@@ -211,7 +217,7 @@ public class LimitExecutor {
       if (displaceTarget != null) {
[…]
