### 2025-12-27 60c07b4
Add override options for displace  (#693)
Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>

* Add grammar for displacing by volume and displacing by units

* Add LimitDisplacementType enum and visitor methods for parsing

* Implement execution logic for explicit displacement types

* Add ChangeLiveTests integration tests for displacement types

* Add front-end UI support for displacement type dropdown

* Add documentation for displacement type options
