### 2025-12-27 89d5dda
Update beta changelog with displacement options
Added details about displacement options enhancement.
