+++ engine/src/main/java/org/kigalisim/engine/support/LimitExecutor.java
@@ -37,8 +37,7 @@ import org.kigalisim.engine.recalc.StreamUpdateBuilder;
 import org.kigalisim.engine.state.Scope;
 import org.kigalisim.engine.state.SimulationState;
 import org.kigalisim.engine.state.YearMatcher;
-import org.kigalisim.lang.operation.CapOperation;
-import org.kigalisim.lang.operation.FloorOperation;
+import org.kigalisim.lang.operation.DisplacementType;
 
 /**
  * Executor for cap and floor limit operations on streams.
@@ -103,7 +102,7 @@ public class LimitExecutor {
    * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
   public void executeCap(String stream, EngineNumber amount, YearMatcher yearMatcher,
-      String displaceTarget, CapOperation.LimitDisplacementType displacementType) {
+      String displaceTarget, DisplacementType displacementType) {
     if (!EngineSupportUtils.getIsInRange(yearMatcher, engine.getYear())) {
       return;
     }
@@ -144,7 +143,7 @@ public class LimitExecutor {
    * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
   public void executeFloor(String stream, EngineNumber amount, YearMatcher yearMatcher,
-      String displaceTarget, FloorOperation.LimitDisplacementType displacementType) {
+      String displaceTarget, DisplacementType displacementType) {
     if (!EngineSupportUtils.getIsInRange(yearMatcher, engine.getYear())) {
       return;
     }
@@ -183,7 +182,7 @@ public class LimitExecutor {
    * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
   private void capWithPercent(String stream, EngineNumber amount, String displaceTarget,
-      CapOperation.LimitDisplacementType displacementType) {
+      DisplacementType displacementType) {
     UnitConverter unitConverter = EngineSupportUtils.createUnitConverterWithTotal(
         engine, stream);
     EngineNumber currentValueRaw = engine.getStream(stream);
@@ -260,7 +259,7 @@ public class LimitExecutor {
    * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
   private void capWithValue(String stream, EngineNumber amount, String displaceTarget,
-      CapOperation.LimitDisplacementType displacementType) {
+      DisplacementType displacementType) {
     UnitConverter unitConverter = EngineSupportUtils.createUnitConverterWithTotal(
         engine, stream);
     EngineNumber currentValueRaw = engine.getStream(stream);
@@ -313,10 +312,7 @@ public class LimitExecutor {
    * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
   private void floorWithPercent(String stream, EngineNumber amount, String displaceTarget,
-      FloorOperation.LimitDisplacementType displacementTypeFloor) {
-    // Convert FloorOperation.LimitDisplacementType to CapOperation.LimitDisplacementType
-    CapOperation.LimitDisplacementType displacementType =
-        convertFloorToCapDisplacementType(displacementTypeFloor);
+      DisplacementType displacementType) {
     UnitConverter unitConverter = EngineSupportUtils.createUnitConverterWithTotal(
         engine, stream);
     EngineNumber currentValueRaw = engine.getStream(stream);
@@ -393,10 +389,7 @@ public class LimitExecutor {
    * @param displacementType The type of displacement (EQUIVALENT, BY_VOLUME, or BY_UNITS)
    */
[…]
