+++ engine/src/main/java/org/kigalisim/lang/operation/FloorDisplacingOperation.java
@@ -75,8 +75,7 @@ public class FloorDisplacingOperation implements Operation {
     YearMatcher yearMatcher = parsedDuring.buildYearMatcher(machine);
 
     Engine engine = machine.getEngine();
-    // Use EQUIVALENT for legacy displacing operations (backward compatibility)
     engine.floor(stream, result, yearMatcher, displaceTarget,
-        FloorOperation.LimitDisplacementType.EQUIVALENT);
+        DisplacementType.EQUIVALENT);
   }
 }

[…]
