### 2025-12-29 b27bd89
Style fixes for 20251229 release.
Style fixes and other code cleanup for release on 2025-12-29. Note
Claude did participate but in executing mostly mechanical changes
described in detail following existing patterns.

Co-Authored-By: Claude <noreply@anthropic.com>
