+++ engine/src/main/java/org/kigalisim/lang/operation/CapDisplacingOperation.java
@@ -75,8 +75,7 @@ public class CapDisplacingOperation implements Operation {
     YearMatcher yearMatcher = parsedDuring.buildYearMatcher(machine);
 
     Engine engine = machine.getEngine();
-    // Use EQUIVALENT for legacy displacing operations (backward compatibility)
     engine.cap(stream, result, yearMatcher, displaceTarget,
-        CapOperation.LimitDisplacementType.EQUIVALENT);
+        DisplacementType.EQUIVALENT);
   }
 }
