### 2025-12-29 f6795e9
Refactor key parsing in incrementYear loops.
Parse the key once in the loop and pass UseKey to private methods
instead of parsing the key string repeatedly in each method.

Co-Authored-By: Claude <noreply@anthropic.com>
