+++ humans.txt
@@ -31,10 +31,13 @@ Douglas McCauley
 Irene Papst
 Claudia Paratori (National Ozone Unit, Ministry of Environment, Chile)
 Alejandro Ramirez-Pabon
+Sara Alcantara Rodriguez (SERMINAT)
 Angelica Monserrat Jara Rios (Ministerio del Medio Ambiente)
 Frederico San Martini
+National Ozone Office, DGIELGCA/SEMARNAT (Mexico)
 Suzanne Spencer
 K V Varadharajan (SRF Limited India)
+Fery Wihastoro (Ministry of Environment, Republic of Indonesia)
 Elina Yuen
 
 And over a dozen countries and supporting organizations during beta testing.
