+++ humans.txt
@@ -19,6 +19,7 @@ Org: DSE / UC Berkeley
 
 /* OTHER CONTRIBUTORS */
 
+Ashraful Ambia (United Nations Development Programme)
 Amanda Anderson-You
 Kevin Basoa (Ministry of Environment, Chile)
 Tina Birmpili
