+++ paper/paper.bib
@@ -455,3 +455,49 @@ url="https://doi.org/10.1007/979-8-8688-0351-2_12"
   author={Brewer, Cynthia and Harrower, Mark and Sheesley, Ben and Woodruff, Andy and Heyman, David},
   year={2013}
 }
+
+@misc{credits,
+  title={Kigali Sim Credits},
+  author={{Kigali Sim Community}},
+  year={2026},
+  url={https://kigalisim.org/guide/credits.html},
+  note={Community contributors and acknowledgments}
+}
+
+@misc{oweg,
+  title={HFC Modelling Agenda for the 46th Meeting of the Open-Ended Working Group},
+  author={Natarajan, Balaji and Pottinger, Sam and de Bruyn, Magali and Birmpili, Tina},
+  year={2025},
+  publisher={Multilateral Fund Secretariat},
+  url={https://ozone.unep.org/system/files/documents/MLF_HFC_ModellingAgenda.pdf},
+  note={Open-Ended Working Group of the Parties to the Montreal Protocol}
+}
+
+@misc{excom,
+  title={Report of the Ninety-Seventh Meeting of the Executive Committee},
+  author={{Executive Committee of the Multilateral Fund}},
+  year={2025},
+  publisher={United Nations Environment Programme},
+  url={https://www.multilateralfund.org/api/drupal-documents/download/file/8f132cb8-e84c-418e-88c2-e55d9798a33d?filename=9702.pdf},
+  note={Document 9702}
+}
+
+@misc{seasianetwork,
+  title={List of Participants: South East Asia and the Pacific Ozone Officers Network Meeting},
+  author={{South East Asia and the Pacific Ozone Officers Network}},
+  year={2025},
+  month={December},
+  publisher={UNEP OzonAction},
+  url={https://admin.ozonactionmeetings.org/system/files/2025-12/List%20of%20participants_0.pdf?token=EPR0bhi2ZJMA6r5XChspmPU174-eovkf7b6lKbdJHHg&timestamp=1766625365},
+  note={Regional network meeting}
+}
+
+@misc{pacificnetwork,
+  title={Final Agenda: Pacific Island Countries Ozone Officers Network Meeting},
+  author={{Pacific Island Countries Ozone Officers Network}},
+  year={2025},
+  month={September},
+  publisher={UNEP OzonAction},
+  url={https://admin.ozonactionmeetings.org/system/files/2025-09/Final%20Agenda_Network%20Meeting%20Fiji%20Sep%2025.pdf?token=-4AgUCZDrrxipNhhPHCSk_yjRE3xPsrZkDHS6vQOPZo&timestamp=1767648637},
+  note={Fiji regional network meeting}
+}
