+++ paper/paper.md
@@ -37,19 +37,22 @@ Signed by all UN member states, the Montreal Protocol successfully phased out 99
 
 Kigali Sim provides the first open source reusable full lifecycle toolkit addressing Montreal Protocol-controlled substances. Consulting over a dozen countries and supporting organizations, a community of on-the-ground scientists, analysts, and policy-makers inform its flexible design. With either code or UI-based editing, it simulates emissions, energy, substance consumption, equipment populations, trade, and policy either in browser or JVM while following treaty conventions [@mlf_guidelines].
 
-# Design
-Kigali Sim serves Article 5 nations [@article5], Implementing Agencies, analysts, researchers, project implementers, and other organizations. Its dual-interface design supports domain experts in atmospheric science and environmental policy with varied programming expertise.
+# Research Impact Statement
+In addition to official participation in international policy-relevant treaty meetings [@oweg; @excom; @pacificnetwork; @seasianetwork] and extended co-design with over a dozen nations and supporting organizations, multiple governments publicly acknowledge use of the software [@credits].
+
+# Software Design
+Kigali Sim serves Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations. Its dual-interface design supports domain experts in atmospheric science and environmental policy with varied programming expertise.
 
 ## Simulation engine
-Some countries and supporting organizations work from trade records and industry surveys while others from bank estimates or observed emissions. The engine propagates user-provided values through substance flows and lifecycles to estimate the unmeasured quantities. It then layers complex policy interventions such as permitting and recycling on top of the "business as usual" scenario.
+Some countries and supporting organizations work from trade records and others from industry census or observed emissions. As opposed to a unidirectional structure assuming the location of observation within systems, the engine propagates user-provided values through substance flows and lifecycles to estimate the unmeasured quantities. It then layers complex policy interventions such as permitting and recycling on top of the "business as usual" scenario.
 
 ![Diagram showing data flow through Kigali Sim simulation engine from input data (trade records, bank estimates, or equipment surveys) through stock and flow calculations to output metrics.\label{fig:architecture}](KigaliEngine.svg){width="100%"}
 
 ## Dual-Interface Design
-We find many Kigali Sim users do not identify as programmers and, reflecting empirical findings that domain experts with "limited programming knowledge" benefit from domain-specific languages [@dsl_evaluation], our software progresses from a UI-based editor to DSL code-based authoring. In all cases, simulations execute locally, ensuring privacy.
+We find many Kigali Sim users do not identify as programmers and, reflecting empirical findings that domain experts with "limited programming knowledge" benefit from domain-specific languages [@dsl_evaluation], our software progresses from a UI-based editor to DSL code-based authoring instead of a general purpose language. In all cases, simulations execute locally, ensuring privacy.
 
 ### UI-based authoring
-The UI-based point-and-click editor familiarizes the user with Kigali Sim's concepts, creating a loop-based design [@loop_based_design; @core_loops] where small GUI-based changes automatically translate to code run for immediate feedback. Note that the web interface first progressively exposes functionality through sequenced disclosure [@hayashida_structure; @hayashida_video; @pyafscgap] as an on-ramp into a more open design [@open_world_zelda; @plastics].
[…]
