### 2026-01-05 63e5566
Additional edits to comply with new JOSS format.
