+++ paper/paper.bib
@@ -501,3 +501,13 @@ url="https://doi.org/10.1007/979-8-8688-0351-2_12"
   url={https://admin.ozonactionmeetings.org/system/files/2025-09/Final%20Agenda_Network%20Meeting%20Fiji%20Sep%2025.pdf?token=-4AgUCZDrrxipNhhPHCSk_yjRE3xPsrZkDHS6vQOPZo&timestamp=1767648637},
   note={Fiji regional network meeting}
 }
+
+@misc{nbc,
+  title={UC Berkeley offers tech to United Nations to fight global pollution},
+  author={Shaban, Bigad},
+  year={2025},
+  month={December},
+  publisher={NBC Bay Area},
+  url={https://www.nbcbayarea.com/news/local/uc-berkeley-tech-united-nations-pollution/3994242/},
+  note={Investigative Unit}
+}
