### 2026-01-05 43e932a
Additional edits to comply with new JOSS format.
