### 2026-01-05 15c40c0
Minor wording fix.
