### 2026-01-05 071ce15
More minor wording tweaks.
