+++ paper/paper.md
@@ -33,12 +33,12 @@ bibliography: paper.bib
 With active international use, Kigali Sim offers reusable stock and flow modeling of Montreal Protocol-controlled substances. Focusing on the Kigali Amendment which reduces hydrofluorocarbons that contribute to climate change through warming potentials up to thousands of times greater than carbon dioxide, this open source platform aids both atmospheric science and policy development. Supporting a diverse community with heterogeneous programming expertise, this parallelized toolkit interoperates visual no-code rapid model development and a domain-specific language (DSL), democratizing advanced computational tools. Through WebAssembly (WASM) or Java Virtual Machine (JVM), Kigali Sim offers portable, private, and rigorous simulation in support of history's most successful international environmental treaty.
 
 # Statement of need
-Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, related research and policy analysis involves considering complex economic, technological, and policy interactions [@complex], typically through closed source ad-hoc models. National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo] and the current lack of an open source domain-oriented toolkit for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
+Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, related research and policy analysis involves considering complex economic, technological, and policy interactions [@complex], typically through closed source ad-hoc models. National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo] and the current lack of open source domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
 
-Therefore, Kigali Sim provides the first open reusable lifecycle model focused on Montreal Protocol-controlled substances. Co-created with an international community, on-the-ground scientists, analysts, and policy-makers inform its flexible design. With either code or UI-based editing, it simulates emissions, energy, substance consumption, equipment populations, trade, and policy either in browser or JVM while following treaty conventions [@mlf_guidelines].
+Therefore, Kigali Sim provides the first open reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. Co-created with an international community, on-the-ground scientists, analysts, and policy-makers inform its flexible design. With either code or UI-based editing, it simulates emissions, energy, substance consumption, equipment populations, trade, and policy either in browser or JVM while following treaty conventions [@mlf_guidelines].
 
 # Research Impact Statement
-Participating in official international policy-relevant treaty meetings [@oweg; @excom; @pacificnetwork; @seasianetwork] and more than a year of co-design with over a dozen nations and supporting organizations, multiple governments publicly acknowledge use of the software [@credits] which has received news media coverage [@nbc].
+Participating in official international policy-relevant treaty meetings [@oweg; @excom; @pacificnetwork; @seasianetwork] with more than a year of co-design with over a dozen nations and supporting organizations, multiple governments publicly acknowledge use of the software [@credits] which has received news media coverage [@nbc].
 
 # Software Design
 Kigali Sim serves Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations. Its dual-interface design supports domain experts in atmospheric science and environmental policy with varied programming expertise.
