+++ paper/paper.md
@@ -35,13 +35,13 @@ With active international use, Kigali Sim offers reusable stock and flow modelin
 # Statement of need
 Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, related research and policy analysis involves considering complex economic, technological, and policy interactions [@complex], typically through closed source ad-hoc models. National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo] and the current lack of open source domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
 
-Therefore, Kigali Sim provides the first open reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. Co-created with an international community, on-the-ground scientists, analysts, and policy-makers inform its flexible design. With either code or UI-based editing, it simulates emissions, energy, substance consumption, equipment populations, trade, and policy either in browser or JVM while following treaty conventions [@mlf_guidelines].
+Therefore, serving Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations, Kigali Sim provides the first open reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. Co-created with an international community, on-the-ground scientists, analysts, and policy-makers inform its flexible design. With either code or UI-based editing, it simulates emissions, energy, substance consumption, equipment populations, trade, and policy either in browser or JVM while following treaty conventions [@mlf_guidelines].
 
 # Research Impact Statement
 Participating in official international policy-relevant treaty meetings [@oweg; @excom; @pacificnetwork; @seasianetwork] with more than a year of co-design with over a dozen nations and supporting organizations, multiple governments publicly acknowledge use of the software [@credits] which has received news media coverage [@nbc].
 
 # Software Design
-Kigali Sim serves Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations. Its dual-interface design supports domain experts in atmospheric science and environmental policy with varied programming expertise.
+Kigali Sim's engine supports domain experts in atmospheric science and environmental policy with varied programming expertise through a dual-interface design.
 
 ## Simulation engine
 Countries and supporting organizations work with varied data from trade records to industry censuses to observed emissions. Therefore, as opposed to a unidirectional structure, this engine propagates user-provided values from any location in the system through substance flows and lifecycles to estimate the unmeasured quantities. It then layers complex policy interventions such as permitting and recycling on top of the "business as usual" scenario.
