### 2026-01-06 c8e519d
Minor rewording.
