### 2026-01-06 0f0d736
Minor reword.
