### 2026-01-06 03af124
More trimming.
