### 2026-01-06 8c4b4a5
Use Kigali Sim name in research impact statement.
