### 2026-01-06 d499579
More trimming!
