+++ paper/paper.md
@@ -36,7 +36,7 @@ With active international use, Kigali Sim offers reusable stock and flow modelin
 Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, related research and policy analysis involves considering complex economic, technological, and policy interactions [@complex], typically through closed source ad-hoc models. National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo] and the current lack of open domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses]. Therefore, serving Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations, Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. With either code or UI-based editing, it simulates emissions, energy, substance consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines].
 
 # Research Impact Statement
-More than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year. On-the-ground scientists, analysts, and policy-makers inform its flexible structure. Participating in international treaty-related meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge working with the community project [@credits], open software which also recieved media coverage [@nbc].
+More than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year. On-the-ground scientists, analysts, and policy-makers inform its flexible structure. Participating in international treaty-related meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge working with the community project [@credits], open software which also received media coverage [@nbc].
 
 # Software Design
 Kigali Sim's engine supports domain experts in atmospheric science and environmental policy with varied programming expertise through a dual-interface design.
