### 2026-01-06 6965ae8
More trimming.
