### 2026-01-06 ce1a524
Some further trimming.
