+++ paper/paper.md
@@ -32,26 +32,25 @@ bibliography: paper.bib
 # Summary
 With active international use, Kigali Sim offers reusable stock and flow modeling of Montreal Protocol-controlled substances. Focusing on the Kigali Amendment which reduces hydrofluorocarbons that contribute to climate change through warming potentials up to thousands of times greater than carbon dioxide, this open source platform aids both atmospheric science and policy development. Supporting a diverse community with heterogeneous programming expertise, this parallelized toolkit interoperates visual no-code rapid model development and a domain-specific language (DSL), democratizing advanced computational tools. Through WebAssembly (WASM) or Java Virtual Machine (JVM), Kigali Sim offers portable, private, and rigorous simulation in support of history's most successful international environmental treaty.
 
-# Statement of need
-Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute].
+# Statement of Need
+Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, research and policy analysis for these HFCs involves modeling complex economic, technological, energy, and policy interactions [@complex].
 
-However, related research and policy analysis involves considering complex economic, technological, and policy interactions [@complex], typically through closed source ad-hoc models. National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo] and the current lack of open domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
+## State of field
+National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo]. Requiring closed source [@hfcoutlook] or ad-hoc models [@adhoc], the current lack of open domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
 
-Therefore, serving Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations, Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. With either code or UI-based editing, it simulates emissions, energy, substance consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines].
-
-# Research Impact Statement
-More than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year. On-the-ground scientists, analysts, and policy-makers inform its flexible structure. Participating in international treaty-related meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge working with the community project [@credits], open software which also received media coverage [@nbc].
+## Research impact statement
+Serving Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations, Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances for emissions, energy, substance consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines]. More than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year, spanning scientists, analysts, and policy-makers. Participating in many international meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge using and contributing to this community project [@credits] which also received media coverage [@nbc].
 
 # Software Design
[…]
