+++ paper/paper.bib
@@ -511,3 +511,23 @@ url="https://doi.org/10.1007/979-8-8688-0351-2_12"
   url={https://www.nbcbayarea.com/news/local/uc-berkeley-tech-united-nations-pollution/3994242/},
   note={Investigative Unit}
 }
+
+@misc{hfcoutlook,
+  title={HFC Outlook: Description and Specification},
+  author={{Gluckman Consulting}},
+  year={2023},
+  month={November},
+  url={https://www.gluckmanconsulting.com/wp-content/uploads/2023/11/HFC-Outlook-Description-and-Specification.pdf},
+  note={HFC modeling tool specification}
+}
+
+@misc{adhoc,
+  title={Preparation of Kigali HFC Implementation Plans},
+  author={{Multilateral Fund Secretariat}},
+  year={2024},
+  month={July},
+  day={8},
+  howpublished={Presentation at 46th Meeting of the Open-Ended Working Group of the Parties to the Montreal Protocol},
+  url={https://ozone.unep.org/system/files/documents/OEWG\%20KIP\%20Side\%20Event\%20MLFS_070824.pdf},
+  note={Side event presentation discussing ad-hoc modeling approaches}
+}
