+++ paper/paper.md
@@ -61,7 +61,7 @@ To support those new to programming, the UI-based point-and-click editor acclima
 We find some simulations' complexity cumbersome in UI-based authoring. Therefore, our QubecTalk domain-specific language facilitates expression of complex models in human-readable syntax inspired by but distinct to HyperTalk [@smallhypertalk]. Mirrored by the UI-editor, QubecTalk speaks in treaty terminology: directly translating terms of art like `cap import to 85% during years 2025 to 2035` into simulations. This also supports uncertainty quantification, conditional logic, and policy stacking. With optional AI assistant compatibility via the llms.txt specification [@llmstxt], users may author scripts in a web-based programming portal [@portal] or outside with direct JVM invocation.
 
 ## Limitations and future work
-Leaving the following for future work, Kigali Sim can model energy consumption but only estimates direct emissions^[Users may calculate indirect emissions using exported and energy mix data [@energy_mix].]. Additionally, it applies treaty trade attribution but will only attribute charge prior to equipment sale entirely to either the importer or exporter^[Local assembly can be modeled as domestic production.]. However, it will receive updates as official guidance changes in the future [@mlf_guidelines].
+Leaving the following for future work, Kigali Sim can model energy consumption but only estimates direct emissions^[Users may calculate indirect emissions using outside energy mix data [@energy_mix].]. Additionally, it applies treaty trade attribution but will only attribute charge prior to equipment sale entirely to either the importer or exporter^[Local assembly can be modeled as domestic production.]. However, it will receive updates as official guidance changes in the future [@mlf_guidelines].
 
 # Implementation
 Migrated from an original JavaScript implementation for performance and portability, Kigali Sim runs browser-based via WASM or JVM [@teavm]. Without transmitting simulations to external servers, both enable privacy-preserving zero-installation parallel local computation with BigDecimal [@bigdecimal] after ANTLR [@antlr] interprets QubecTalk code.
