### 2026-01-06 c32dc72
Consolidate duplicate citation.
