+++ paper/paper.md
@@ -36,7 +36,7 @@ With active international use, Kigali Sim offers reusable stock and flow modelin
 Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, research and policy analysis for these HFCs involves modeling complex economic, technological, energy, and policy interactions [@complex].
 
 ## State of field
-National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo]. Requiring closed source [@hfcoutlook] or ad-hoc models [@adhoc], the current lack of open domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
+National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo]. Requiring closed source [@hfcoutlook] or ad-hoc models [@complex], the current lack of open domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
 
 ## Research impact statement
 Serving Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations, Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances for emissions, energy, substance consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines]. More than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year, spanning scientists, analysts, and policy-makers. Participating in many international meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge using and contributing to this community project [@credits] which also received media coverage [@nbc].
