### 2026-01-06 b91a403
More trimming.
