+++ paper/paper.md
@@ -36,7 +36,7 @@ With active international use, Kigali Sim offers reusable stock and flow modelin
 Signed by all UN member states, the Montreal Protocol successfully phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, research and policy analysis for these HFCs involves modeling complex economic, technological, energy, and policy interactions [@complex].
 
 ## State of field
-National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo]. Requiring closed source [@hfcoutlook] or ad-hoc models [@complex], the current lack of open domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
+National Ozone Units (NOUs) and others who possess deep domain knowledge may work with limited resources [@noo]. Currently requiring closed source [@hfcoutlook] or ad-hoc models [@complex], the lack of open domain-oriented tools for holistic stock and flow simulation may inhibit transparent accessible modeling [@oses].
 
 ## Research impact statement
 Serving Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations, Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances for emissions, energy, substance consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines]. More than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year, spanning scientists, analysts, and policy-makers. Participating in many international meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge using and contributing to this community project [@credits] which also received media coverage [@nbc].
@@ -45,7 +45,7 @@ Serving Article 5 nations [@article5], Implementing Agencies, analysts, research
 Kigali Sim's engine supports domain experts in atmospheric science and environmental policy with varied programming expertise through a dual-interface design.
 
 ## Flexible engine
-Countries and supporting organizations work with varied information from trade records to industry census data to observed emissions. As diverse methodologies measure stocks, Kigali Sim pushes information from known to unknown parts of the system, providing automated unit conversions which may depend on equipment properties. As opposed to a unidirectional structure with a single entry-point, this propagates limited user-provided values through substance flows and lifecycles to estimate unmeasured quantities. It then layers complex policy interventions such as permitting and recycling on top of the triangulated "business as usual" scenario.
+Countries and supporting organizations work with varied information from trade records to industry census data to observed emissions. As diverse methodologies measure stocks, Kigali Sim pushes information from known to unknown parts of the system, providing automated unit conversions dependent on equipment properties. As opposed to a unidirectional structure with a single entry-point, this propagates limited user-provided values through substance flows and lifecycles to estimate unmeasured quantities. It then layers complex policy interventions such as permitting and recycling on top of the triangulated "business as usual" scenario.
 
 ![Diagram showing data flow through Kigali Sim simulation engine from input data (trade records, bank estimates, or equipment surveys) through stock and flow calculations to output metrics.\label{fig:architecture}](KigaliEngine.svg){width="100%"}
 
@@ -53,18 +53,18 @@ Countries and supporting organizations work with varied information from trade r
[…]
