### 2026-01-06 401da37
More trimming.
