### 2026-01-06 20557a5
More trimming.
