### 2026-01-06 d54fa99
More trimming.
