### 2026-01-06 f27c530
Minor trimming.
