### 2026-01-20 b2e3cd3
Add Tutorial 3a AI-assisted version and update index
Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.com>
