### 2026-01-20 5017c89
Add Tutorial 4a AI-assisted version and update index
Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.com>
