### 2026-01-20 5b372ae
Add Tutorial 6a AI-assisted version and update index
Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.com>
