### 2026-01-20 ea5994e
Add alphabetical sorting to UI lists and update changelog
Sort applications, substances, policies, and simulations alphabetically in UI-based editor lists and checkboxes. Policy selections in simulation stanzas are now also sorted alphabetically in generated code.

Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.com>
