### 2026-01-20 a546b8d
Update tutorials to mention ordering.
