### 2026-01-20 9b1d6a2
Update tutorials to mention ordering.
