### 2026-01-20 38cc9a9
Wrap AI prompts at 80 characters in tutorials 3a, 4a, and 6a
Improve readability of example prompts by wrapping long lines at 80
characters in both markdown and HTML versions of the AI tutorials.

Co-authored-by: A Samuel Pottinger <sam.pottinger@berkeley.edu>
