### 2026-01-20 01991c7
Add test assertions and flip file to diagnose tutorial 6 order bug
- Add Test 3 to testTutorial06 asserting Combined <= Permit in 2034
- Add testTutorial06Flip test for flipped policy order (Permit then Recycling)
- Create tutorial_06_flip.qta with reversed policy order for testing

The flip test is expected to FAIL, demonstrating the order dependency bug
where recycling after permit undoes the cap with 100% induction.
