### 2026-01-20 d7b71e2
Refactor ui_editor.js into focused modules for maintainability.
Split monolithic ui_editor.js (3,253 lines) into 5 focused files:
- ui_editor_support.js: Standalone command init/read functions
- ui_editor_app.js: ApplicationsListPresenter
- ui_editor_sim.js: SimulationListPresenter
- ui_editor_action.js: ConsumptionListPresenter, PolicyListPresenter,
  ReminderPresenter, StreamSelectionAvailabilityUpdater
- ui_editor.js: UiEditorPresenter, SubstanceTablePresenter,
  ActiveSubstancesResult

Updated importmaps in index.html and test.html.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
