+++ editor/js/ui_editor.js
@@ -3,1994 +3,15 @@
  *
  * @license BSD, see LICENSE.md.
  */
-import {ParsedYear, YearMatcher} from "duration";
-import {EngineNumber} from "engine_number";
+import {Program} from "ui_translator_components";
 import {MetaSerializer, MetaChangeApplier} from "meta_serialization";
-import {GwpLookupPresenter} from "known_substance";
-import {NumberParseUtil} from "number_parse_util";
-
+import {DuplicateEntityPresenter} from "duplicate_util";
+import {ApplicationsListPresenter} from "ui_editor_app";
+import {SimulationListPresenter} from "ui_editor_sim";
 import {
-  Application,
-  Command,
-  DefinitionalStanza,
-  LimitCommand,
-  Program,
-  RechargeCommand,
-  RecycleCommand,
-  ReplaceCommand,
-  RetireCommand,
-  SimulationScenario,
-  SubstanceBuilder,
-} from "ui_translator_components";
-import {
-  NameConflictResolution,
-  resolveNameConflict,
-  resolveSubstanceNameConflict,
-  DuplicateEntityPresenter,
-} from "duplicate_util";
-import {
-  ENABLEABLE_STREAMS,
-  ALWAYS_ON_STREAMS,
-  VALID_YEAR_KEYWORDS,
-} from "ui_editor_const";
-import {
-  STREAM_TARGET_SELECTORS,
-  updateDurationSelector,
-  setupDurationSelector,
-  buildSetupListButton,
-  setFieldValue,
-  getFieldValue,
-  getSanitizedFieldValue,
-  setListInput,
-  getListInput,
-  setEngineNumberValue,
-  getEngineNumberValue,
-  invertNumberString,
-  validateNumericInputs,
-  validateSimulationDuration,
-  setDuring,
-  buildUpdateCount,
-  setupDialogInternalLinks,
-} from "ui_editor_util";
-
-/**
- * Manages the UI for listing and editing applications.
- *
- * Manages the UI for listing and editing applications where these refer to
- * collections of substances based on use like commercial refrigeration.
- */
-class ApplicationsListPresenter {
-  /**
-   * Creates a new ApplicationsListPresenter.
-   *
-   * @param {HTMLElement} root - Root DOM element for the applications list.
-   * @param {Function} getCodeObj - Callback to get the current code object.
-   * @param {Function} onCodeObjUpdate - Callback when code object is updated.
-   */
-  constructor(root, getCodeObj, onCodeObjUpdate) {
-    const self = this;
-    self._root = root;
-    self._dialog = self._root.querySelector(".dialog");
-    self._getCodeObj = getCodeObj;
-    self._onCodeObjUpdate = onCodeObjUpdate;
-    self._editingName = null;
-    self._setupDialog();
-    self.refresh();
-  }
-
-  /**
-   * Refreshes the applications list display.
-   *
-   * @param {Object} codeObj - Current code object.
-   */
-  refresh(codeObj) {
-    const self = this;
-    self._refreshList(codeObj);
-  }
-
-  /**
-   * Updates the applications list UI with current data.
-   *
-   * @param {Object} codeObj - Current code object from which to extract
-   *     applications.
-   * @private
-   */
-  _refreshList(codeObj) {
-    const self = this;
-    const appNames = self._getAppNames();
-    const itemList = d3.select(self._root).select(".item-list");
-
-    itemList.html("");
-    const newItems = itemList.selectAll("li").data(appNames).enter().append("li");
-
-    newItems.attr("aria-label", (x) => x);
-
-    const buttonsPane = newItems.append("div").classed("list-buttons", true);
-
-    newItems
-      .append("div")
-      .classed("list-label", true)
[…]
