+++ editor/js/ui_editor_app.js
@@ -1,0 +1,229 @@
+/**
+ * Presenter for managing the applications list in the UI editor.
+ *
+ * @license BSD, see LICENSE.md.
+ */
+import {MetaSerializer, MetaChangeApplier} from "meta_serialization";
+import {Application} from "ui_translator_components";
+import {
+  NameConflictResolution,
+  resolveNameConflict,
+  DuplicateEntityPresenter,
+} from "duplicate_util";
+import {
+  buildSetupListButton,
+  getSanitizedFieldValue,
+  setupDialogInternalLinks,
+} from "ui_editor_util";
+
+/**
+ * Manages the UI for listing and editing applications.
+ *
+ * Manages the UI for listing and editing applications where these refer to
+ * collections of substances based on use like commercial refrigeration.
+ */
+class ApplicationsListPresenter {
+  /**
+   * Creates a new ApplicationsListPresenter.
+   *
+   * @param {HTMLElement} root - Root DOM element for the applications list.
+   * @param {Function} getCodeObj - Callback to get the current code object.
+   * @param {Function} onCodeObjUpdate - Callback when code object is updated.
+   */
+  constructor(root, getCodeObj, onCodeObjUpdate) {
+    const self = this;
+    self._root = root;
+    self._dialog = self._root.querySelector(".dialog");
+    self._getCodeObj = getCodeObj;
+    self._onCodeObjUpdate = onCodeObjUpdate;
+    self._editingName = null;
+    self._setupDialog();
+    self.refresh();
+  }
+
+  /**
+   * Refreshes the applications list display.
+   *
+   * @param {Object} codeObj - Current code object.
+   */
+  refresh(codeObj) {
+    const self = this;
+    self._refreshList(codeObj);
+  }
+
+  /**
+   * Updates the applications list UI with current data.
+   *
+   * @param {Object} codeObj - Current code object from which to extract
+   *     applications.
+   * @private
+   */
+  _refreshList(codeObj) {
+    const self = this;
+    const appNames = self._getAppNames();
+    const itemList = d3.select(self._root).select(".item-list");
+
+    itemList.html("");
+    const newItems = itemList.selectAll("li").data(appNames).enter().append("li");
+
+    newItems.attr("aria-label", (x) => x);
+
+    const buttonsPane = newItems.append("div").classed("list-buttons", true);
+
+    newItems
+      .append("div")
+      .classed("list-label", true)
+      .text((x) => x);
+
+    buttonsPane
+      .append("a")
+      .attr("href", "#")
+      .on("click", (event, x) => {
+        event.preventDefault();
+        self._showDialogFor(x);
+      })
+      .text("edit")
+      .attr("aria-label", (x) => "edit " + x);
+
+    buttonsPane.append("span").text(" | ");
+
+    buttonsPane
+      .append("a")
+      .attr("href", "#")
+      .on("click", (event, x) => {
+        event.preventDefault();
+        const message = "Are you sure you want to delete " + x + "?";
+        const isConfirmed = confirm(message);
+        if (isConfirmed) {
+          const codeObj = self._getCodeObj();
+          codeObj.deleteApplication(x);
+          self._onCodeObjUpdate(codeObj);
+        }
+      })
+      .text("delete")
+      .attr("aria-label", (x) => "delete " + x);
+  }
+
+  /**
+   * Sets up the dialog window for adding/editing applications.
+   *
+   * @private
+   */
+  _setupDialog() {
+    const self = this;
+    const addLink = self._root.querySelector(".add-link");
+    addLink.addEventListener("click", (event) => {
+      self._showDialogFor(null);
+      event.preventDefault();
+    });
+
+    const closeButton = self._root.querySelector(".cancel-button");
[…]
