+++ editor/js/ui_editor_sim.js
@@ -1,0 +1,295 @@
+/**
+ * Presenter for managing the simulations list in the UI editor.
+ *
+ * @license BSD, see LICENSE.md.
+ */
+import {ParsedYear, YearMatcher} from "duration";
+import {SimulationScenario} from "ui_translator_components";
+import {
+  NameConflictResolution,
+  resolveNameConflict,
+  DuplicateEntityPresenter,
+} from "duplicate_util";
+import {
+  updateDurationSelector,
+  setupDurationSelector,
+  buildSetupListButton,
+  getFieldValue,
+  getSanitizedFieldValue,
+  setFieldValue,
+  validateNumericInputs,
+  validateSimulationDuration,
+  setupDialogInternalLinks,
+} from "ui_editor_util";
+
+/**
+ * Manages the UI for listing and editing simulations.
+ */
+class SimulationListPresenter {
+  /**
+   * Creates a new SimulationListPresenter.
+   *
+   * @param {HTMLElement} root - The root DOM element for the simulation list.
+   * @param {Function} getCodeObj - Callback to get the current code object.
+   * @param {Function} onCodeObjUpdate - Callback when the code object is updated.
+   */
+  constructor(root, getCodeObj, onCodeObjUpdate) {
+    const self = this;
+    self._root = root;
+    self._dialog = self._root.querySelector(".dialog");
+    self._getCodeObj = getCodeObj;
+    self._onCodeObjUpdate = onCodeObjUpdate;
+    self._editingName = null;
+    self._setupDialog();
+    self.refresh();
+  }
+
+  /**
+   * Visually and functionally enables the simulation list interface.
+   */
+  enable() {
+    const self = this;
+    self._root.classList.remove("inactive");
+  }
+
+  /**
+   * Visually and functionally disables the simulation list interface.
+   */
+  disable() {
+    const self = this;
+    self._root.classList.add("inactive");
+  }
+
+  /**
+   * Refreshes the simulation list display.
+   *
+   * @param {Object} codeObj - Current code object to display.
+   */
+  refresh(codeObj) {
+    const self = this;
+    self._refreshList(codeObj);
+  }
+
+  /**
+   * Refreshes the simulation list display.
+   *
+   * @param {Object} codeObj - The current code object.
+   * @private
+   */
+  _refreshList(codeObj) {
+    const self = this;
+    const simulationNames = self._getSimulationNames();
+    const itemList = d3.select(self._root).select(".item-list");
+
+    itemList.html("");
+    const newItems = itemList.selectAll("li").data(simulationNames).enter().append("li");
+
+    newItems.attr("aria-label", (x) => x);
+
+    const buttonsPane = newItems.append("div").classed("list-buttons", true);
+
+    newItems
+      .append("div")
+      .classed("list-label", true)
+      .text((x) => x);
+
+    buttonsPane
+      .append("a")
+      .attr("href", "#")
+      .on("click", (event, x) => {
+        event.preventDefault();
+        self._showDialogFor(x);
+      })
+      .text("edit")
+      .attr("aria-label", (x) => "edit " + x);
+
+    buttonsPane.append("span").text(" | ");
+
+    buttonsPane
+      .append("a")
+      .attr("href", "#")
+      .on("click", (event, x) => {
+        event.preventDefault();
+        const message = "Are you sure you want to delete " + x + "?";
+        const isConfirmed = confirm(message);
+        if (isConfirmed) {
+          const codeObj = self._getCodeObj();
+          codeObj.deleteScenario(x);
+          self._onCodeObjUpdate(codeObj);
+        }
+      })
+      .text("delete")
+      .attr("aria-label", (x) => "delete " + x);
+  }
+
+  /**
+   * Sets up the dialog for adding/editing simulations.
+   *
[…]
