+++ editor/js/ui_editor_action.js
@@ -1,0 +1,1673 @@
+/**
+ * Presenters for managing consumption and policy actions in the UI editor.
+ *
+ * @license BSD, see LICENSE.md.
+ */
+import {ParsedYear, YearMatcher} from "duration";
+import {EngineNumber} from "engine_number";
+import {MetaSerializer, MetaChangeApplier} from "meta_serialization";
+import {GwpLookupPresenter} from "known_substance";
+import {NumberParseUtil} from "number_parse_util";
+import {
+  Application,
+  Command,
+  DefinitionalStanza,
+  RetireCommand,
+  SubstanceBuilder,
+} from "ui_translator_components";
+import {
+  NameConflictResolution,
+  resolveNameConflict,
+  resolveSubstanceNameConflict,
+  DuplicateEntityPresenter,
+} from "duplicate_util";
+import {ENABLEABLE_STREAMS, ALWAYS_ON_STREAMS} from "ui_editor_const";
+import {
+  STREAM_TARGET_SELECTORS,
+  updateDurationSelector,
+  setupDurationSelector,
+  buildSetupListButton,
+  setFieldValue,
+  getFieldValue,
+  getSanitizedFieldValue,
+  setListInput,
+  getListInput,
+  setEngineNumberValue,
+  getEngineNumberValue,
+  validateNumericInputs,
+  setDuring,
+  buildUpdateCount,
+  setupDialogInternalLinks,
+} from "ui_editor_util";
+import {
+  initSetCommandUi,
+  readSetCommandUi,
+  initChangeCommandUi,
+  readChangeCommandUi,
+  initLimitCommandUi,
+  readLimitCommandUi,
+  initRechargeCommandUi,
+  readRechargeCommandUi,
+  initRecycleCommandUi,
+  readRecycleCommandUi,
+  initReplaceCommandUi,
+  readReplaceCommandUi,
+  readDurationUi,
+} from "ui_editor_support";
+
+/**
+ * Manages stream selection availability based on context.
+ */
+class StreamSelectionAvailabilityUpdater {
+  constructor(container = document, context = null) {
+    const self = this;
+    self._container = container;
+    self._context = context;
+  }
+
+  /**
+   * Gets enabled streams for the current context.
+   * @param {Object} codeObj - The code object containing substances data.
+   * @param {string} substanceName - The name of the substance to check.
+   * @param {string} context - Optional context override ('consumption' or 'policy').
+   * @returns {Array<string>} Array of enabled stream names.
+   */
+  getEnabledStreamsForCurrentContext(codeObj, substanceName, context = null) {
+    const actualContext = context || this._context;
+
+    if (actualContext === "consumption") {
+      return this._getCurrentEnabledStreamsFromCheckboxes();
+    } else if (actualContext === "policy") {
+      return this._getCurrentEnabledStreamsFromCode(codeObj, substanceName);
+    }
+
+    return [];
+  }
+
+  /**
+   * Gets the first enabled option from a select element.
+   * @param {HTMLSelectElement} selectElement - The select element to check.
+   * @returns {string} Value of the first enabled option, or 'sales' as fallback.
+   */
+  getFirstEnabledOption(selectElement) {
+    const options = selectElement.querySelectorAll("option:not([disabled])");
+    return options.length > 0 ? options[0].value : "sales";
+  }
+
+  /**
+   * Updates stream option disabled states based on enabled streams.
+   * @param {HTMLSelectElement} selectElement - The select element to update.
+   * @param {Array<string>} enabledStreams - Array of enabled stream names.
+   */
+  updateStreamOptionStates(selectElement, enabledStreams) {
+    const options = selectElement.querySelectorAll("option");
[…]

[…]
