### 2026-01-20 fe1cf06
Add policy ordering controls to simulation dialog (non-functional).
- Add "specify policy order" link with .enable-policy-order-holder class
- Add HTML template #sim-order-controls-template for move controls
- Add _renderOrderControls method to SimulationListPresenter
- Add CSS styling for ordering controls
- Add unit tests for template rendering logic

Controls render for all policies but are not yet functional.
Click handlers will be added in Component 4.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
