### 2026-01-20 2531c85
Add mode detection and render ordering for policy list.
Implements logic to determine simple vs explicit ordering mode
based on selected policies, and orders policy rendering accordingly.
Includes 7 unit tests for mode detection and render order.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
