### 2026-01-20 7efb6b1
Add visibility controls for policy ordering mode.
Implements show/hide logic for simple vs explicit ordering:
- _showForSimpleOrdering() shows "specify policy order" link, hides move controls
- _showForExplicitOrdering() hides link, shows move controls
- Click listener enables switching to explicit mode
- Removes .sort() from _parseObj() to preserve explicit ordering
- Adds 3 unit tests for visibility methods

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
