### 2026-01-20 615ad6c
feat(reorder): implement policy reordering functionality
Add core reordering capability to policy dialog by implementing event-delegated
click handlers for move before/after links, extracting _renderPolicyCheckboxes()
method for re-rendering, and adding _movePolicyUp() and _movePolicyDown()
methods with comprehensive test coverage. Checkbox states preserved across
re-renders, supporting both simple and explicit ordering modes.

- Add event delegation for move-policy-up/down-link clicks in _setupDialog()
- Extract _renderPolicyCheckboxes() from _showDialogFor() to enable re-rendering
- Implemen
[…]
