### 2026-01-20 e4d8fca
feat(ui): implement position-based visibility for policy move controls (Component 5)
Implement fine-grained visibility controls for move links in the policy ordering interface:
- Hide "move before" link for first policy (cannot move up)
- Hide "move after" link for last policy (cannot move down)
- Hide separator when either adjacent link is hidden for clean presentation

Add _updateMoveControlVisibility() method that queries all policy labels, determines each
position, and applies position-specific visibility rules. Method is called from
_renderPolicyCheckboxes() after mode-based visibility to
[…]
