+++ editor/test/test_ui_editor_sim.js
@@ -341,6 +341,8 @@ function buildUiEditorSimTests() {
         _renderOrderControls: SimulationListPresenter.prototype._renderOrderControls,
         _showForExplicitOrdering: SimulationListPresenter.prototype._showForExplicitOrdering,
         _showForSimpleOrdering: SimulationListPresenter.prototype._showForSimpleOrdering,
+        _updateMoveControlVisibility:
+          SimulationListPresenter.prototype._updateMoveControlVisibility,
         _orderControlsTemplate: document.getElementById("sim-order-controls-template").innerHTML,
       };
 
@@ -383,6 +385,8 @@ function buildUiEditorSimTests() {
           _renderOrderControls: SimulationListPresenter.prototype._renderOrderControls,
           _showForExplicitOrdering: SimulationListPresenter.prototype._showForExplicitOrdering,
           _showForSimpleOrdering: SimulationListPresenter.prototype._showForSimpleOrdering,
+          _updateMoveControlVisibility:
+            SimulationListPresenter.prototype._updateMoveControlVisibility,
           _orderControlsTemplate: document.getElementById("sim-order-controls-template").innerHTML,
         };
 
@@ -494,6 +498,239 @@ function buildUiEditorSimTests() {
         "Array should remain unchanged when policy not found",
       );
     });
+
+    QUnit.test(
+      "_updateMoveControlVisibility hides move-up for first policy",
+      function (assert) {
+        const dialogElement = document.createElement("div");
+        dialogElement.innerHTML = `
+      <div class="policy-check-label">
+        <span class="move-policy-control">
+          <a href="#" class="move-policy-up-link">move before</a>
+          <span class="move-policy-sep"> / </span>
+          <a href="#" class="move-policy-down-link">move after</a>
+        </span>
+      </div>
+      <div class="policy-check-label">
+        <span class="move-policy-control">
+          <a href="#" class="move-policy-up-link">move before</a>
+          <span class="move-policy-sep"> / </span>
+          <a href="#" class="move-policy-down-link">move after</a>
+        </span>
+      </div>
+      <div class="policy-check-label">
+        <span class="move-policy-control">
+          <a href="#" class="move-policy-up-link">move before</a>
+          <span class="move-policy-sep"> / </span>
+          <a href="#" class="move-policy-down-link">move after</a>
+        </span>
+      </div>
+    `;
+
+        const mockPresenter = {
+          _dialog: dialogElement,
+          _updateMoveControlVisibility:
+            SimulationListPresenter.prototype._updateMoveControlVisibility,
+        };
+
+        mockPresenter._updateMoveControlVisibility();
+
+        const policyLabels = dialogElement.querySelectorAll(".policy-check-label");
+        const firstLabel = policyLabels[0];
+        const secondLabel = policyLabels[1];
+
+        // First policy should have move-up hidden
+        const firstMoveUp = firstLabel.querySelector(".move-policy-up-link");
+        const firstMoveDown = firstLabel.querySelector(".move-policy-down-link");
+        const firstSep = firstLabel.querySelector(".move-policy-sep");
+
+        assert.equal(firstMoveUp.style.display, "none", "First policy move-up should be hidden");
+        assert.equal(
+          firstMoveDown.style.display,
+          "inline",
+          "First policy move-down should be visible",
+        );
+        assert.equal(firstSep.style.display, "none", "First policy separator should be hidden");
+
+        // Second (middle) policy should have both visible
+        const secondMoveUp = secondLabel.querySelector(".move-policy-up-link");
+        const secondMoveDown = secondLabel.querySelector(".move-policy-down-link");
+        const secondSep = secondLabel.querySelector(".move-policy-sep");
+
+        assert.equal(
+          secondMoveUp.style.display,
[…]
