### 2026-01-21 f0f09ab
test(ui_editor_support): expand QUnit test coverage for validateInductionInput and readDurationUi
Add comprehensive unit test coverage for ui_editor_support.js pure functions:
- validateInductionInput: 11 tests covering valid percentages, boundary values, error cases, and edge cases
- readDurationUi: 8 tests covering all duration types (in year, during all years, starting in year, ending in year, during years) and ParsedYear object creation

New test file: test_ui_editor_support.js with 19 total tests. All tests passing with no linting errors.
Component 6 complete: validates critical validatio
[…]
