+++ editor/test/test_ui_editor_support.js
@@ -1,0 +1,243 @@
+/**
+ * Tests for ui_editor_support.js functions.
+ *
+ * @license BSD, see LICENSE.md.
+ */
+import {validateInductionInput, readDurationUi} from "ui_editor_support";
+import {EngineNumber} from "engine_number";
+import {ParsedYear} from "duration";
+
+function buildUiEditorSupportTests() {
+  QUnit.module("ui_editor_support", function (hooks) {
+    QUnit.module("validateInductionInput", function () {
+      QUnit.test("accepts valid numeric values", function (assert) {
+        const result0 = validateInductionInput("0");
+        assert.ok(result0 instanceof EngineNumber, "0 should return EngineNumber");
+        assert.equal(result0.getValue(), 0, "0 should have value 0");
+
+        const result50 = validateInductionInput("50");
+        assert.ok(result50 instanceof EngineNumber, "50 should return EngineNumber");
+        assert.equal(result50.getValue(), 50, "50 should have value 50");
+
+        const result100 = validateInductionInput("100");
+        assert.ok(result100 instanceof EngineNumber, "100 should return EngineNumber");
+        assert.equal(result100.getValue(), 100, "100 should have value 100");
+      });
+
+      QUnit.test("preserves original string in EngineNumber", function (assert) {
+        const result = validateInductionInput("50.5");
+        assert.ok(result instanceof EngineNumber, "Should return EngineNumber");
+        assert.equal(result.getValue(), 50.5, "Should have correct numeric value");
+        assert.equal(result.getOriginalString(), "50.5", "Should preserve original string");
+      });
+
+      QUnit.test("accepts 'default' keyword", function (assert) {
+        const result = validateInductionInput("default");
+        assert.equal(result, "default", "Should return 'default' string");
+      });
+
+      QUnit.test("rejects empty string with descriptive error", function (assert) {
+        assert.throws(
+          () => validateInductionInput(""),
+          /Induction rate is required/,
+          "Empty string should throw error about required field",
+        );
+      });
+
+      QUnit.test("rejects negative values", function (assert) {
+        assert.throws(
+          () => validateInductionInput("-10"),
+          /out of range.*0-100/,
+          "Negative value should throw range error",
+        );
+      });
+
+      QUnit.test("rejects values over 100", function (assert) {
+        assert.throws(
+          () => validateInductionInput("150"),
+          /out of range.*0-100/,
+          "Value over 100 should throw range error",
+        );
+      });
+
+      QUnit.test("rejects non-numeric strings", function (assert) {
+        assert.throws(
+          () => validateInductionInput("abc"),
+          /Invalid induction rate.*Must be a number/,
+          "Non-numeric string should throw invalid format error",
+        );
+      });
+
+      QUnit.test("rejects NaN values", function (assert) {
+        assert.throws(
+          () => validateInductionInput("NaN"),
+          /Invalid induction rate.*Must be a number/,
+          "NaN string should throw invalid format error",
+        );
+      });
+
+      QUnit.test("handles whitespace in numeric values", function (assert) {
+        const result = validateInductionInput("  50  ");
+        assert.ok(result instanceof EngineNumber, "Should handle whitespace");
+        assert.equal(result.getValue(), 50, "Should parse numeric value correctly");
+        assert.equal(result.getOriginalString(), "50", "Should trim whitespace");
+      });
+
+      QUnit.test("accepts boundary values", function (assert) {
+        const result0 = validateInductionInput("0");
+        assert.equal(result0.getValue(), 0, "0 is valid boundary");
+
+        const result100 = validateInductionInput("100");
+        assert.equal(result100.getValue(), 100, "100 is valid boundary");
+      });
+
+      QUnit.test("accepts decimal values within range", function (assert) {
+        const result = validateInductionInput("33.33");
[…]
