### 2026-01-21 bc7256c
feat(ui_editor): update JSDoc and sort imports alphabetically
