### 2026-01-21 5b93854
refactor(ui_editor_action): clean up formatting and JSDoc
Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
