### 2026-01-21 6a60370
refactor(ui_editor_action): improve StreamSelectionAvailabilityUpdater clarity
Remove constructor defaults and add comprehensive JSDoc. Clarify context parameter
and simplify defensive null checking for checkboxes. Remove getEnables type
introspection. Extract _hasNoSubstanceSelected private method. All tests pass.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
