### 2026-01-21 f36ae6e
refactor(consumption): extract methods and break up setupDialog Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
