+++ editor/js/ui_editor_action.js
@@ -105,6 +105,7 @@ class StreamSelectionAvailabilityUpdater {
 
   /**
    * Gets the first enabled option from a select element.
+   *
    * @param {HTMLSelectElement} selectElement - The select element to check.
    * @returns {string} Value of the first enabled option, or 'sales' as fallback.
    */
@@ -115,6 +116,7 @@ class StreamSelectionAvailabilityUpdater {
 
   /**
    * Updates stream option disabled states based on enabled streams.
+   *
    * @param {HTMLSelectElement} selectElement - The select element to update.
    * @param {Array<string>} enabledStreams - Array of enabled stream names.
    */
@@ -137,6 +139,7 @@ class StreamSelectionAvailabilityUpdater {
 
   /**
    * Updates all stream target dropdowns within the container.
+   *
    * @param {Array<string>} enabledStreams - Array of enabled stream names.
    * @param {Function} filterFn - Optional filter function (deprecated).
    */
@@ -151,6 +154,7 @@ class StreamSelectionAvailabilityUpdater {
 
   /**
    * Refreshes all stream target dropdowns for a specific substance.
+   *
    * @param {Object} codeObj - The code object containing substances data.
    * @param {string} substanceName - The name of the substance whose streams changed.
    */
@@ -450,27 +454,44 @@ class ConsumptionListPresenter {
       .attr("href", "#")
       .on("click", (event, x) => {
         event.preventDefault();
-        const message = "Are you sure you want to delete " + x + "?";
-        const isConfirmed = confirm(message);
-        if (isConfirmed) {
-          const codeObj = self._getCodeObj();
-          const objIdentifierRegex = /\"([^\"]+)\" for \"([^\"]+)\"/;
-          const match = x.match(objIdentifierRegex);
-          const substance = match[1];
-          const application = match[2];
-          codeObj.deleteSubstance(application, substance);
-          self._onCodeObjUpdate(codeObj);
-        }
+        self._handleDeleteConsumption(x);
       })
       .text("delete")
       .attr("aria-label", (x) => "delete " + x);
   }
 
   /**
+   * Handles deletion of a consumption record.
+   *
+   * Prompts the user for confirmation before deleting the specified consumption.
+   * Parses the consumption name to extract substance and application, then
+   * removes it from the code object.
+   *
+   * @param {string} consumptionName - The full name of the consumption to delete.
+   * @private
+   */
+  _handleDeleteConsumption(consumptionName) {
+    const self = this;
+    const message = "Are you sure you want to delete " + consumptionName + "?";
+    const isConfirmed = confirm(message);
+    if (!isConfirmed) {
+      return;
+    }
+
+    const codeObj = self._getCodeObj();
+    const objIdentifierRegex = /\"([^\"]+)\" for \"([^\"]+)\"/;
+    const match = consumptionName.match(objIdentifierRegex);
+    const substance = match[1];
+    const application = match[2];
+    codeObj.deleteSubstance(application, substance);
+    self._onCodeObjUpdate(codeObj);
+  }
+
+  /**
    * Sets up the dialog for adding/editing consumption records.
    *
-   * Sets up the dialog for adding/editing consumption records, initializing
-   * tabs and event handlers.
+   * Orchestrates initialization of all dialog components including tabs,
+   * buttons, command lists, reminders, GWP lookup, and stream checkboxes.
    *
    * @private
    */
@@ -478,6 +499,24 @@ class ConsumptionListPresenter {
     const self = this;
 
     self._tabs = new Tabby("#" + self._dialog.querySelector(".tabs").id);
+
+    self._setupDialogButtons();
+    self._setupDialogCommandLists();
+    self._setupDialogReminders();
+    self._setupDialogGwpLookup();
[…]
