### 2026-01-21 b4b8e90
refactor(policy): extract substance listener and simplify save
