### 2026-01-21 fecf281
refactor(ui_editor_app): format D3 chain for readability
