### 2026-01-21 0461ed6
refactor(sim): improve code clarity and JSDoc documentation Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
