+++ editor/js/ui_editor_sim.js
@@ -85,19 +85,20 @@ class SimulationListPresenter {
     const itemList = d3.select(self._root).select(".item-list");
 
     itemList.html("");
-    const newItems = itemList.selectAll("li").data(simulationNames).enter().append("li");
+    const newItems = itemList.selectAll("li")
+      .data(simulationNames)
+      .enter()
+      .append("li");
 
     newItems.attr("aria-label", (x) => x);
 
     const buttonsPane = newItems.append("div").classed("list-buttons", true);
 
-    newItems
-      .append("div")
+    newItems.append("div")
       .classed("list-label", true)
       .text((x) => x);
 
-    buttonsPane
-      .append("a")
+    buttonsPane.append("a")
       .attr("href", "#")
       .on("click", (event, x) => {
         event.preventDefault();
@@ -108,8 +109,7 @@ class SimulationListPresenter {
 
     buttonsPane.append("span").text(" | ");
 
-    buttonsPane
-      .append("a")
+    buttonsPane.append("a")
       .attr("href", "#")
       .on("click", (event, x) => {
         event.preventDefault();
@@ -163,26 +163,38 @@ class SimulationListPresenter {
       self._showForExplicitOrdering();
     });
 
-    // Add click delegation for move up/down links
     self._dialog.addEventListener("click", (event) => {
-      const target = event.target;
-
-      // Handle move up (move before)
-      if (target.classList.contains("move-policy-up-link")) {
-        event.preventDefault();
-        const policyName = target.getAttribute("data-policy-name");
-        self._movePolicyUp(policyName);
-        return;
-      }
-
-      // Handle move down (move after)
-      if (target.classList.contains("move-policy-down-link")) {
-        event.preventDefault();
-        const policyName = target.getAttribute("data-policy-name");
-        self._movePolicyDown(policyName);
-        return;
-      }
+      self._handleMoveLinkClick(event);
     });
+  }
+
+  /**
+   * Handles click events for move up/down links within the policy list.
+   *
+   * Uses event delegation to handle clicks on move policy links. When a move
+   * up link is clicked, moves the policy before the previous policy. When a
+   * move down link is clicked, moves the policy after the next policy.
+   *
+   * @param {Event} event - The click event.
+   * @private
+   */
+  _handleMoveLinkClick(event) {
+    const self = this;
+    const target = event.target;
+
+    if (target.classList.contains("move-policy-up-link")) {
+      event.preventDefault();
+      const policyName = target.getAttribute("data-policy-name");
+      self._movePolicyUp(policyName);
+      return;
+    }
+
+    if (target.classList.contains("move-policy-down-link")) {
+      event.preventDefault();
+      const policyName = target.getAttribute("data-policy-name");
+      self._movePolicyDown(policyName);
+      return;
+    }
   }
 
   /**
@@ -219,9 +231,7 @@ class SimulationListPresenter {
       x.getYearEnd(),
     );
 
-    // Get all available policies
-    const allPolicyNames = self
-      ._getCodeObj()
+    const allPolicyNames = self._getCodeObj()
       .getPolicies()
       .map((x) => x.getName());
 
@@ -255,12 +265,12 @@ class SimulationListPresenter {
   /**
    * Determines whether to use simple or explicit ordering mode.
    *
-   * Simple ordering mode is used if any of these conditions are true:
-   * - No policies are selected
-   * - Only one policy is selected
-   * - Selected policies are in ascending alphabetical order
+   * Uses simple ordering mode if any of these rules are true:
+   * 1. No policies are selected
+   * 2. Only one policy is selected
+   * 3. Selected policies are in ascending alphabetical order
    *
[…]
