### 2026-01-21 dfc9755
refactor(sim): extract policy swap logic into _movePolicy
