### 2026-01-21 37117d7
refactor: rename ui_editor_support to ui_editor_strategy
Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
