### 2026-01-21 cbd15db
refactor: move STREAM_TARGET_SELECTORS to ui_editor_const.js
Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
