### 2026-01-21 0b9c256
refactor(editor): improve readability in ui_editor_action.js Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
