### 2026-01-21 2432aca
refactor(ui_editor_sim): add hasClass helper and explicit offset
