### 2026-01-21 40bfcfa
docs: add ui_editor_* descriptions to JSDoc header Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
