### 2026-01-22 431d7d2
Add failing tests for policy stacking with percentage changes
Add StackingLiveTests with two test cases comparing policy stacking
behavior between percentage-based and explicit absolute changes.
The percentage test fails while the explicit test passes, revealing
a bug in how percentage changes interact with caps in stacked policies.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
