### 2026-01-22 98f371b
Clear policy checkboxes on new scenario.
Clear the policy checkboxes in the scenario dialog when the user elects to make a new policy. Currently, the prior selection caries over.
