### 2026-01-22 270570a
fix(stacking): update lastSpecified when cap/floor/displacement modifies streams
When policies are stacked (e.g., using "Permit" then "Big Change"), percentage-based
changes in later policies need to use values modified by earlier policies. This fix
ensures that:

1. LimitExecutor updates lastSpecified when a cap or floor is violated, so subsequent
   percentage changes use the capped/floored value as their base
2. DisplaceExecutor updates lastSpecified for displaced streams/substances, so
   subsequent percentage changes include the displaced amount
3. Component streams (domestic/import) are 
[…]
