+++ engine/src/main/java/org/kigalisim/engine/support/DisplaceExecutor.java
@@ -30,6 +30,7 @@ import org.kigalisim.engine.Engine;
 import org.kigalisim.engine.number.EngineNumber;
 import org.kigalisim.engine.number.UnitConverter;
 import org.kigalisim.engine.state.Scope;
+import org.kigalisim.engine.state.SimulationState;
 import org.kigalisim.lang.operation.DisplacementType;
 
 /**
@@ -247,6 +248,9 @@ public class DisplaceExecutor {
         Optional.empty(),
         Optional.empty()
     );
+
+    // Update lastSpecified for the displaced stream so subsequent percentage changes use new value
+    updateLastSpecifiedAfterDisplacement(displaceTarget, engine.getScope());
   }
 
   /**
@@ -295,6 +299,9 @@ public class DisplaceExecutor {
     // Use custom recalc kit with destination substance's properties for correct GWP calculation
     shortcuts.changeStreamWithDisplacementContext(stream, displaceChange, destinationScope);
 
+    // Update lastSpecified for the displaced stream in destination scope
+    updateLastSpecifiedAfterDisplacement(stream, destinationScope);
+
     // Restore original scope
     engine.setSubstance(originalSubstance);
   }
@@ -330,6 +337,9 @@ public class DisplaceExecutor {
           Optional.empty(),
           Optional.empty()
       );
+
+      // Update lastSpecified for the displaced stream
+      updateLastSpecifiedAfterDisplacement(displaceTarget, engine.getScope());
     } else {
       Scope currentScope = engine.getScope();
       Scope destinationScope = currentScope.getWithSubstance(displaceTarget);
@@ -339,6 +349,47 @@ public class DisplaceExecutor {
           displaceChange,
           destinationScope
       );
+
+      // Update lastSpecified for the displaced stream in destination scope
+      updateLastSpecifiedAfterDisplacement(stream, destinationScope);
+    }
+  }
+
+  /**
+   * Updates lastSpecified for a stream after displacement to ensure subsequent percentage changes
+   * use the new displaced value.
+   *
+   * <p>When displacement adds volume to a stream, the lastSpecified value must be updated
+   * so that subsequent percentage-based changes (e.g., "change sales by 5%") apply to the
+   * new value that includes the displaced amount, not the original value.</p>
+   *
+   * @param stream The stream that received the displacement
+   * @param scope The scope (UseKey) for the displaced stream
+   */
+  private void updateLastSpecifiedAfterDisplacement(String stream, Scope scope) {
+    SimulationState simulationState = engine.getStreamKeeper();
+
+    // Get the current value of the displaced stream
+    String originalSubstance = engine.getScope().getSubstance();
+    boolean needsScopeSwitch = !scope.getSubstance().equals(originalSubstance);
+
+    if (needsScopeSwitch) {
+      engine.setSubstance(scope.getSubstance());
+    }
+
+    EngineNumber currentValue = engine.getStream(stream);
+    simulationState.setLastSpecifiedValue(scope, stream, currentValue);
+
+    // For "sales" stream, also update lastSpecified for component streams (domestic/import)
+    if ("sales".equals(stream)) {
+      EngineNumber domesticValue = engine.getStream("domestic");
+      EngineNumber importValue = engine.getStream("import");
+      simulationState.setLastSpecifiedValue(scope, "domestic", domesticValue);
+      simulationState.setLastSpecifiedValue(scope, "import", importValue);
+    }
+
+    if (needsScopeSwitch) {
+      engine.setSubstance(originalSubstance);
     }
   }
 }
