### 2026-01-23 d479c7c
test(change): add failing test for negative stream values
Add testChangeNegativeValue() to demonstrate the bug where change
operations can result in negative stream values. The test intentionally
FAILS to document that when a stream has 5 kg and receives a -10 kg
change, it incorrectly becomes -5 kg instead of being clamped to 0 kg.
Also add change_negative_test.qta example file for the test.

This failing test establishes the baseline for Component 2 which will
fix the ChangeExecutor to clamp negative values to zero.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
