### 2026-01-23 2c62447
fix(change): prevent negative stream values from change operations
Implement clamping logic in ChangeExecutor to ensure change operations cannot result in negative stream values. When a change would exceed the current stream value, the change is automatically clamped to bring the stream to exactly zero instead.

- Add clampChangeToPreventNegative() helper method for reusable clamping logic
- Modify handleVolumeChange() to clamp volume-based changes
- Modify handleUnitsChange() to clamp unit-based changes in both branches
- Modify applyStandardPercentageChange() to clamp percentage changes
- Mo
[…]
