+++ engine/src/main/java/org/kigalisim/engine/support/ChangeExecutor.java
@@ -42,6 +42,30 @@ public class ChangeExecutor {
   }
 
   /**
+   * Clamps a change amount to prevent negative stream values.
+   *
+   * <p>This method ensures that applying the change will not result in a negative stream value.
+   * If the proposed change would make the stream negative, it is clamped to bring the stream
+   * to exactly zero instead.</p>
+   *
+   * @param currentValue The current stream value in kg
+   * @param proposedChange The proposed change amount in kg (can be negative)
+   * @return The clamped change amount in kg that will not result in negative values
+   */
+  private BigDecimal clampChangeToPreventNegative(EngineNumber currentValue,
+      BigDecimal proposedChange) {
+    BigDecimal currentKg = currentValue.getValue();
+    BigDecimal newValue = currentKg.add(proposedChange);
+
+    // If new value would be negative, clamp change to bring stream to exactly 0
+    if (newValue.compareTo(BigDecimal.ZERO) < 0) {
+      return currentKg.negate();  // Change of -currentKg brings stream to 0
+    }
+
+    return proposedChange;  // No clamping needed
+  }
+
+  /**
    * Execute a change operation by routing to the appropriate handler.
    *
    * @param stream The stream identifier to modify
@@ -116,7 +140,10 @@ public class ChangeExecutor {
     UnitConverter unitConverter = EngineSupportUtils.createUnitConverterWithTotal(engine, stream);
 
     EngineNumber convertedDelta = unitConverter.convert(amount, currentValue.getUnits());
-    BigDecimal newAmount = currentValue.getValue().add(convertedDelta.getValue());
+
+    // Clamp the change to prevent negative values
+    BigDecimal clampedDelta = clampChangeToPreventNegative(currentValue, convertedDelta.getValue());
+    BigDecimal newAmount = currentValue.getValue().add(clampedDelta);
     EngineNumber outputWithUnits = new EngineNumber(newAmount, currentValue.getUnits());
 
     StreamUpdate update = new StreamUpdateBuilder()
@@ -177,9 +204,14 @@ public class ChangeExecutor {
     EngineNumber currentValue = engine.getStream(stream, Optional.of(useKeyEffective), Optional.empty());
     BigDecimal percentageValue = amount.getValue();
     BigDecimal changeAmount = lastSpecified.getValue().multiply(percentageValue).divide(new BigDecimal("100"));
-    BigDecimal newTotalValue = currentValue.getValue().add(changeAmount);
 
-    EngineNumber newTotal = new EngineNumber(newTotalValue, lastSpecified.getUnits());
+    // Clamp to prevent negative values
+    BigDecimal proposedNewValue = currentValue.getValue().add(changeAmount);
+    BigDecimal clampedNewValue = proposedNewValue.compareTo(BigDecimal.ZERO) < 0
+        ? BigDecimal.ZERO
+        : proposedNewValue;
+
+    EngineNumber newTotal = new EngineNumber(clampedNewValue, lastSpecified.getUnits());
     boolean subtractRecycling = "units".equals(lastSpecified.getUnits());
     StreamUpdate update = new StreamUpdateBuilder()
         .setName(stream)
@@ -209,9 +241,14 @@ public class ChangeExecutor {
 
     BigDecimal percentageValue = amount.getValue();
     BigDecimal changeAmount = lastSpecified.getValue().multiply(percentageValue).divide(new BigDecimal("100"));
-    BigDecimal newTotalValue = lastSpecified.getValue().add(changeAmount);
 
-    EngineNumber newTotal = new EngineNumber(newTotalValue, lastSpecified.getUnits());
+    // Clamp to prevent negative values
[…]
