### 2026-01-23 79de24b
test(stacking): add year 2035 non-negative assertions to StackingLiveTests
Add regression protection for Component 2 fix by validating that stream
values remain non-negative in complex policy stacking scenarios. Added
helper method assertDomesticNonNegative() and year 2035 assertions for
all 5 scenarios (BAU, Big Change, Permit, Before, After) in both
testPolicyStackingOrderPercent and testPolicyStackingOrderExplicit.
Year 2035 is critical test point where Permit policy caps sales to 0 mt.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
