### 2026-01-23 affb17a
docs(change): document non-negative clamping behavior in Change command
Updated documentation in llms-full.txt and qubectalk_commands.html to explain
that change operations prevent streams from going negative. When a change would
result in a negative value, the stream is clamped to exactly zero, and only the
actual amount removed (not the requested amount) is used for downstream calculations.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
