### 2026-01-23 a5a5d54
style: add final modifier to fix VariableDeclarationUsageDistance warnings
Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
