### 2026-01-23 470e0d9
test(stacking): add testDoubleSet for set command stacking behavior
Add new test that validates how multiple set commands stack, confirming
that the last policy in the chain determines the final value. Includes
stacking_double_set.qta with BAU, Single Set, and Combined Set scenarios.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
