### 2026-01-23 472e707
test(stacking): add testDoubleCap for cap command stacking behavior
Created testDoubleCap in StackingLiveTests.java and stacking_double_cap.qta. Tests validate that multiple cap commands stack correctly with the most restrictive (lowest) cap taking effect regardless of application order. Five scenarios tested: BAU (10 kg), Top Cap (5 kg), Bottom Cap (2 kg), Combined Strict (2 kg), Combined Loose (2 kg).

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
