### 2026-01-23 390b7a0
test(stacking): add testDoubleFloor for floor command stacking behavior
Implements Component 3 of the stacking tests expansion. Validates that multiple
floor commands stack correctly with the least restrictive (highest) floor taking
effect regardless of application order. Includes five test scenarios (BAU, Top
Floor, Bottom Floor, Combined Strict, and Combined Loose) with expected year 2
values verified.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
