### 2026-01-23 cf7b276
test(stacking): add testDoubleChange for change command stacking behavior
Add two QTA files (stacking_double_change.qta and stacking_double_change_current.qta)
to test stacking behavior of explicit and percentage-based change commands. Both
variations (using '%' and '% current' syntax) produce identical results, confirming
equivalent behavior. Added testDoubleChange() and testDoubleChangeCurrent() methods
to StackingLiveTests.java to validate five scenarios: BAU (10 kg), Explicit Change
(20 kg), Percent Change (11 kg), Explicit First (22 kg), and Percent First (21 kg).

Co-Authored-By: Claude 
[…]
