### 2026-01-23 72eaa35
test(stacking): add testSetAndChange for set and change command interaction
Add testSetAndChange test in StackingLiveTests.java and corresponding
stacking_set_and_change.qta file. Tests validate the interaction between
set and change commands when policies are stacked in different orders.

Five scenarios tested:
- BAU: 10 kg (no policies)
- Set: 20 kg (set policy only)
- Change: 11 kg (change policy only, +10%)
- Set First: 22 kg (set then change, 20 kg + 10% = 22 kg)
- Change First: 20 kg (change then set, overwrites to 20 kg)

SPECIFICATION DEVIATION DOCUMENTED:
Original spec expected "Chang
[…]
