### 2026-01-23 8f67ab0
refactor(stacking): rename change and cap tests to match convention
Rename testPolicyStackingOrderPercent → testChangeAndCapPercent
Rename testPolicyStackingOrderExplicit → testChangeAndCapExplicit

QTA files renamed:
- stacking.qta → stacking_change_and_cap_percent.qta
- stacking_explicit.qta → stacking_change_and_cap_explicit.qta

This aligns with the naming convention established in Components 1-5
where all stacking tests use stacking_ prefix and descriptive operation names.

No functional changes - all test logic and assertions remain identical.

Co-Authored-By: Claude Opus 4.5 <noreply@an
[…]
