### 2026-01-23 d0c837b
refactor(ChangeExecutor): improve clarity with better naming and extract helper
