### 2026-01-23 d3b0990
refactor(DisplaceExecutor): clarify variable name, remove redundant comments
