### 2026-01-23 24e4278
refactor(LimitExecutor): consolidate lastSpecified updates
