+++ engine/src/main/java/org/kigalisim/engine/support/LimitExecutor.java
@@ -213,13 +213,7 @@ public class LimitExecutor {
           .build();
       engine.executeStreamUpdate(update);
 
-      // Update lastSpecified so subsequent percentage changes use the capped value
-      simulationState.setLastSpecifiedValue(scope, stream, newCappedValue);
-
-      // For "sales" stream, also update lastSpecified for component streams (domestic/import)
-      if ("sales".equals(stream)) {
-        updateComponentStreamsLastSpecified(simulationState, scope);
-      }
+      setCurrentValueAsLastSpecified(scope, stream);
 
       if (displaceTarget != null) {
         EngineNumber finalInKg = engine.getStream(stream);
@@ -244,14 +238,7 @@ public class LimitExecutor {
           Optional.empty()
       );
 
-      // Update lastSpecified so subsequent percentage changes use the capped value
-      EngineNumber cappedValue = engine.getStream(stream);
-      simulationState.setLastSpecifiedValue(scope, stream, cappedValue);
-
-      // For "sales" stream, also update lastSpecified for component streams (domestic/import)
-      if ("sales".equals(stream)) {
-        updateComponentStreamsLastSpecified(simulationState, scope);
-      }
+      setCurrentValueAsLastSpecified(scope, stream);
 
       displaceExecutor.execute(stream, amount, changeAmount, displaceTarget, displacementType);
     }
@@ -298,16 +285,8 @@ public class LimitExecutor {
         .build();
     engine.executeStreamUpdate(update);
 
-    // Update lastSpecified so subsequent percentage changes use the capped value
-    SimulationState simulationState = engine.getStreamKeeper();
     Scope scope = engine.getScope();
-    EngineNumber cappedValue = engine.getStream(stream);
-    simulationState.setLastSpecifiedValue(scope, stream, cappedValue);
-
-    // For "sales" stream, also update lastSpecified for component streams (domestic/import)
-    if ("sales".equals(stream)) {
-      updateComponentStreamsLastSpecified(simulationState, scope);
-    }
+    setCurrentValueAsLastSpecified(scope, stream);
 
     if (displaceTarget != null) {
       EngineNumber cappedInKg = engine.getStream(stream);
@@ -372,13 +351,7 @@ public class LimitExecutor {
           .build();
       engine.executeStreamUpdate(update);
 
-      // Update lastSpecified so subsequent percentage changes use the floored value
-      simulationState.setLastSpecifiedValue(scope, stream, newFloorValue);
-
-      // For "sales" stream, also update lastSpecified for component streams (domestic/import)
-      if ("sales".equals(stream)) {
-        updateComponentStreamsLastSpecified(simulationState, scope);
-      }
+      setCurrentValueAsLastSpecified(scope, stream);
 
       if (displaceTarget != null) {
         EngineNumber finalInKg = engine.getStream(stream);
@@ -403,14 +376,7 @@ public class LimitExecutor {
           Optional.empty()
       );
 
-      // Update lastSpecified so subsequent percentage changes use the floored value
-      EngineNumber flooredValue = engine.getStream(stream);
-      simulationState.setLastSpecifiedValue(scope, stream, flooredValue);
-
-      // For "sales" stream, also update lastSpecified for component streams (domestic/import)
-      if ("sales".equals(stream)) {
-        updateComponentStreamsLastSpecified(simulationState, scope);
-      }
+      setCurrentValueAsLastSpecified(scope, stream);
 
       displaceExecutor.execute(stream, amount, changeAmount, displaceTarget, displacementType);
     }
@@ -457,21 +423,34 @@ public class LimitExecutor {
         .build();
     engine.executeStreamUpdate(update);
 
-    // Update lastSpecified so subsequent percentage changes use the floored value
[…]
