### 2026-01-23 81de440
test(stacking): add testChangeAndFloor for floor and change interaction
Add testChangeAndFloor method to StackingLiveTests.java and corresponding
stacking_change_and_floor.qta file. This creates a mirror-image test of the
existing change and cap tests, validating that floor operations correctly
enforce minimum thresholds while interacting with change operations.

Key design:
- Uses positive change percentages (+7%, +3%) instead of negative
- Progressive floor values (1.5-3.0 mt) increasing over time
- Inverted assertion logic (> instead of <) for floor comparisons
- Six test scenarios validati
[…]
