### 2026-01-23 1f099c1
test(stacking): add set with cap units switch test (Component 8)
Tests interaction between set (units-based) and cap (percentage-based)
policies when stacked. Validates that "last policy wins" principle applies:
Set First produces 0 kg because cap to 0% is applied after set, and Cap First
produces > 10 kg because set overrides the prior cap. This is correct and
consistent with the stacking specification.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
