### 2026-01-23 dfd84fd
test(stacking): implement set with cap units switch (Component 8)
When Set policy specifies domestic in units (e.g., 'set domestic to 10 units')
and Cap policy follows with percentage (e.g., 'cap domestic to 0%'), the cap
was incorrectly using getStream() which returns kg values, causing recharge
to not be added since the StreamUpdate was in kg instead of units.

Modified LimitExecutor.capWithPercent() and floorWithPercent() to use
getLastSpecifiedValue() when it has equipment units, preserving unit-based
tracking through cap/floor operations. This ensures implicit recharge
calculation is trig
[…]
