### 2026-01-23 afc52c3
test(stacking): add testSetWithFloorUnitsSwitch (Component 9)
Tests interaction between set (units-based) and floor (percentage-based) policies.
Mirrors Component 8 structure with floor enforcing minimums instead of maximums.
Five scenarios: BAU (~10 kg), Set (>=10 kg), Floor (~10 kg), Set First (>=10 kg),
Floor First (~10 kg). Validates units/kg switching during stacking operations.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
