### 2026-01-23 35cafde
test(stacking): add testChangeWithUnitsSwitch test (Component 10)
Tests interaction between change (units-based) and cap (percentage-based) policies across five scenarios: BAU (10 kg), Change (+10 units = 20 kg), Cap (0%), Change First (20 kg then capped to 0 kg), and Cap First (capped to 0 kg then +10 units = 10 kg). Validates that change commands properly specify units and that units/kg switching works correctly when stacked with percentage-based caps.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
