### 2026-01-23 abbc627
test(stacking): add displace with units switch tests (Component 11)
Add six displace tests with cap and floor policies across three scenarios:
- Scenario A/D: Units/kg switching with triggered cap/floor
- Scenario B/E: Kg/units switching with triggered cap/floor
- Scenario C/F: Non-triggered cap/floor with displace and set

Tests validate that displace reports in same units as target substance,
and that cap/floor changes units only when policy is triggered.
