### 2026-01-23 006993a
test(stacking): add testRecycleAndCap for recycle and cap interaction (Component 12)
Tests interaction between recycling (recovery/reuse/induction) and cap policies when stacked together. Validates that policy application order affects final outcomes by comparing two scenarios: Recycling First (domestic recycling before sales permit cap) and Permit First (sales permit cap before domestic recycling). For HFC-134a, Permit First yields higher domestic kg than Recycling First due to differences in when the cap is applied relative to recycling additions.

Co-Authored-By: Claude Opus 4.5 <noreply@an
[…]
