+++ engine/src/test/java/org/kigalisim/validate/StackingLiveTests.java
@@ -1371,6 +1371,100 @@ public class StackingLiveTests {
   }
 
   /**
+   * Test interaction between recycling and cap policies when stacked.
+   * This validates that the order of applying recycling and cap policies affects the outcome.
+   * When recycling is applied first, it adds to the base (with 100% induction) before the cap
+   * is applied. When the cap is applied first, it reduces the amount before recycling adds to it.
+   * Therefore, "Permit First" should result in higher domestic kg than "Recycling First" because
+   * the cap operates on a larger base in the former case.
+   */
+  @Test
+  public void testRecycleAndCap() throws IOException {
+    String qtaPath = "../examples/stacking_recycle_and_cap.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    // Run Recycling First scenario (Domestic Recycling then Sales Permit)
+    Stream<EngineResult> recyclingFirstResults = KigaliSimFacade.runScenario(
+        program, "Recycling First", progress -> {});
+    List<EngineResult> recyclingFirstList = recyclingFirstResults.collect(Collectors.toList());
+    EngineResult recyclingFirst2030 = LiveTestsUtil.getResult(
+        recyclingFirstList.stream(), 2030, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(recyclingFirst2030, "Should have Recycling First result for 2030");
+    final double recyclingFirstDomestic = recyclingFirst2030.getDomestic().getValue().doubleValue();
+
+    // Run Permit First scenario (Sales Permit then Domestic Recycling)
+    Stream<EngineResult> permitFirstResults = KigaliSimFacade.runScenario(
+        program, "Permit First", progress -> {});
+    List<EngineResult> permitFirstList = permitFirstResults.collect(Collectors.toList());
+    EngineResult permitFirst2030 = LiveTestsUtil.getResult(
+        permitFirstList.stream(), 2030, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(permitFirst2030, "Should have Permit First result for 2030");
+    final double permitFirstDomestic = permitFirst2030.getDomestic().getValue().doubleValue();
+
+    // Run BAU scenario for reference
+    Stream<EngineResult> bauResults = KigaliSimFacade.runScenario(
+        program, "BAU", progress -> {});
+    List<EngineResult> bauList = bauResults.collect(Collectors.toList());
+    EngineResult bau2030 = LiveTestsUtil.getResult(
+        bauList.stream(), 2030, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(bau2030, "Should have BAU result for 2030");
+    final double bauDomestic = bau2030.getDomestic().getValue().doubleValue();
+
+    // Run Recycling scenario for reference
+    Stream<EngineResult> recyclingResults = KigaliSimFacade.runScenario(
+        program, "Recycling", progress -> {});
+    List<EngineResult> recyclingList = recyclingResults.collect(Collectors.toList());
+    EngineResult recycling2030 = LiveTestsUtil.getResult(
+        recyclingList.stream(), 2030, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(recycling2030, "Should have Recycling result for 2030");
+    final double recyclingDomestic = recycling2030.getDomestic().getValue().doubleValue();
+
+    // Run Permit scenario for reference
+    Stream<EngineResult> permitResults = KigaliSimFacade.runScenario(
+        program, "Permit", progress -> {});
+    List<EngineResult> permitList = permitResults.collect(Collectors.toList());
+    EngineResult permit2030 = LiveTestsUtil.getResult(
[…]
