### 2026-01-23 40279a4
test(stacking): add testCapRecycleDisplace for cap and recycling with displacement (Component 13)
Tests interaction between cap with displacement and recycling policies with five scenarios: BAU (10kg primary), Cap (8kg primary + 2kg displaced to secondary), Recycling (5kg primary with 100% recovery), Cap First (3kg primary after cap then recycling reduces recharge), Recycling First (5kg primary as recycling prevents cap trigger). Validates that displacement only occurs when cap triggers and recycling with 0% induction reduces recharge demand.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic
[…]
