+++ engine/src/test/java/org/kigalisim/validate/StackingLiveTests.java
@@ -1465,6 +1465,142 @@ public class StackingLiveTests {
   }
 
   /**
+   * Test interaction between cap with displacement and recycling policies.
+   * This validates that cap with displacement and recycling with 0% induction
+   * interact correctly, with displacement only occurring when the cap triggers.
+   * Recycling with 0% induction reduces recharge demand (displacement behavior),
+   * so it can prevent a cap from triggering if applied first.
+   */
+  @Test
+  public void testCapRecycleDisplace() throws IOException {
+    String qtaPath = "../examples/stacking_cap_recycle_displace.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    // Run BAU scenario
+    Stream<EngineResult> bauResults = KigaliSimFacade.runScenario(
+        program, "BAU", progress -> {});
+    List<EngineResult> bauList = bauResults.collect(Collectors.toList());
+    EngineResult bauHfc2 = LiveTestsUtil.getResult(
+        bauList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(bauHfc2, "Should have BAU HFC-134a result for year 2");
+    final double bauHfcDomestic = bauHfc2.getDomestic().getValue().doubleValue();
+
+    EngineResult bauR600a2 = LiveTestsUtil.getResult(
+        bauList.stream(), 2, "Domestic Refrigeration", "R-600a");
+    assertNotNull(bauR600a2, "Should have BAU R-600a result for year 2");
+    final double bauR600aDomestic = bauR600a2.getDomestic().getValue().doubleValue();
+
+    // Run Cap scenario
+    Stream<EngineResult> capResults = KigaliSimFacade.runScenario(
+        program, "Cap", progress -> {});
+    List<EngineResult> capList = capResults.collect(Collectors.toList());
+    EngineResult capHfc2 = LiveTestsUtil.getResult(
+        capList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(capHfc2, "Should have Cap HFC-134a result for year 2");
+    final double capHfcDomestic = capHfc2.getDomestic().getValue().doubleValue();
+
+    EngineResult capR600a2 = LiveTestsUtil.getResult(
+        capList.stream(), 2, "Domestic Refrigeration", "R-600a");
+    assertNotNull(capR600a2, "Should have Cap R-600a result for year 2");
+    final double capR600aDomestic = capR600a2.getDomestic().getValue().doubleValue();
+
+    // Run Recycling scenario
+    Stream<EngineResult> recyclingResults = KigaliSimFacade.runScenario(
+        program, "Recycling", progress -> {});
+    List<EngineResult> recyclingList = recyclingResults.collect(Collectors.toList());
+    EngineResult recyclingHfc2 = LiveTestsUtil.getResult(
+        recyclingList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(recyclingHfc2, "Should have Recycling HFC-134a result for year 2");
+    final double recyclingHfcDomestic = recyclingHfc2.getDomestic().getValue().doubleValue();
+
+    EngineResult recyclingR600a2 = LiveTestsUtil.getResult(
+        recyclingList.stream(), 2, "Domestic Refrigeration", "R-600a");
+    assertNotNull(recyclingR600a2, "Should have Recycling R-600a result for year 2");
+    final double recyclingR600aDomestic = recyclingR600a2.getDomestic().getValue().doubleValue();
+
[…]
