### 2026-01-23 bd642ea
docs(stacking): document units tracking behavior for cap, floor, change, and displace (Component 14)
Update llms-full.txt and qubectalk_commands.html to document how QubecTalk commands
interact with units-based versus volume-based tracking:
- Cap/Floor: preserve units tracking when triggered if last value was in units
- Change: preserves current tracking mode without forcing conversion
- Displace: operates in target stream's current units without mode change

Changes are minimal and surgical, focused on the specific behaviors validated by
Components 8-11 stacking tests.
