### 2026-01-23 df05d12
refactor(displacement): remove unnecessary comment in applyUnitsDisplacementDifferentSubstance
