### 2026-01-23 632eccf
test(stacking): add change and recycle stacking test
Add testChangeAndRecycle to StackingLiveTests with corresponding QTA file.
Tests interaction between change (-4% current) and recycle (50% with 0%
induction at recharge) policies when stacked.

Current observed behavior:
- Change Then Recycle: both effects accumulate (lowest domestic kg)
- Recycle Then Change: change overrides recycling effect (equals Change alone)

This documents the current stacking behavior for investigation.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
