+++ engine/src/test/java/org/kigalisim/validate/StackingLiveTests.java
@@ -1601,6 +1601,149 @@ public class StackingLiveTests {
   }
 
   /**
+   * Test interaction between change and recycle policies when stacked.
+   * This validates that the order of applying change and recycle policies affects the outcome.
+   *
+   * <p>Setup:
+   * <ul>
+   *   <li>domestic: 10 mt</li>
+   *   <li>priorEquipment: 1000 units</li>
+   *   <li>initial charge: 0.1 kg / unit</li>
+   *   <li>recharge: 10% with 0.1 kg / unit</li>
+   *   <li>BAU change: +5% current each year</li>
+   *   <li>Change Tax policy: -4% current each year</li>
+   *   <li>Recycle policy: 50% recovery with 100% reuse with 0% induction at recharge</li>
+   * </ul>
+   *
+   * <p>Expected outcomes in year 2:
+   * <ul>
+   *   <li>Change alone: Lower than BAU (negative change reduces growth)</li>
+   *   <li>Recycle alone: Lower than BAU (recycling with 0% induction reduces virgin demand)</li>
+   *   <li>Change then Recycle: Lower than both Change alone and Recycle alone
+   *       (both effects accumulate when recycle is applied after change)</li>
+   *   <li>Recycle then Change: Equal to Change alone (change recalculates from base,
+   *       effectively overriding the recycling effect on recharge)</li>
+   *   <li>Change then Recycle: Lower than Recycle then Change
+   *       (order matters: applying recycle last preserves both effects)</li>
+   * </ul>
+   */
+  @Test
+  public void testChangeAndRecycle() throws IOException {
+    String qtaPath = "../examples/stacking_change_and_recycle.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    // Run BAU scenario
+    Stream<EngineResult> bauResults = KigaliSimFacade.runScenario(
+        program, "BAU", progress -> {});
+    List<EngineResult> bauList = bauResults.collect(Collectors.toList());
+    EngineResult bau2 = LiveTestsUtil.getResult(
+        bauList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(bau2, "Should have BAU result for year 2");
+    final double bauDomestic = bau2.getDomestic().getValue().doubleValue();
+
+    // Run Change scenario
+    Stream<EngineResult> changeResults = KigaliSimFacade.runScenario(
+        program, "Change", progress -> {});
+    List<EngineResult> changeList = changeResults.collect(Collectors.toList());
+    EngineResult change2 = LiveTestsUtil.getResult(
+        changeList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(change2, "Should have Change result for year 2");
+    final double changeDomestic = change2.getDomestic().getValue().doubleValue();
+
+    // Run Recycle scenario
+    Stream<EngineResult> recycleResults = KigaliSimFacade.runScenario(
+        program, "Recycle", progress -> {});
+    List<EngineResult> recycleList = recycleResults.collect(Collectors.toList());
+    EngineResult recycle2 = LiveTestsUtil.getResult(
+        recycleList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(recycle2, "Should have Recycle result for year 2");
+    final double recycleDomestic = recycle2.getDomestic().getValue().doubleValue();
+
+    // Run Change Then Recycle scenario
+    Stream<EngineResult> changeThenRecycleResults = KigaliSimFacade.runScenario(
+        program, "Change Then Recycle", progress -> {});
+    List<EngineResult> changeThenRecycleList = changeThenRecycleResults.collect(Collectors.toList());
[…]
