### 2026-01-23 ed77a0f
test(stacking): update assertions to expected behavior (FAILING)
Update testChangeAndRecycle assertions to reflect expected stacking behavior:
- Recycle Then Change should be lower than Change alone
- Recycle Then Change should be lower than Change Then Recycle

Current behavior does not match expectations:
- Recycle Then Change equals Change alone (10160.64 kg)
- Change Then Recycle (9652.14 kg) is lower than Recycle Then Change

This suggests a potential bug where the change policy recalculates from
the base value, overriding the recycling effect when applied after recycle.

Co-Authored-By: 
[…]
