+++ beta_changelog.md
@@ -16,11 +16,19 @@ The following changes have been adopted and released.
 
 ### Manual ordering of policies in UI editor
 
-**Status**: Released January 23, 2026
+**Status**: Released January 26, 2026
 
 **Classification**: Enhancement
 
 Added ability to manually specify the ordering of policy application / policy stacking. This allows for order-dependent policy scenarios to be defined without requiring use of the code editor. For example, if a percent decrease (like from an educational program) should be applied before or after a cap (like from a permitting system). Further described in [#703](https://github.com/SchmidtDSE/kigali-sim/issues/703).
+
+### Fix basis for stacked change policy
+
+**Status**: Released January 26, 2026.
+
+**Classification**: Bug
+
+When stacking policies, it is possible for change statements with `% current` to use the original basis from BAU and not that set by a floor or cap. See [#711](https://github.com/SchmidtDSE/kigali-sim/pull/711). There is some ambiguity in what is desireable but, given explicit control of stacking order, this was fixed to use the cap or floor as basis (if cap or floor were violated) instead of the original value for `% current`.
 
 ### Alphabetical sorting in UI lists
 
