### 2026-01-24 b219a36
fix(change): account for recycling share in percentage change calculations
When stacking recycling with percentage changes, the change operation now
accounts for recycled material that displaces virgin production. The effective
base for percentage calculations includes the stream's proportional share of
recycling, ensuring percentage changes represent total demand changes rather
than just changes to the virgin production stream.

Modified applyStandardPercentageChange() to use an effective base that includes
the stream's recycling share, and added getStreamRecyclingShare() helper to
calculate 
[…]
