### 2026-01-24 ee68926
fix(change): return full recycling share for sales stream
When the stream is "sales" (combined domestic + import), return the
full recycling amount rather than zero.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
