### 2026-01-24 428a913
refactor(change): use switch expression and add sales stream test
- Convert if-else chain in getStreamRecyclingShare to switch expression
- Add testChangeAndRecycleSales test using sales stream instead of domestic
- Add stacking_change_and_recycle_sales.qta example file

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
