+++ engine/src/test/java/org/kigalisim/validate/StackingLiveTests.java
@@ -1738,6 +1738,130 @@ public class StackingLiveTests {
   }
 
   /**
+   * Test interaction between change and recycle policies when stacked, using sales stream.
+   * This is a variation of testChangeAndRecycle that uses "change sales" instead of
+   * "change domestic" in the policy to verify recycling share works for sales stream.
+   *
+   * <p>Setup is identical to testChangeAndRecycle except the Change Tax policy uses
+   * "change sales by -4 % current each year" instead of "change domestic".</p>
+   *
+   * <p>Expected outcomes should be the same as testChangeAndRecycle since sales = domestic + import
+   * and import is 0 in this test.</p>
+   */
+  @Test
+  public void testChangeAndRecycleSales() throws IOException {
+    String qtaPath = "../examples/stacking_change_and_recycle_sales.qta";
+    ParsedProgram program = KigaliSimFacade.parseAndInterpret(qtaPath);
+    assertNotNull(program, "Program should not be null");
+
+    // Run BAU scenario
+    Stream<EngineResult> bauResults = KigaliSimFacade.runScenario(
+        program, "BAU", progress -> {});
+    List<EngineResult> bauList = bauResults.collect(Collectors.toList());
+    EngineResult bau2 = LiveTestsUtil.getResult(
+        bauList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(bau2, "Should have BAU result for year 2");
+    final double bauDomestic = bau2.getDomestic().getValue().doubleValue();
+
+    // Run Change scenario
+    Stream<EngineResult> changeResults = KigaliSimFacade.runScenario(
+        program, "Change", progress -> {});
+    List<EngineResult> changeList = changeResults.collect(Collectors.toList());
+    EngineResult change2 = LiveTestsUtil.getResult(
+        changeList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(change2, "Should have Change result for year 2");
+    final double changeDomestic = change2.getDomestic().getValue().doubleValue();
+
+    // Run Recycle scenario
+    Stream<EngineResult> recycleResults = KigaliSimFacade.runScenario(
+        program, "Recycle", progress -> {});
+    List<EngineResult> recycleList = recycleResults.collect(Collectors.toList());
+    EngineResult recycle2 = LiveTestsUtil.getResult(
+        recycleList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(recycle2, "Should have Recycle result for year 2");
+    final double recycleDomestic = recycle2.getDomestic().getValue().doubleValue();
+
+    // Run Change Then Recycle scenario
+    Stream<EngineResult> changeThenRecycleResults = KigaliSimFacade.runScenario(
+        program, "Change Then Recycle", progress -> {});
+    List<EngineResult> changeThenRecycleList = changeThenRecycleResults.collect(Collectors.toList());
+    EngineResult changeThenRecycle2 = LiveTestsUtil.getResult(
+        changeThenRecycleList.stream(), 2, "Domestic Refrigeration", "HFC-134a");
+    assertNotNull(changeThenRecycle2, "Should have Change Then Recycle result for year 2");
+    final double changeThenRecycleDomestic = changeThenRecycle2.getDomestic().getValue().doubleValue();
+
+    // Run Recycle Then Change scenario
+    Stream<EngineResult> recycleThenChangeResults = KigaliSimFacade.runScenario(
+        program, "Recycle Then Change", progress -> {});
+    List<EngineResult> recycleThenChangeList = recycleThenChangeResults.collect(Collectors.toList());
+    EngineResult recycleThenChange2 = LiveTestsUtil.getResult(
[…]
