### 2026-01-24 b3a02c2
Clarify stacking change policy details in changelog
Clarified the classification and updated the explanation regarding stacking policies. Added references to related pull requests and issues, and noted changes made to documentation and tests.
