### 2026-01-24 #9377 @kthyng
@sampottinger Thank you for all your effort here! It is indeed a busy time with the changes here. Your submission will be considered under our previous scope guidelines, but it is fine that you updated the paper. I will work to find you an editor
