### 2026-01-26 #9377 @willgearty
:wave: @arbennett @ali-ramadhan @GregorBecker @rosecers, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. We ask for reviews to start in the next 2 weeks and aim to finish within 4-6 weeks.
