### 2026-01-27 825e0a8
Update change log
Clarified recent release in annotated change log.
