### 2026-01-27 6f79f0a
Add tutorial 11 expansion to changelog.
