### 2026-01-28 5e6cd17
fix(sales): add induced demand back when setting sales with recycling
When setting sales (e.g., from a cap), the code subtracted recycling
but never added back the induced demand. With 100% induction, this
caused virgin sales to be incorrectly reduced.

The fix ensures that after subtracting recycling, we add back the
induced amount based on the induction rates for each recovery stage.
This makes the cap correctly account for induction, so with 100%
induction the order of cap vs recycling policies no longer matters.

Also adds test for tutorial 06 induction basis and updates stacking
test to e
[…]
