+++ engine/src/test/java/org/kigalisim/validate/StackingLiveTests.java
@@ -1429,12 +1429,15 @@ public class StackingLiveTests {
     assertNotNull(permit2030, "Should have Permit result for 2030");
     final double permitDomestic = permit2030.getDomestic().getValue().doubleValue();
 
-    // Assert: Permit First should have higher domestic than Recycling First in 2030
-    // When recycling is applied first, the cap operates on the increased base (recycled added).
-    // When permit is applied first, the cap reduces before recycling adds to it.
-    assertTrue(permitFirstDomestic > recyclingFirstDomestic,
+    // Assert: With 100% induction, Permit First and Recycling First should be equal
+    // With 100% induction, recycling adds ON TOP of virgin sales (doesn't displace).
+    // The cap correctly accounts for induced demand, so the order doesn't matter:
+    // - Recycling First: recycling calculates amount, cap sets virgin (accounting for induction)
+    // - Permit First: cap sets virgin (accounting for induction), recycling adds on top
+    double orderTolerance = 0.1; // kg
+    assertEquals(permitFirstDomestic, recyclingFirstDomestic, orderTolerance,
         String.format(
-            "Permit First domestic (%.2f kg) should be higher than "
+            "With 100%% induction, Permit First domestic (%.2f kg) should equal "
             + "Recycling First domestic (%.2f kg) in 2030",
             permitFirstDomestic, recyclingFirstDomestic));
 
