### 2026-02-04 #9377 @rosecers
@willgearty, not mine either, sorry

> On Jan 28, 2026, at 05:05, Gregor Becker ***@***.***> wrote:
>
>
> GregorBecker
>  left a comment
> (openjournals/joss-reviews#9377)
>  <https://github.com/openjournals/joss-reviews/issues/9377#issuecomment-3810641090>
> Hi @willgearty <https://github.com/willgearty>, not my field, sorry.
>
