### 2026-02-04 #9377 @sampottinger
Hello @willgearty ! I hope this is acceptable but I found some more options for your consideration.

Awaiting response:

 - [Andrew Bennett](https://reviewers.joss.theoj.org/reviewers/986): `@arbennett`
 - [Ali Ramadhan](https://reviewers.joss.theoj.org/reviewers/700): `@ali-ramadhan`

For your consideration:

 - [Ahanaf Aziz](https://reviewers.joss.theoj.org/reviewers/23836): `@Ahanaf-Aziz` (already reviewing another submission)
 - [Lionel Constantin](https://reviewers.joss.theoj.org/reviewers/4913): `@lionel42` (already reviewing another submission)
 - [Jack Leland](https://reviewers.joss.theoj.org/reviewers/17303): `@jackleland` (already reviewing another submission)
 - ~~[Angela Rossi](https://reviewers.joss.theoj.org/reviewers/20208): `@rossiangela`~~
 - [Dominic Orchard](https://reviewers.joss.theoj.org/reviewers/7466): `@dorchard`
 - [Chirag Hareshkumar Shah](https://reviewers.joss.theoj.org/reviewers/3749): `@shahchiragh` (already reviewing 2 other submissions)
 - [Fred Shone](https://reviewers.joss.theoj.org/reviewers/3393): `@fredshone`

Declined:

 - [Gregor Becker](https://reviewers.joss.theoj.org/reviewers/2752): `@GregorBecker`
 - [Rosecers](https://reviewers.joss.theoj.org/reviewers/1672) : `@rosecers`

Thank you very much again for your time and help.
