### 2026-02-05 #9377 @fredshone
Hello, looks slick, there has clearly been a lot of effort by someone or thing 🤷🏻‍♂️ , so I think it will sail through the SE criteria.

However, I am not at all close to modelling pollutants, so (for example) specifying chemicals for consumption is new to me, and generally commenting on the overall value of the work is tricky.

Happy to review. But perhaps a second reviewer should be closer to this domain?
