### 2026-02-05 #9377 @fredshone
@sampottinger is there an original JS implementation somewhere?
