### 2026-02-05 #9377 @sampottinger
Hello @fredshone! Thank you for your time and help as well as your warm response. Also, kind of you to inquire about the prior JS implementation which can be found back in Git history at https://github.com/SchmidtDSE/kigali-sim/tree/74b5b30f9a060f8fa1d4ff1165fc54d144c9495f. I should mention that the software has grown quite a bit since then. We transitioned to Java / WASM right before we began an extended period of testing and co-design with some countries and supporting organizations. Even so, you can probably see how much of that original structure and core ideas carried over into the current implementation.
