### 2026-02-09 #9377 @willgearty
:wave: @lionel42 I know you are already reviewing another submission for JOSS, but would you be able to review this one as well? We're particularly looking for someone with expertise in emissions and pollutants, so you would be perfect. However, if you're not available, would you know of anyone else that might be able to review this submission?
