### 2026-02-23 #10104 @editorialbot
[REVIEW]: Kigali Sim: Open source simulation toolkit for modeling substances and policies related to the Montreal Protocol
**Submitting author:** @sampottinger (<a href="http://orcid.org/0000-0002-0458-4985">A Samuel Pottinger</a>)
**Repository:** https://github.com/SchmidtDSE/kigali-sim
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1.8
