### 2026-03-03 #10104 @lionel42
Hello honor for me as well,

As a starting comment, I started reading the paper and checked the sources, but already after reading the statement of need section, I realized you have issues with the hyper links of the sources.

Some sources don't have hyperlinks at all, if you cite a study you should at least put the doi. Others cite to non specific documents and some are even dead links.

Can you fix this ?
