### 2026-03-03 #10104 @lionel42
@sampottinger  Can you also check https://kigalisim.org/ ? I tried to follow the quickstart steps,  I am able to download the results, but nothing is shown on the graphs
