### 2026-03-03 SchmidtDSE/montreal-protocol#757 @sampottinger
Update quick start code to show more dynamic line
Users typically start with [the tutorials](https://kigalisim.org/guide/tutorial_02.html) which use 25 tonnes (mt) of substance. In [Chart.JS](https://www.chartjs.org/), the auto scale feature leaves some room above the line so it is very apparent. This is visible in the [Tutorial 2 results video](https://kigalisim.org/guide/tutorial_02.html#interpreting-results), for example.

However, the current [quick start](https://github.com/SchmidtDSE/kigali-sim?tab=readme-ov-file#quickstart) code uses 1 mt which does render correctly in the application but it appears at the top of the screen so may be less visually apparent. **Please see the attached screenshot.** This could lead to a poor experience for someone using the quick start instead.

I apologize for this issue @lionel42 and thank you for raising it. I am conversing with my co-authors this morning but, because the long form tutorials have more room (and videos!) to explain this to the user, I may suggest that the quick start have a `change` statement to make the line both more visually apparent and to also better demonstrate features of the tool. I hope to resolve this shortly.

This is related to https://github.com/openjournals/joss-reviews/issues/10104.

![Image](https://github.com/user-attachments/assets/9f6fbaa3-d9ad-430b-ad42-cfb52b0c37ee)
