### 2026-03-03 SchmidtDSE/montreal-protocol#758 @sampottinger
Double check links from paper
Thank you @lionel42 for your help and guidance within https://github.com/openjournals/joss-reviews/issues/10104. This is in response to the issue about links from our paper.

I reviewed our links. My very sincere apologies as I did find that https://ozone.unep.org/system/files/documents/OEWG/%20KIP/%20Side/%20Event/%20MLFS_070824.pdf no longer resolves. My apologies for this issue. We will update accordingly. All of the other links did resolve for me but I will ask on https://github.com/openjournals/joss-reviews/issues/10104 if any others were giving you trouble.

There are 4 resources which do not have links that I will add shortly:

 - The large contribution of projected HFC emissions to future climate forcing: https://www.pnas.org/doi/10.1073/pnas.0902817106 (DOI [10.1073/pnas.0902817106](https://doi.org/10.1073/pnas.0902817106))
 - Success of Montreal Protocol Demonstrated by Comparing High-Quality UV Measurements with "World Avoided" Calculations from Two Chemistry-Climate Models: https://pubmed.ncbi.nlm.nih.gov/31481668/ (DOI [10.1038/s41598-019-48625-z](https://doi.org/10.1038/s41598-019-48625-z))
 - https://report.opensustain.tech/chapters/index.html (DOI [10.5281/zenodo.7771632](https://doi.org/10.5281/zenodo.7771632) but let's include both DOI and direct link)
 - http://sciencedirect.com/science/article/pii/S163107131830124X?__cf_chl_tk=90Ww7vDbvGvD5mMm3xlWkJhMI.FhPyC3K8nTPL.HaHw-1772557381-1.0.1.1-_99GVBFTBB5SaZegeW7nvZOyJiO5gtKOre6tB2LY83w (DOI [10.1016/j.crte.2018.09.002](https://doi.org/10.1016/j.crte.2018.09.002))

My apologies for this inconvenience. I wonder if the DOI got cut off for these 4 when I recently removed redundant links (some entries had both a link and DOI causing two links to be rendered in the JOSS paper engine).
