### 2026-03-03 3ac5947
Improve quick start for visibility and features.
In response to @lionel42 feedback, modifying the quickstart mentioned in the README to have a line changing over time. This serves two purposes:

 - Improved visibility for the line itself depending on chartjs auto scale feature which may have the line visibile at the top of the chart.
 - Demonstration of the `change` command in addition to `set` in case users are not taking advantage of the [the tutorials](https://kigalisim.org/guide/).

Closes #757. As this was made with feedback from @lionel42, listing as coauthor with gratitude.

Co-authored
[…]
