### 2026-03-03 #10104 @sampottinger
Humm this might be failing due to [issues ongoing with GitHub](https://www.githubstatus.com/) but it does generate using the JOSS action in the repo itself. I’ll try again to regenerate the pdf shortly when the GitHub actions outage is resolved.
