### 2026-03-03 #10104 @sampottinger
Hello @lionel42! Thank you very much for your time and your feedback. My apologies for the delay in my reply due to the [GitHub outage](https://www.githubstatus.com/). Please see below for both of these comments.

**Quick start**: I opened https://github.com/SchmidtDSE/kigali-sim/issues/757. I believe the graph may be rendering but the line is at the top of the chart (see [issue](https://github.com/SchmidtDSE/kigali-sim/issues/757) for further explanation). Please let me know if that is consistent with what you were seeing.

My apologies, I typically hear from users who go through [the tutorials](https://kigalisim.org/guide/) (where there’s a video and additional explanatory text) instead of the quick start. I am sorry and agree that can be confusing or easy to miss. In discussion with my co-authors this morning, we landed on the change at https://github.com/SchmidtDSE/kigali-sim/pull/759.

I have you listed as a co-author in recognition of your help in identifying this issue. I had a spot check this morning but please let me know if it is OK to merge with you included or, if you prefer, I can remove the co-author attribution. Thanks!

**Citations**: I opened and have resolved https://github.com/SchmidtDSE/kigali-sim/issues/758. My apologies, I found the broken link due to an escaping issue in my `paper.bib`. That said, I am very sorry but I only found that one link as broken going through all of the citations. Was there another? Finally, my sincere apologies and the 4 citations without links are resolved (I left ISBN  as is as it looks like JOSS builds the link on its own). Those were inadvertently removed when I tried to fix an unrelated rendering issue previously. https://github.com/SchmidtDSE/kigali-sim/pull/760 has been merged.

**Thanks**: I appreciate your thoughts and your care here. Already making things better!
