### 2026-03-04 SchmidtDSE/montreal-protocol#761 @lionel42
Paper review comments
This is linked to the review https://github.com/openjournals/joss-reviews/issues/10104

In general, I find the writing style a bit complicated to read. (long sentences, not going directly to the point)
Can you try to re-read it and improve some of the formulations ?

## General questions

* Is the software also adapted to other ghgs ? methane, HFOs ? Is this something we should mention in the paper ?

## state of the field
*  Would be nice to also add the name of the models in the text, not just as reference.
* check the english in this section, the text is a bit complicated to understand.

## research impact statement
* you could develop a bit more on the history of the tool, instead of putting all references at one (ex. the tool was requested by ... in ..., it was then designed by ..., finally implemented by ..., now maintained by community ...)
* `Simulating emissions, energy, consumption, equipment populations, trade,` out of scope for this section, should be in the tool description
* It seems that the effort is quite broad, but I see only you as developer on github. This is a bit odd in a community project. Can you explain how that is ?

## Software design: flexible engine

* `from trade records to industry census data to observed emissions`  : from ... to ... to` is weird formulation
* Figure 1: improve this figure by coloring boxes based on user policy input/ data / calculated quantities/ ...

## Dual-interface design
* first sentence is too long and not clear
* Since you go in details about the 2 kinds of authoring later, this section could be made smaller to avoid duplication

## UI based authoring

* I like that you cite video game development sources, however, I feel that there should be more appropriate sources for software development principles (section on code based authoring is much better in that regard)
* You cite multiple times previous work of your own about the SW principles. Here again, I find it would be more appropriate to cite specific sources on those developement principles.

## Code based authoring
* Can you specify that you developped your own DSL QubecTalk specifically for this project ? (if I understood correctly)

## Limitations
* footnote 1 and 2: you can put this directly in the text instead of footnotes

## Implementation
* I would move this section at the top of the SW design section, because it gives a history of the SW and it feels more natural in the progression.
* you mention JVM, but citation Konsin,2024 is about TeaVM, why this confusion ?
* Figure 3: can you add some part to explain that after the results are calculated they are shown or displayed in the UI ?

## Thanks section
* There it would be clearer if you enumerate each tool one by one, with the appropriate citation. A list of name and citations as it is now, is really hard to read.

## Funding
* make a full sentence instead of just `no conflicts`

## AI usage disclosure
* same as for Thanks section but with the AI tools.
