+++ paper/paper.bib
@@ -1,46 +1,12 @@
 @article{pyafscgap, doi = {10.21105/joss.05593}, url = {https://doi.org/10.21105/joss.05593}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {86}, pages = {5593}, author = {Pottinger, A Samuel and Zarpellon, Giulia}, title = {Pyafscgap.org: Open source multi-modal Python-based tools for NOAA AFSC RACE GAP}, journal = {Journal of Open Source Software} }
 
-@article{montreal_protocol_success,
-  title     = "Success of Montreal Protocol demonstrated by comparing
-               high-quality {UV} measurements with ``world avoided''
-               calculations from two chemistry-climate models",
-  author    = "McKenzie, Richard and Bernhard, Germar and Liley, Ben and
-               Disterhoft, Patrick and Rhodes, Steve and Bais, Alkiviadis and
-               Morgenstern, Olaf and Newman, Paul and Oman, Luke and Brogniez,
-               Colette and Simic, Stana",
-  abstract  = "The Montreal Protocol on Substances that Deplete the Ozone Layer
-               has been hailed as the most successful environmental treaty ever
-               (
-               https://www.unenvironment.org/news-and-stories/story/montreal-protocol-triumph-treaty
-               ). Yet, although our main concern about ozone depletion is the
-               subsequent increase in harmful solar UV radiation at the Earth's
-               surface, no studies to date have demonstrated its effectiveness
-               in that regard. Here we use long-term UV Index (UVI) data
-               derived from high-quality UV spectroradiometer measurements to
-               demonstrate its success in curbing increases in UV radiation.
-               Without this landmark agreement, UVI values would have increased
-               at mid-latitude locations by approximately 20\% between the
-               early 1990s and today and would approximately quadruple at
-               mid-latitudes by 2100. In contrast, an analysis of UVI data from
-               multiple clean-air sites shows that maximum daily UVI values
-               have remained essentially constant over the last ~20 years in
-               all seasons, and may even have decreased slightly in the
-               southern hemisphere, especially in Antarctica, where effects of
-               ozone depletion were larger. Reconstructions of the UVI from
-               total ozone data show evidence of increasing UVI levels in the
-               1980s, but unfortunately, there are no high-quality UV
-               measurements available prior to the early 1990s to confirm these
-               increases with direct observations.",
-  journal   = "Sci. Rep.",
-  publisher = "Springer Science and Business Media LLC",
-  volume    =  9,
-  number    =  1,
-  pages     = "12332",
-  month     =  sep,
-  year      =  2019,
-  doi       = {10.1038/s41598-019-48625-z},
-  copyright = "https://creativecommons.org/licenses/by/4.0",
-  language  = "en"
+@misc{montreal_protocol_success,
+  title     = "The {Montreal} {Protocol}: triumph by treaty",
+  author    = "Molina, Mario and Zaelke, Durwood",
+  year      = {2017},
+  month     = nov,
+  publisher = {{United Nations Environment Programme}},
+  url       = {https://www.unep.org/news-and-stories/story/montreal-protocol-triumph-treaty}
 }
 
 @misc{ozone,
