### 2026-03-04 7f86672
Exchange McKenzie et al for Molina and Zaelke.
Based on feedback from @lionel42 as seen in https://github.com/SchmidtDSE/kigali-sim/issues/758#issuecomment-3995905609, we exchanged an indirect citation for a more direct quote from a UNEP published article.
