### 2026-03-04 05198a5
Clarify non-treaty greenhouse gasses.
Related to feedback at #761, clarifying that substances outside of the Montreal Protocol are out of scope for Kigali Sim due to differing definitions / norms caused by different controlling international law.
