+++ paper/paper.md
@@ -61,7 +61,7 @@ To support beginning programmers, the UI-based point-and-click editor acclimates
 We find some simulations' complexity cumbersome in UI-based authoring. Therefore, our QubecTalk domain-specific language facilitates expression of complex models in human-readable syntax inspired by but distinct to HyperTalk [@smallhypertalk]. Mirrored by the UI-editor, QubecTalk speaks in treaty terminology, translating terms of art into simulations. This also supports uncertainty quantification, conditional logic, and policy stacking. With optional AI assistant compatibility via the llms.txt specification [@llmstxt], users may author scripts in a web-based programming portal [@portal] or outside with direct JVM invocation.
 
 ## Limitations and future work
-Kigali Sim can model energy consumption but only estimates direct emissions^[Users may calculate indirect emissions using outside energy mix data [@energy_mix].]. Additionally, it applies treaty trade attribution but will only attribute charge prior to equipment sale entirely to either the importer or exporter^[Local assembly can be modeled as domestic production.]. However, it will receive updates as official guidance changes in the future [@mlf_guidelines].
+Kigali Sim can model energy consumption but only estimates direct emissions^[Users may calculate indirect emissions using outside energy mix data [@energy_mix].]. Additionally, it embodies Montreal Protocol definitions and rules such as initial charge to exporter and it will receive updates as official guidance changes in the future [@mlf_guidelines]. Therefore, substances not subject to Monteral Protocol accounting norms are out of scope.
 
 # Implementation
 Migrated from an original JavaScript implementation for performance and portability, Kigali Sim runs via WASM^[Zero-install browser-based.] or JVM [@teavm]. Without transmitting simulations to external servers, both enable privacy-preserving parallel local computation with BigDecimal [@bigdecimal] after ANTLR [@antlr] interpretation.
