### 2026-03-04 69161cb
Clarify state of field.
As part of #761, name HFC Outlook and clarify the connection to the NOUs. Minor phrasing clean up.
