### 2026-03-04 ac586b6
Improvements in structure in paper.
Incorporating improvements in structure regarding discussion of functionality. THis is part of #761.
