### 2026-03-04 a19cb3f
Improvements in structure in paper.
Incorporating improvements in structure regarding discussion of functionality. THis is part of #761.
