### 2026-03-04 9056c83
Add contribution section.
Added contribution section as part of #761. This further highlights the contribution of the project and describes how the project came about. Also, we clarify ongoing funding.
