+++ paper/paper.bib
@@ -136,6 +136,26 @@
   note = {Technical paper}
 }
 
+@misc{ciera,
+  author = {Martinez, Ciera},
+  title = {Data, Interactive Tools, and {AI} for Actionable Change: Ending Plastic Pollution},
+  year = {2024},
+  url = {https://dse.berkeley.edu/news/data-interactive-tools-and-ai-actionable-change-ending-plastic-pollution},
+  note = {DSE Berkeley News}
+}
+
+@article{plasticssci,
+  author = {A. Samuel Pottinger and Roland Geyer and Nivedita Biyani and Ciera C. Martinez and Neil Nathan and Molly R. Morse and Chao Liu and Shanying Hu and Magali de Bruyn and Carl Boettiger and Elijah Baker and Douglas J. McCauley},
+  title = {Pathways to reduce global plastic waste mismanagement and greenhouse gas emissions by 2050},
+  journal = {Science},
+  volume = {386},
+  number = {6726},
+  pages = {1168--1173},
+  year = {2024},
+  doi = {10.1126/science.adr3837},
+  url = {https://www.science.org/doi/abs/10.1126/science.adr3837}
+}
+
 @misc{plastics,
   title={Combining Game Design and Data Visualization to Inform Plastics Policy: Fostering Collaboration between Science, Decision-Makers, and Artificial Intelligence}, 
   author={A Samuel Pottinger and Nivedita Biyani and Roland Geyer and Douglas J McCauley and Magali de Bruyn and Molly R Morse and Neil Nathan and Kevin Koy and Ciera Martinez},
