### 2026-03-05 259d09a
Minor word tweak related to #761.
