### 2026-03-05 af84f96
Minor word tweak related to #761.
