+++ paper/paper.md
@@ -39,7 +39,7 @@ Signed by all UN member states, the Montreal Protocol phased out 99% of ozone-de
 Currently, HFC Outlook offers a powerful but closed source model [@hfcoutlook]. Otherwise, organizations typically build private complex ad-hoc models [@complex].
 
 ## Contribution
-Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. The Multilateral Fund for the Implementation of the Montreal Protocol (MLF) approached the University of California Schmidt Center for Data Science and Environment (DSE) during a workshop featuring our interactive simulation tool which sought to support policy design for a global plastics pollution treaty [@plasticssci]. Co-design sessions sought to employ that project's game design-inspired techniques [@plastics] to Montreal Protocol opportunities in providing reusable software serving the science and policy communities. Those sessions identified needs in modeling emissions, energy, consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines]. Designed jointly by the MLF and the DSE, Kigali Sim was implemented by DSE in consultation with users from various governments and supporting organizations [@credits].
+Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. The Multilateral Fund for the Implementation of the Montreal Protocol (MLF) approached the University of California Schmidt Center for Data Science and Environment (DSE) during a workshop featuring our interactive simulation tool which sought to support policy design for a global plastics pollution treaty [@plasticssci]. Co-design sessions sought to employ that project's game design-inspired techniques [@plastics] to Montreal Protocol opportunities in providing reusable software serving the science and policy communities. Those sessions identified needs in modeling emissions, energy, consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines]. Designed jointly by the MLF and the DSE, Kigali Sim was implemented by DSE in collaboration with users from various governments and supporting organizations [@credits].
 
 ## Research impact statement
 Kigali Sim serves Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations. We particularly focus on National Ozone Units (NOUs), espeically those with limited resources [@noo] for whom developing ad-hoc models may be burdensome. More than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year. This effort spanned scientists, analysts, and policy-makers. Participating in many international meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge contributing to this community project [@credits] which also received media coverage [@nbc].
