### 2026-03-05 697e3e4
Move footnotes into body.
Move the footnotes into body and reowrk citation location for improved clarity as a result. Part of #761.
