+++ paper/paper.md
@@ -64,10 +64,10 @@ To support beginning programmers, the UI-based point-and-click editor acclimates
 Many Kigali Sim users do not identify as programmers and, reflecting empirical findings that domain experts with "limited programming knowledge" benefit from DSLs [@dsl_evaluation], we created the QubecTalk domain-specific language to facilitate expression of complex models in human-readable syntax inspired by but distinct to HyperTalk [@smallhypertalk]. Mirrored by the UI-editor, QubecTalk speaks in treaty terminology, translating terms of art into simulations. This also supports uncertainty quantification, conditional logic, and policy stacking. With optional AI assistant compatibility via the llms.txt specification [@llmstxt], users may author scripts in a web-based programming portal [@portal] or outside with direct JVM invocation.
 
 ## Limitations and future work
-Kigali Sim can model energy consumption but only estimates direct emissions^[Users may calculate indirect emissions using outside energy mix data [@energy_mix].]. Additionally, it embodies Montreal Protocol definitions and rules such as initial charge to exporter and it will receive updates as official guidance changes in the future [@mlf_guidelines]. Therefore, substances not subject to Montreal Protocol accounting norms are out of scope.
+Kigali Sim can model energy consumption but only estimates direct emissions. Users may calculate indirect emissions using outside energy mix data [@energy_mix]. Additionally, it embodies Montreal Protocol definitions and rules such as initial charge to exporter and it will receive updates as official guidance changes in the future [@mlf_guidelines]. Therefore, substances not subject to Montreal Protocol accounting norms are out of scope.
 
 # Implementation
-Migrated from an original JavaScript implementation for performance and portability, Kigali Sim runs via WASM^[Zero-install browser-based.] or JVM [@teavm]. Without transmitting simulations to external servers, both enable privacy-preserving parallel local computation with BigDecimal [@bigdecimal] after ANTLR [@antlr] interpretation.
+Migrated from an original JavaScript implementation for performance and portability, Kigali Sim runs via WASM [@teavm] which is browser-based without requiring install or via JVM. Without transmitting simulations to external servers, both enable privacy-preserving parallel local computation with BigDecimal [@bigdecimal] after ANTLR [@antlr] interpretation.
 
 ![Diagram describing multi-modal execution in which simulations run across different platforms.\label{fig:execution}](KigaliExecution.svg){width="100%"}
 
