### 2026-03-05 ee9c4fd
Move implemetntation to before design.
Move the implementation section to be before software design. As part of #761.
