### 2026-03-05 8d46d55
Move implemetntation to before design.
Move the implementation section to be before software design. As part of #761.
