### 2026-03-05 79861be
Expand figure 1 input / output description.
Expand the input and output figure caption to clarify that these values can be either inputs or otuputs. Part of #761.
