### 2026-03-05 d034c9c
Explain Hayashida sequence.
Make the hayashida sequence explicit in the paper, responding to #761.
