### 2026-03-05 78622c6
Explain Hayashida sequence.
Make the hayashida sequence explicit in the paper, responding to #761.
