### 2026-03-05 4e27c9e
Expand out thanks section.
Enumerate each open source package and AI provider in a bulleted list. Part of #761.
