### 2026-03-05 1d4b01e
Expand out thanks section.
Enumerate each open source package and AI provider in a bulleted list. Part of #761.
