### 2026-03-05 571a3ac
Clean up for clarity and brevity.
Continuing to respond to #761, complete re-read and trimming to maintain JOSS guidance on length.
