+++ paper/paper.md
@@ -30,22 +30,22 @@ bibliography: paper.bib
 ---
 
 # Summary
-Actively used internationally, Kigali Sim offers stock and flow modeling of Montreal Protocol-controlled substances. Focusing on the Kigali Amendment which reduces hydrofluorocarbons that contribute to climate change through warming potentials up to thousands of times greater than carbon dioxide, this open source platform aids both atmospheric science and policy development. Supporting a diverse community with heterogeneous programming expertise, this parallelized toolkit interoperates visual no-code rapid model development and a domain-specific language (DSL), democratizing advanced computational tools. Through WebAssembly (WASM) or Java Virtual Machine (JVM), Kigali Sim offers portable, private, and rigorous simulation in support of history's most successful international environmental treaty.
+Kigali Sim offers stock and flow modeling of Montreal Protocol-controlled substances. This includes hydrofluorocarbons that contribute to climate change through warming potentials much greater than carbon dioxide. Supporting a diverse community with heterogeneous programming expertise, this parallelized toolkit democratizes advanced computational tools by interoperating visual no-code rapid model development and a domain-specific language (DSL). Either through WebAssembly (WASM) or Java Virtual Machine (JVM), Kigali Sim affords portable, private, and rigorous simulation in support of history's most successful international environmental treaty, aiding both atmospheric science and policy development.
 
 # Statement of Need
-Signed by all UN member states, the Montreal Protocol phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, research and policy analysis for these HFCs involves modeling complex economic, technological, energy, and policy interactions [@complex].
+Signed by all UN member states, the Montreal Protocol phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, research and policy analysis for these substances involves modeling complex economic, technological, energy, and policy interactions [@complex].
 
 ## State of field
-Currently, HFC Outlook offers a powerful but closed source model [@hfcoutlook]. Otherwise, organizations typically build private complex ad-hoc models [@complex].
+On the current public market, only the proprietary HFC Outlook offers a full lifecycle reusable model for the Monteral Protocol [@hfcoutlook]. Otherwise, organizations typically build private ad-hoc models [@complex].
 
 ## Contribution
-Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. The Multilateral Fund for the Implementation of the Montreal Protocol (MLF) approached the University of California Schmidt Center for Data Science and Environment (DSE) during a workshop featuring our interactive simulation tool which sought to support policy design for a global plastics pollution treaty [@plasticssci]. Co-design sessions sought to employ that project's game design-inspired techniques [@plastics] to Montreal Protocol opportunities in providing reusable software serving the science and policy communities. Those sessions identified needs in modeling emissions, energy, consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines]. Designed jointly by the MLF and the DSE, Kigali Sim was implemented by DSE in collaboration with users from various governments and supporting organizations [@credits].
[…]
