### 2026-03-05 e9f08db
Clean up for clarity and brevity.
Continuing to respond to #761, complete re-read and trimming to maintain JOSS guidance on length.
