### 2026-03-05 SchmidtDSE/kigali-sim#761 @sampottinger
Thank you very much @lionel42 for your time and care in helping us further improve our work. Our most sincere gratitude for this detailed feedback which we have worked to incorporate as described below. I have connected with some of my other authors and hope to merge #763 shortly. CC @iamciera.

> Is the software also adapted to other ghgs?

Thank you for this generous question as we think about the future of this work. The controlling regulation and norms which influence analysis are different for different classes of substances, typically due to international law designation. So, while in theory Kigali Sim could be adapted to other greenhouse gasses, we are employing trade attribution, established GWPs, and definitions which are provided by the Montreal Protocol for the substances it controls (and as used by related scholarly bodies like the [TEAP](https://ozone.unep.org/science/assessment/teap)). We reference “Montreal Protocol-controlled substances” as the classes of substances under the purview of the treaty have grown over time.

We apologize for this confusion and recognize how parts of our paper may have failed to address this clearly. We have made modifications to further clarify this point.

> Would be nice to also add the name of the models in the text, not just as reference.

Our appreciation for helping us make this section of our paper more useful and specific. We have expanded this section but I did want to clarify that the only other reusable model we are aware of that incorporates treaty conventions is HFC Outlook as cited. This is a proprietary Excel / VBA-based model.

> check the english in this section, the text is a bit complicated to understand.

My apologies and thank you for flagging this. I believe some parts of the paper are a little dense. This was done in order to respect the JOSS paper length guidance. However, we agree that some parts were confusing. I have tried to simplify sentence structure here and invite further feedback.

> you could develop a bit more on the history of the tool, instead of putting all references at one

Thank you for this invitation to expand. We have included a new section in response to your comments.

> `Simulating emissions, energy, consumption, equipment populations, trade,` out of scope for this section, should be in the tool description

Thank you for helping us improve the structure of our work and this has been refined.

> It seems that the effort is quite broad, but I see only you as developer on github. This is a bit odd in a community project. Can you explain how that is ?

Thank you for inquiring! Though I am one of a small number of folks who have actually pushed to GitHub, we are very lucky to have many collaborators and active users providing labor into the open source process as further described on our [our credits page](https://kigalisim.org/guide/credits.html). In particular, some of the users in the community from various governments and organizations helped us with invaluable feedback and testing to get to this point. We regret that the GitHub contributors pulse doesn’t show their work when it is outside of programming itself. Even so, we still try to highlight them on the website when given permission.

I should say that we have been engaged in the effort since the summer of 2024 and you are actually seeing the second implementation (the prior implementation was discussed in our pre-review at https://github.com/openjournals/joss-reviews/issues/9377). For many in this community, this is their first time programming which is part of why we invested in the DSL and the web-based interface as well as those applications of game design concepts for “silent tutorialization” and progression to programming over time (as distinct from but related to [our plastics work](https://global-plastics-tool.org/), see [Pottinger et al 2024](https://www.science.org/doi/10.1126/science.adr3837) and [Pottinger et al 2026](https://arxiv.org/abs/2312.11359v3)).

Funding for the current team is ongoing to maintain this project though, through continued partnerships with users, we hope to foster more contributors!

> `from trade records to industry census data to observed emissions` : `from ... to ... to` is weird formulation

We agree and have revised.

> Figure 1: improve this figure by coloring boxes based on user policy input/ data / calculated quantities/ ...

Apologies for the confusion here and we have tried to improve the caption of the figure to better describe its role.

As seen in [the user’s guide](https://kigalisim.org/guide/), one can provide information on number of units of equipment, trade volumes, measured emissions, and others. We’ve even see someone do it through maintenance records! All this means that, in practice, different countries will have different parts of this picture (sometimes even heterogenous for different substances). Therefore, all of these actually can be inputs or outputs (if you specify trade data on amount of substance imported, we calculate how many units of equipment that likely indicates or, if specifying number of units, we can calculate the expected demand). In other words, based on that user feedback, we have different entry points and the missing values are determined.

> first sentence is too long and not clear … - Since you go in details about the 2 kinds of authoring later, this section could be made smaller to avoid duplication

Thank you for that feedback and have proposed removing most of it to help with compliance with paper length guidance.

> I like that you cite video game development sources, however, I feel that there should be more appropriate sources for software development principles (section on code based authoring is much better in that regard)

> You cite multiple times previous work of your own about the SW principles. Here again, I find it would be more appropriate to cite specific sources on those developement principles.

We appreciate you highlighting this opportunity to refine this section. We are employing a Hayashida sequence in the UI-based editor which is a very specific form of progressive disclosure which worked well in the plastics project to create an “invisible” tutorial (see [Pottinger et al 2026](https://arxiv.org/abs/2312.11359v3)) and this tool’s design was based on that project specifically.

That said, we cite the game design resources because, while this is increasingly common in interactive media for entertainment, a review of the literature finds that it is still not common in non-entertainment settings. Indeed, we cite [Pottinger and Zerpellon](https://joss.theoj.org/papers/10.21105/joss.05593) because it remains one of the few places we know of in the literature to apply it in a participatory data science context.

Even so, we understand how the brevity of this section buries this connection and we have modified this section to better describe these patterns. With mention of the plastics tool in the contribution section, we also streamlined the citations here.

> Can you specify that you developped your own DSL QubecTalk specifically for this project ? (if I understood correctly)

We appreciate your double checking this point. Yes, we did make our own DSL and have clarified in the paper.

> footnote 1 and 2: you can put this directly in the text instead of footnotes

Thank you and we agree that this is better structured. This is resolved in the paper.

> you mention JVM, but citation Konsin,2024 is about TeaVM, why this confusion ?

Our apologies for the confusion. The TeaVM citation is about how the JVM is made available through WASM. We have moved the location of the citation in that sentence for improved clarity.

> I would move this section at the top of the SW design section, because it gives a history of the SW and it feels more natural in the progression.

Thank you for helping to improve the flow of the paper. We have incorporated this feedback.

> Figure 3: can you add some part to explain that after the results are calculated they are shown or displayed in the UI ?

My apologies as I was unsure if you meant in the caption or figure itself. I have added to the caption.

> There it would be clearer if you enumerate each tool one by one, with the appropriate citation. A list of name and citations as it is now, is really hard to read.

> same as for Thanks section but with the AI tools.

Thank you for this feedback to further improve the readability of these sections. We have incorporated your suggestions.

> make a full sentence instead of just `no conflicts`

Thank you for drawing our attention to this and we have expanded this.

P.S. We are still trying to adhere to the pre-2026 paper length guidelines since we are tagged pre-2026. However, if needed, we will discuss with the editor. It appears the JOSS editors have increased the allowed length for post-2026 papers, corresponding to the recent journal change of scope (see https://github.com/openjournals/joss/commit/ac183457b4078928f2ea2a44442d10e119e0d15e)
