### 2026-03-12 #10104 @willgearty
@sampottinger Thanks for addressing that feedback so quickly! While we wait on the second review, you may want to look into the DOI comments raised by the bot [here](https://github.com/openjournals/joss-reviews/issues/10104#issuecomment-3945183253). Please review the "MISSING DOIs" section and also double check if there are publications/archives for any of the packages you cite that currently are missing a DOI ("SKIP DOIs" section).
