### 2026-03-13 SchmidtDSE/kigali-sim#765 @sampottinger
Check missing DOIs
Thanks for your help @willgearty! Related to https://github.com/openjournals/joss-reviews/issues/10104, I am just taking another look at the missing DOIs (see below). Some of these likely have a DOI through the UN library though others (like open source projects) may be more variable. I’ve started this process but just formalizing an issue to capture this for tying to an incoming PR.

- No DOI given, and none found for title: The Montreal Protocol: triumph by treaty
- No DOI given, and none found for title: Scientific Assessment of Ozone Depletion: 2022
- No DOI given, and none found for title: HyperTalk: The Language for the Rest of Us
- No DOI given, and none found for title: Using Game Design to Inform a Plastics Treaty: Fos...
- No DOI given, and none found for title: Core Loops in Gameplay
- No DOI given, and none found for title: How Nintendo Solved Zelda’s Open World Problem
- No DOI given, and none found for title: The Structure of Fun: Learning from Super Mario 3D...
- No DOI given, and none found for title: Super Mario 3D World’s 4 Step Level Design
- No DOI given, and none found for title: Programming Portals
- No DOI given, and none found for title: Why DSLs? A Collection of Anecdotes
- No DOI given, and none found for title: Domain-Specific Languages (DSLs)
- No DOI given, and none found for title: ANTLR (ANother Tool for Language Recognition)
- No DOI given, and none found for title: Ace: The High Performance Code Editor for the Web
- No DOI given, and none found for title: Apache Commons CSV
- No DOI given, and none found for title: Chart.js: Simple yet flexible JavaScript charting ...
- No DOI given, and none found for title: D3.js: Data-Driven Documents
- No DOI given, and none found for title: Papa Parse: Powerful CSV Parser for JavaScript
- No DOI given, and none found for title: Prism: Lightweight, extensible syntax highlighter
- No DOI given, and none found for title: QUnit: A JavaScript Unit Testing Framework
- No DOI given, and none found for title: Tabby: Lightweight JavaScript plugin for managing ...
- No DOI given, and none found for title: Webpack: A static module bundler for modern JavaSc...
- No DOI given, and none found for title: TeaVM: Build fast, modern web applications in Java
- No DOI given, and none found for title: Policies, Procedures, Guidelines and Criteria of t...
- No DOI given, and none found for title: Jgraph/Drawio: Draw.io is a JavaScript, client-sid...
- No DOI given, and none found for title: Claude
- No DOI given, and none found for title: GitHub Copilot
- No DOI given, and none found for title: JetBrains AI Assistant
- No DOI given, and none found for title: Replit AI Agent
- No DOI given, and none found for title: Preparation of Kigali HFC Implementation Plans
- No DOI given, and none found for title: National Ozone Officers’ Capacity Building
- No DOI given, and none found for title: Montreal Protocol on Substances that Deplete the O...
- No DOI given, and none found for title: The /llms.txt file
- No DOI given, and none found for title: Colorbrewer 2.0
- No DOI given, and none found for title: Kigali Sim Credits
- No DOI given, and none found for title: HFC Modelling Agenda for the 46th Meeting of the O...
- No DOI given, and none found for title: Report of the Ninety-Seventh Meeting of the Execut...
- No DOI given, and none found for title: List of Participants: South East Asia and the Paci...
- No DOI given, and none found for title: Final Agenda: Pacific Island Countries Ozone Offic...
- No DOI given, and none found for title: UC Berkeley offers tech to United Nations to fight...
- No DOI given, and none found for title: HFC Outlook: Description and Specification
