### 2026-03-13 #10104 @sampottinger
Our appreciation @willgearty for helping further improve the paper! This is much stronger with these changes and thank you for your help. Updated proof available in thread, resolving https://github.com/SchmidtDSE/kigali-sim/issues/765.
