### 2026-03-16 3ed5b0e
Clarify broad applicability of the tool.
Related to #761, responding to @lionel42 feedback on broad
applicability of tool to halocarbons and further clarifying HFCs /
Kigali Amendment in paper.
