+++ paper/paper.md
@@ -30,16 +30,16 @@ bibliography: paper.bib
 ---
 
 # Summary
-Kigali Sim offers stock and flow modeling of Montreal Protocol-controlled substances. This includes hydrofluorocarbons that contribute to climate change through warming potentials much greater than carbon dioxide. Supporting a diverse community with heterogeneous programming expertise, this parallelized toolkit democratizes advanced computational tools by interoperating visual no-code rapid model development and a domain-specific language (DSL). Either through WebAssembly (WASM) or Java Virtual Machine (JVM), Kigali Sim affords portable, private, and rigorous simulation in support of history's most successful international environmental treaty, aiding both atmospheric science and policy development.
+Kigali Sim offers stock and flow modeling of Montreal Protocol-controlled substances. This includes hydrofluorocarbons at the center of the Kigali Amendment. These potent substances contribute to climate change through warming potentials much greater than carbon dioxide. Supporting a diverse community with heterogeneous programming expertise, this parallelized toolkit democratizes advanced computational tools by interoperating visual no-code rapid model development and a domain-specific language (DSL). Either through WebAssembly (WASM) or Java Virtual Machine (JVM), Kigali Sim affords portable, private, and rigorous simulation in support of history's most successful international environmental treaty, aiding both atmospheric science and policy development.
 
 # Statement of Need
-Signed by all UN member states, the Montreal Protocol phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework to hydrofluorocarbons which contribute significantly to climate change [@kigali_amendment; @contribute]. However, research and policy analysis for these substances involves modeling complex economic, technological, energy, and policy interactions [@complex].
+Signed by all UN member states, the Montreal Protocol phased out 99% of ozone-depleting substances [@ozone]. Already the most successful international environmental treaty [@montreal_protocol_success], its ambitious 2016 Kigali Amendment extends this multilateral framework and related international funding to hydrofluorocarbons (HFC), potent substances which contribute significantly to climate change [@kigali_amendment; @contribute]. However, research and policy analysis for these substances involves modeling complex economic, technological, energy, and policy interactions [@complex].
 
 ## State of field
 On the current public market, only the proprietary HFC Outlook offers a reusable full lifecycle model for the Montreal Protocol [@hfcoutlook]. Otherwise, organizations typically build private ad-hoc models, a multi-dataset process incorporating treaty-specific conventions [@complex].
 
 ## Contribution
-Kigali Sim provides the first open source reusable lifecycle modeling toolkit focused on Montreal Protocol-controlled substances. It emerged after the Multilateral Fund for the Implementation of the Montreal Protocol (MLF) met the University of California Schmidt Center for Data Science and Environment (DSE) during a 2023 presentation of an interactive simulation tool developed by DSE and their partners for a global plastics pollution treaty [@plasticssci]. From that initial conversation, co-design sessions in 2024 sought to consider game design-inspired techniques for participatory data science similar to that prior project [@plasticsgame] but within the Montreal Protocol's science and policy context. Kigali Sim's design ultimately focuses on reducing barriers to entry in modeling emissions, energy, consumption, equipment populations, trade, and policy while following treaty conventions [@mlf_guidelines]. Finally, Kigali Sim was implemented by DSE in collaboration with MLF as well as users from various governments and supporting organizations [@credits].
[…]
