### 2026-03-16 1932af3
Clarify broad applicability of the tool.
Related to #761, responding to @lionel42 feedback on broad
applicability of tool to halocarbons and further clarifying HFCs /
Kigali Amendment in paper.
