### 2026-03-16 a969741
Clarify BigDecimal and ANTLR
Clarifying why BigDecimal was used specifically and discuss the need to
support numbers at wide scales. Additionally, clarify the role of
ANTLR. Related to #761 in response to @lionel42 feedback.
