+++ paper/paper.md
@@ -45,7 +45,7 @@ Kigali Sim provides the first open source reusable lifecycle modeling toolkit fo
 Kigali Sim serves Article 5 nations [@article5], Implementing Agencies, analysts, researchers, and related organizations. This includes National Ozone Units (NOUs) with limited resources [@noo] for whom developing ad-hoc models may be burdensome. Though usage remains anonymized, more than a dozen nations and supporting organizations co-designed Kigali Sim over more than a year. This effort spanned scientists, analysts, and policy-makers. Participating in many international meetings [@oweg; @excom; @pacificnetwork; @seasianetwork], multiple governments publicly acknowledge contributing to this community project [@credits]. This open source effort also received media coverage [@nbc].
 
 # Implementation
-Migrated from an original JavaScript implementation for performance and portability, Kigali Sim runs via JVM or browser-based WASM through TeaVM [@teavm]. Without transmitting simulations to external servers, both modalities enable privacy-preserving parallel local computation with BigDecimal [@bigdecimal] after ANTLR [@antlr] interpretation.
+Migrated from an original JavaScript implementation for performance and portability, Kigali Sim runs via JVM or browser-based WASM through TeaVM [@teavm]. Without transmitting simulations to external servers, both modalities enable privacy-preserving parallel local computation. We use BigDecimal [@bigdecimal] to support the wide numeric scales frequently encountered in simulations, often through unit conversions leveraging equipment properties. Finally, ANTLR provides DSL interpretation [@antlr].
 
 ![Diagram describing multi-modal execution in which simulations run across different platforms with simulation results displayed via a web browser.\label{fig:execution}](KigaliExecution.svg){width="100%"}
 
