### 2026-03-16 d6df8e2
Clarify engine graph traversal.
As part of #761, responding to helpful @lionel42 feedback to make the
graph traversal mechanism more clear.
