### 2026-03-16 5b5f325
Clarify engine graph traversal.
As part of #761, responding to helpful @lionel42 feedback to make the
graph traversal mechanism more clear.
