+++ paper/paper.bib
@@ -477,3 +477,29 @@ url="https://doi.org/10.1007/979-8-8688-0351-2_12"
   url={https://www.gluckmanconsulting.com/wp-content/uploads/2023/11/HFC-Outlook-Description-and-Specification.pdf},
   note={HFC modeling tool specification}
 }
+
+@incollection{article3,
+  title     = "{No. 26369. Montreal Protocol on Substances that Deplete
+                the Ozone Layer. Concluded at Montreal on 16 September 1987}",
+  author    = "{United Nations}",
+  booktitle = "United Nations Treaty Series",
+  volume    = {1522},
+  publisher = "United Nations Publications",
+  year      = {1989},
+  doi       = {10.18356/e497d068-en-fr},
+  url       = {https://www.un-ilibrary.org/content/books/9789210453806s002-c013}
+}
+
+@techreport{decisioni12a,
+  title       = "{Decision I/12A}: Clarification of Terms and Definitions:
+                  Controlled Substances (in Bulk)",
+  author      = "{First Meeting of the Parties to the Montreal Protocol}",
+  institution = "United Nations Environment Programme",
+  type        = "Report of the First Meeting of the Parties",
+  number      = "UNEP/OzL.Pro.1/5",
+  year        = {1989},
+  url         = {https://ozone.unep.org/treaties/montreal-protocol/meetings/first-meeting-parties/decisions/decision-i12a-clarification-terms-and-definitions-controlled-substances-bulk},
+  note        = "No DOI assigned; official UN document symbol UNEP/OzL.Pro.1/5.
+                  Clarifies that controlled substances in pre-charged equipment
+                  are attributed to the exporting (manufacturing) country"
+}
