### 2026-03-16 4c68492
Simplification of DSL introduction.
As part of #761, incorporating @lionel42 suggested improvements in
introducing the DSL.
