### 2026-03-16 4507642
Simplification of DSL introduction.
As part of #761, incorporating @lionel42 suggested improvements in
introducing the DSL.
