### 2026-03-16 acce30b
Fix typo in paper.
As part of #761, fix a typo identified by @lionel42.
