### 2026-03-16 a7e7b61
Fix typo in paper.
As part of #761, fix a typo identified by @lionel42.
