### 2026-03-16 cb8ea21
Clarification in limitations and future work
As part of #761, incorporating suggestions in limitations and future
work from @lionel42.
