+++ paper/paper.md
@@ -78,7 +78,7 @@ Given the goal to introduce the user loops [@core_loops], this sequence only run
 Many Kigali Sim users do not identify as programmers. Prior empirical work suggests that domain experts with "limited programming knowledge" benefit from DSLs [@dsl_evaluation]. Therefore, we created the QubecTalk domain-specific language to facilitate expression of complex models in human-readable syntax. This DSL is inspired by HyperTalk [@smallhypertalk]. Mirrored by the UI-editor, QubecTalk speaks in treaty terminology, translating terms of art into simulations for improved usability [@dsl_comprehension]. This also supports uncertainty quantification, conditional logic, and policy stacking. With optional AI assistant compatibility via the llms.txt specification [@llmstxt], users may author scripts in a web-based programming portal [@portal] or outside with direct JVM invocation.
 
 ## Limitations and future work
-Kigali Sim focuses only on the Montreal Protocol. First, it can model energy consumption but only estimates direct emissions. Users may use Kigali Sim to calculate indirect emissions by exporting from the tool to join with outside energy mix data [@energy_mix]. Additionally, it embodies Montreal Protocol definitions such as trade attribution rules and it will receive updates as official guidance changes in the future [@mlf_guidelines]. Therefore, substances not subject to Montreal Protocol conventions are out of scope.
+Kigali Sim focuses only on the Montreal Protocol. First, it can model energy consumption but only estimates direct emissions. Users may use Kigali Sim to calculate indirect emissions by exporting consumption data from the tool to join with outside energy mix data [@energy_mix]. Additionally, it embodies Montreal Protocol definitions such as trade attribution rules [@article3; @decisioni12a] and it will receive updates as official guidance changes in the future [@mlf_guidelines]. Therefore, substances not subject to Montreal Protocol conventions are out of scope.
 
 # Acknowledgments
 BSD-licensed.
