### 2026-03-16 d5cbfcd
Clarification in limitations and future work
As part of #761, incorporating suggestions in limitations and future
work from @lionel42.
