### 2026-03-16 919e565
Wording improvement in dual interface design
As part of #761, incorporating suggestions from @lionel42 in dual
interface design introduction.
