### 2026-03-18 SchmidtDSE/kigali-sim#761 @lionel42
@sampottinger Do you simply feed my comments to the AI or do you acutally think about what I write ?
