### 2026-03-18 SchmidtDSE/kigali-sim#761 @willgearty
Reading over this back-and-forth, I can see how @sampottinger's tendency towards being very polite and professional could maybe come across as AI-generated (unfortunately this is the world we now live in). However, these kinds of responses are very familiar to me in academic publishing, where being mean or impolite to a reviewer could mean the rejection of your paper. I don't see any indicators here that @sampottinger is misleading us, and it seems he genuinely cares about the feedback @lionel42 has provided and wants to make sure he has addressed those concerns satisfactorily.
