### 2026-03-19 857ef8e
Respond to #761 comments on Hayashida.
Responding to continued comments regarding the Hayashida sequence as mentioned by @lionel42 in #761. This removes pyafscgap and Brown while changing to mention Kishōtenketsu instead of Hayashida to name the pattern.
