+++ paper/paper.bib
@@ -1,5 +1,3 @@
-@article{pyafscgap, doi = {10.21105/joss.05593}, url = {https://doi.org/10.21105/joss.05593}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {86}, pages = {5593}, author = {Pottinger, A Samuel and Zarpellon, Giulia}, title = {Pyafscgap.org: Open source multi-modal Python-based tools for NOAA AFSC RACE GAP}, journal = {Journal of Open Source Software} }
-
 @article{montreal_protocol_success,
   title     = "Triumph by treaty",
   author    = "Molina, Mario and Zaelke, Durwood",
@@ -146,14 +144,6 @@
   year={2012},
   url={https://www.gamedeveloper.com/design/the-structure-of-fun-learning-from-i-super-mario-3d-land-i-s-director},
   note={Interview article. Game Developers Conference 2012}
-}
-
-@misc{hayashida_video,
-  title={Super Mario 3D World's 4 Step Level Design},
-  author={Brown, Mark},
-  year={2015},
-  url={https://youtu.be/dBmIkEvEBtA},
-  publisher={Game Maker's Toolkit}
 }
 
 @misc{portal,
