### 2026-03-19 c87e7a6
Remove BigDecimal and first ANTLR citation.
In response to ongoing discussion in #761 from @lionel42, removed the first reference to ANTLR (still mentioned in acknowledgements) and cut out the reference to BigDecimal as that intention still comes through in our source.
