+++ paper/paper.bib
@@ -285,20 +285,6 @@
   note={Apache License 2.0}
 }
 
-@Inbook{bigdecimal,
-author="Friesen, Jeff",
-title="Math, BigDecimal, and BigInteger",
-bookTitle="Learn Java Fundamentals: A Primer for Java Development and Programming",
-year="2024",
-publisher="Apress",
-address="Berkeley, CA",
-pages="311--324",
-abstract="Mathematics operations are an important part of computing. You learned about basic math operators (such as addition and multiplication) in Chapter 3. Java also provides a Math class in the java.lang package, along with BigDecimal and BigInteger classes in the java.math package. This chapter introduces you to these classes.",
-isbn="979-8-8688-0351-2",
-doi="10.1007/979-8-8688-0351-2_12",
-url="https://doi.org/10.1007/979-8-8688-0351-2_12"
-}
-
 @misc{mlf_guidelines,
   title={Policies, Procedures, Guidelines and Criteria of the Multilateral Fund for the Implementation of the Montreal Protocol},
   author={{Multilateral Fund Secretariat}},
