### 2026-03-20 #10104 @lionel42
## Review questions and comments

This software is a very impressive work. Not only is it very useful but it also has a huge code base and an extensive documentation with useful tutorials. Congratulations to the authors.

I have sadly not have time to go through the whole code, it is really huge and java is a very verbose language. I have check few parts for which I was interested. The structure is well designed and the code is clear and well commented.

I have few questions left and general remarks about the SW for the authors.

### Technical feedback

Why not using Enums for units or steams in the code ?
They usually help in terms of readability and performance.

The frontend (editor) does not seem to have an automated API doc generation. Am I missing something or did you only include automatic doc generation for the backend ?

In the CI you have a MonteCarlo test step `runMonteCarloReplicates`. As far as I understand, this is a test. Why did you put it directly as a step in the CI instead of including it in the existing test pipeline that you have ?

In the docs, you have one markdown and one html folder.
Usually packages have simply a markdown version and they automatically generate html from it using tools like sphinx or mkdocs. As I understand you always have to update the 2 versions separately.
Why did you not follow a automatic generation approach ?
It can be very cumbersome to have 2 versions of the same thing.

Also related to the docs: I have seen that in the repo the videos are given as .gif format, but in the online tutorial they are present as `.webm` files.
I tried to find where and how the conversion is done but could not find any mention of it. Can you let me know how this is done ?

### Further improvements suggestions

(Just ideas to discuss for the future)

Do you already have or plan to implement a "gallery" of simulations ? As a new user for example i would be happy to see how a specific country have developed very complex real life scenarios (example, as a country simulating all substances for all applications for which they have to report these usages) using kigalisim.

You mention at some places in the documentation (like in this tutorial (https://kigalisim.org/guide/tutorial_16.html)) that people can use the output of the simulation to do further analysis. Do you already know of such examples ? Any plans to include such examples as part of kigalisim in the future ? Or do you want to further improve your frontend do be able to support these more complex simulations ?
