### 2026-03-24 #10104 @sampottinger
Hello @lionel42! Thank you very much and please see below. I’ll link issues shortly.

> The frontend (editor) does not seem to have an automated API doc generation. Am I missing something or did you only include automatic doc generation for the backend ?

I apologize that this was confusing. You are correct: currently we only provide automatic HTML generation from structured comments (javadoc) for the Java elements. However, we don’t see any downside to making those resources from the front-end available as well! See https://github.com/SchmidtDSE/kigali-sim/issues/774.

> In the CI you have a MonteCarlo test step `runMonteCarloReplicates`. As far as I understand, this is a test. Why did you put it directly as a step in the CI instead of including it in the existing test pipeline that you have ?

This was motivated by the Monte Carlo step In CI / CD taking a bit of time potentially and also observation that knowing of its failure may be less useful if the simpler (single replicate) tests are already broken. So, we wanted to gate that computational cost in GitHub Actions, running it only if the earlier less computationally intensive steps passed. Please let me know if you feel an additional action is needed here!

> In the docs, you have one markdown and one html folder. Usually packages have simply a markdown version and they automatically generate html from it using tools like sphinx or mkdocs. As I understand you always have to update the 2 versions separately. Why did you not follow a automatic generation approach ? It can be very cumbersome to have 2 versions of the same thing.

This confusion is understandable and thanks for highlighting. We are using some HTML features extensively in the tutorials which are not well supported in markdown like `<summary>` and `<video>`. The markdown versions really are more there to support llms.txt (to save assistant context for users taking advantage of that technology). In practice, it may actually make more sense for the markdown versions to be generated from HTML. We agree with you that this might be a good opportunity to take that step. See https://github.com/SchmidtDSE/kigali-sim/issues/775.

> Also related to the docs: I have seen that in the repo the videos are given as .gif format, but in the online tutorial they are present as `.webm` files. I tried to find where and how the conversion is done but could not find any mention of it. Can you let me know how this is done ?

You can see how the project has evolved here 😅. The new webm files replaced the old gifs.

The superior quality of the webm files has been useful as are the accessibility features afforded by the `<video>` tag (pause, rewind, etc). All that said, those files were uploaded directly. I’m a little mixed on including them in to the repo itself just in service of avoiding larger binary blobs in git. I think we should absolutely remove the gifs. However, I’m curious your perspective @lionel42 on the conflicting advice regarding exclusion of binary blobs from git vs including them to keep the project more self-contained within the repository. Thanks in advance for your thoughts!

EDIT, see https://github.com/SchmidtDSE/kigali-sim/issues/776?reload=1

> Why not using Enums for units or steams in the code? They usually help in terms of readability and performance.

Thanks for this!

Early on, we chose not to go this route because Kigali Sim allows for combinations (like `kg / yr`) of units. That said, handling dimensional analysis was a bit more diffuse / complex in earlier versions of Kigali Sim. Architecturally, things have consolidated since then. Finally, [another project that used some of Kigali Sim’s code as a jumping off point](https://github.com/SchmidtDSE/josh) also had some innovations here.

Let me look more closely and open an issue. I’d like to engage a little conversation with colleagues just to get some advice (I believe, counter-intuitively, early profiling was also showing a switch to enum wasn’t actually saving too many cycles… but it has been some time). However, I’ll respond as soon as I can!

https://github.com/SchmidtDSE/kigali-sim/issues/777

> Do you already have or plan to implement a "gallery" of simulations ? As a new user for example i would be happy to see how a specific country have developed very complex real life scenarios (example, as a country simulating all substances for all applications for which they have to report these usages) using kigalisim.

This is a fantastic suggestion and one that has come up in other contexts as well (a bilateral just today was considering a related question!). Even so, there are some considerations regarding data privacy. Many or even most users of Kigali Sim work with sensitive country-specific data and, depending on the policy of those countries, those always get disclosed to the UN but only sometimes get released in full publicly.

There are certainly folks more on the open side of things. For those doing the public release, I think it could be quite valuable! All that in mind, I think we are still trying to work with our users to see the appetite for this feature relative to the work in maintaining it. Having a discussion thread open for this in an issue is fantastic and I’ll make that shortly. However, if you find it agreeable, we may politely decline to resolve that in the timeframe of the JOSS review if that’s ok.

> You mention at some places in the documentation (like in this tutorial ([https://kigalisim.org/guide/tutorial_16.html](https://kigalisim.org/guide/tutorial_16.html))) that people can use the output of the simulation to do further analysis. Do you already know of such examples ? Any plans to include such examples as part of kigalisim in the future ? Or do you want to further improve your frontend do be able to support these more complex simulations ?

From our engagement with users, the most common case is indirect emissions. However, this particular use case gets a little dicey with IPCC tiers. So far, we’ve been a little reluctant to wade into that just because we feel Kigali Sim is doing a good job at guiding users to analysis compliant with the Montreal Protocol methodologies through a structured DSL / UI flow but offering guidance on that other step just requires a lot of instruction on additional consideration that we worry is a little outside our purview. At this time, if you find it agreeable, we might politely decline to expand out into that territory if that’s ok. It may be something we feel more confident addressing if we get additional maintainers (or funding!). We would value your feedback on this point though!
