### 2026-03-24 #10104 @lionel42
Thanks for you answer @sampottinger . I agree with all your answers.

Please let me know when you are done with the documentations changes discussed, they are the only ones required in the context of this review.

Also, here are my answers to some of your comments.

> This was motivated by the Monte Carlo step In CI / CD taking a bit of time potentially and also observation that knowing of its failure may be less useful if the simpler (single replicate) tests are already broken. So, we wanted to gate that computational cost in GitHub Actions, running it only if the earlier less computationally intensive steps passed. Please let me know if you feel an additional action is needed here!

Not mandatory, but what you could do here is (at least it is what I do in some python test suites) is to mark some tests as "expensive" or similar. Then when you run the test case, you can set a flag to run only the non expensive tests or only the expensive ones. This could be useful in the future if you want to add more tests. In the context of this review the current solution is enough.

> However, I’m curious your perspective @lionel42 on the conflicting advice regarding exclusion of binary blobs from git vs including them to keep the project more self-contained within the repository. Thanks in advance for your thoughts!

I agree that binary blobs should not go on a github repository. However in the context of open source, it would still be good to give full access to the necessary resources. I would suggest an external cloud storage to host the `webm` files. Then when the doc is built, an automatic download from this .
