### 2026-03-25 SchmidtDSE/montreal-protocol#774 @sampottinger
Add JSDoc generation
Part of https://github.com/openjournals/joss-reviews/issues/10104, we currently have javadoc generation and publication. We have JSDoc compliant comments but we are not currently making those available in a similar HTML format. This could happen through a few avenues but we might consider [Docma](https://onury.io/docma/). This would likely need to split into another GitHub Action given the new dependencies introduced. However, it could export / publish through a pathway similar to the current JavaDoc generation. CC @lionel42
