### 2026-03-25 SchmidtDSE/kigali-sim#775 @sampottinger
Consolidate HTML and md in user’s guide
Part of https://github.com/openjournals/joss-reviews/issues/10104. We have dual HTML and markdown documents in the user’s guide and supporting docs. The HTML versions are required due to some tags which, while improving user experience, can’t be expressed in markdown. This includes `<video>` and `<summary>` which are used extensively. It may be best to consolidate these two parallel sets of documents but it may also be best to have md generated from HTML. Note that the md is largely in service of llms.txt. CC @lionel42
