### 2026-03-27 828d0df
Add JSDoc generation for ECMAScript editor (closes #774)
Adds jsdoc as a devDependency, a jsdoc.json config targeting editor/js/,
a new generateJsdoc CI job mirroring the existing generateJavadoc pattern,
and JSDoc links alongside JavaDoc references across README, CONTRIBUTING,
DEVELOPING, the user guide, and the QubecTalk spec.

Co-authored-by: Claude <claude@anthropic.com>

https://claude.ai/code/session_01Kim1yXnwNXArxA63RajKke
