+++ docs/guide/md/tutorial_04a.md
@@ -1,114 +1,0 @@
-# Tutorial 4a: Growth and Trade with AI
-
-Adding economic growth projections and import flows to create more nuanced business-as-usual scenarios, aided by AI.
-
-**Note that this is the AI version of this tutorial.** You can find a version of this tutorial without AI at [Tutorial 4](/guide/md/tutorial_04.md).
-
-## Motivation
-Let's say ABC has economic growth forecasts and data on equipment imports. In this tutorial, we will add growth patterns and import flows to our multi-application model from Tutorial 3, creating a more nuanced business-as-usual baseline to help aid policy analysis. In addition to adding more sophistication to our model, this also demonstrates use of multiple sales streams.
-
-As before, all of this can take quite a bit of clicking. So, we will see if AI can help us out to speed up the process! This can show how AI can help us not just add new things to our simulations but modify our existing work as well.
-
-## Preparing for AI
-If you did the AI version of Tutorial 3, you probably are already familiar with this process. However, just to cover everything, let's step through this again. Most AI assistants can help with Kigali Sim using a standard called [llms.txt](https://llmstxt.org/).
-
-**If you did Tutorial 3 with AI** and the output from your assistant looked good, just continue your same chat session! Continue with [Adding imports for HFC-134a](#adding-imports-for-hfc-134a) section.
-
-**If you are new to the AI** or just want to start fresh, create a new chat session. Then, tell it to look up information about Kigali Sim through a message like this:
-
-```
-Hello! I would like help with building a Kigali Sim simulation. Please read https://kigalisim.org/llms-full.txt?v=20260302 to learn more. Please stick to only features compatible with the UI editor.
-```
-
-Need a little cheat? [Download the Tutorial 3](../tutorial_03.qta) file here. Also, for more capable assistants like Claude, this is enough. However, some assistants cannot access the full internet or won't know how to work with this kind of file out of the box. If your assistant is having issues, instead, attach the [Kigali Sim llms-full.txt](https://kigalisim.org/llms-full.txt?v=20260302) file as an attachment!
-
-When your assistant is ready, next give it your Tutorial 3 file. This can be done by going to the Save File button at the top of the Kigali Sim application and adding it to the chat as an attachment with a message like this:
-
-```
-Great! The attached is the simulation I started working on.
-```
-
-**Note**: More details and troubleshooting steps specific to individual AI assistants are available in [Tutorial 11](/guide/md/tutorial_11.md).
-
-## Adding imports for HFC-134a
-Before we add in growth rates, let's consider trade. Specifically, for brevity, let's have just one substance with imports.
-
-Let's say that ABC imports some but not all of their HFC-134a. Therefore, for **Domestic Refrigeration**, modify your **HFC-134a** consumption record with a prompt like this:
-
-```
-Please modify HFC-134a to enable imports in addition to the domestic manufacture already present. For charge, let's use 0.2 kg / unit for import. Then, let's reduce domestic to 13 mt / year for this substance in 2025. However, also add 11 mt / year in year 2025. Thanks!
-```
-
-When the AI is done, check its work by downloading the file it creates and using the Load File button in Kigali Sim. You should see changes when you click "edit" next to HFC-134a (domestic refrigeration).
-
-## First Economic Growth
-In addition to trade, let's also add in economic growth. For example, these projections might come from industry surveys or from outside modeling efforts. Let's continue with HFC-134a:
-
-```
-Great! Let's continue with HFC-134a. Next, please expect growth +6% of current for all sales (domestic manufacture and import) from 2025 to 2030. Then, have this decrease to +4% of current for all sales from 2031 to 2035.
-```
-
[…]
