+++ docs/guide/md/tutorial_09.md
@@ -1,158 +1,0 @@
-# Tutorial 9: Global Warming Potential Refrigerant Comparison
-
-Demonstrating GWP impact reduction through HFC-134a to R-600a substitution.
-
-## Motivation
-
-Global Warming Potential (GWP) values represent how much a substance contributes to climate change compared to CO2. For example, consider HFC-134a with a GWP of 1,430 kgCO2e / kg versus R-600a which has only 3. This means that each kilogram of HFC-134a has the same climate impact as 477 kilograms of R-600a.
-
-Previous tutorials included investigation of modeling these values but this section takes a closer look at specifying tCO2e for substances, modeling both with and without different policies. More specifically, we'll look at a refrigerant substitution policy that gradually replaces HFC-134a with R-600a in domestic refrigeration. This simulation can track this transition as it dramatically reduces overall climate impact (measured in tCO2e) even when total refrigerant consumption remains similar.
-
-Note that instructions are a bit more brief under the assumption that the earlier tutorials have been completed.
-
-## Setting Up the Business-as-Usual Scenario
-
-In this tutorial, we will move away from our ABC Country case study. If you are continuing from [Tutorial 8](/guide/md/tutorial_08.md), go ahead and click **New File** at the top of the web editor. This will give you a fresh simulation.
-
-First, let's create our baseline scenario with both HFC-134a and R-600a refrigerants in domestic refrigeration. We'll start with HFC-134a as the dominant refrigerant and minimal R-600a consumption.
-
-**Note:** In this feature specific tutorial, we will start from scratch each time. Click **Save File** to save your current work and then click **New File** to start a simulation.
-
-**Step 1: Create the Domestic Refrigeration application**
-- Click **Add Application**
-- Name it "Domestic Refrigeration"
-- Click **Finish**
-
-**Step 2: Add HFC-134a substance**
-- In your Domestic Refrigeration application, click **Add Consumption**
-- On the **General** tab:
-  - Name the substance "HFC-134a" without quotes.
-  - Use Domestic Refrigeration as the **Application**.
-  - Set **GHG equivalency** to 1430 kgCO2e/kg.
-  - Leave **Annual energy consumption** alone for now.
-  - Enable **domestic** manufacture.
-- On the **Equipment** tab:
-  - We will leave **Equipment type** alone for now.
-  - Set **initial charge** to 0.15 kg/unit for domestic.
-  - Set **annual retirement** rate to 5% each year.
-- On the **Servicing** tab:
-  - Set **recharge** to 10% with 0.15 kg/unit in all years
-- On the **Set** tab:
-  - Set **prior equipment** to 1,000,000.0 units in year 2025.
-  - Set **domestic** manufacture to 20 mt in year 2025.
-- Click **Finish**
-
-**Step 3: Add R-600a substance**
-- Again in the Domestic Refrigeration application, **Add Consumption** for R-600a with similar equipment properties except different GWP.
-- On the **General** tab:
-  - Name the substance "R-600a" without quotes.
-  - Use Domestic Refrigeration as the **Application**.
-  - Set **GHG equivalency** to 3 kgCO2e/kg.
-  - Leave **Annual energy consumption** alone for now.
-  - Enable **domestic** manufacture.
-- On the **Equipment** tab:
-  - We will leave **Equipment type** alone for now.
-  - Set **initial charge** to 0.15 kg/unit for domestic.
-  - Set **annual retirement** rate to 5% each year.
-- On the **Servicing** tab:
-  - Set **recharge** to 10% with 0.15 kg/unit in all years.
-- On the **Set** tab:
-  - Set **prior equipment** to 50,000.0 units in year 2025.
-  - Set **domestic** manufacture to 1 mt in year 2025.
-- Click **Finish**
-
-**Step 4: Create baseline simulation**
[…]
