### 2026-03-27 5d07f9b
Generate guide markdown from HTML in CI rather than maintaining dual files
Remove hand-written docs/guide/md/ files and replace with a new
generateGuideMd CI job that converts docs/guide/*.html to GFM markdown
using Pandoc (invoked as executable, consistent with existing spec PDF
generation). The generated md/ directory is downloaded as an artifact
into deployPrep before the guide copy step, so llms.txt paths remain
valid on the hosted site. docs/guide/md/ is now .gitignored.

Closes #775

https://claude.ai/code/session_0171SWXZqKkV8TCPZ3SRfWRa

Co-authored-by: Claude <claude@anthropic.com>
