+++ docs/guide/md/qubectalk_commands.md
@@ -1,307 +1,0 @@
-# QubecTalk Reference: Commands
-
-Commands are the executable statements within QubecTalk stanzas that define substance properties, set values, and configure policies.
-
-## Contents
-
-- [Assume](#assume)
-- [Cap](#cap)
-- [Change](#change)
-- [Define](#define)
-- [Enable](#enable)
-- [Equals (GWP)](#equals-gwp)
-- [Floor](#floor)
-- [Get](#get)
-- [Initial Charge](#initial-charge)
-- [Recharge](#recharge)
-- [Recover](#recover)
-- [Replace](#replace)
-- [Retire](#retire)
-- [Set](#set)
-
-## Assume
-
-**Purpose:** Controls sales carryover behavior, particularly useful for bank tracking scenarios.
-
-**Syntax:** `assume [no|only recharge|continued] streamName during years startYear to endYear`
-
-**Available Streams:** domestic, import, export, sales, bank, equipment
-
-**Examples:**
-```qubectalk
-assume no domestic during year 2030
-assume only recharge sales during years 2025 to onwards
-assume continued import
-assume only recharge bank during years 3 to onwards
-```
-
-**Behavior:**
-- `assume no` - Sets stream to 0 kg (zero all sales)
-- `assume only recharge` - Sales only cover recharge needs, no new equipment
-- `assume continued` - Continue from previous year (default behavior)
-
-## Cap
-
-**Purpose:** Limits consumption to specified levels, optionally with displacement to alternative substances.
-
-**Syntax:**
-```
-cap streamName to amount during years startYear to endYear
-cap streamName to amount displacing "substanceName"|stream during years startYear to endYear
-cap streamName to amount displacing by volume "substanceName"|stream during years startYear to endYear
-cap streamName to amount displacing by units "substanceName"|stream during years startYear to endYear
-```
-
-**Examples:**
-```qubectalk
-cap sales to 80% during years 2027 to 2030
-cap domestic to 0 mt displacing "R-600a" during years 2031 to onwards
-cap import to 50% during years 2028 to 2032
-```
-
-### Displacement Types
-
-When using displacement, you can control how the displaced demand is calculated. Displacement always operates in the target stream's current units (units or kg) without changing its tracking mode:
-
-- **displacing** means equivalent displacement (default). This maintains the same quantity in the units last specified. If last value was in kg, displaces by kg. If last value was in units, displaces by units.
-- **displacing by volume** means volume-based displacement. This always displaces by substance mass (kg), converting from units if necessary using initial charge values. May result in different equipment counts. In other words, this assumes 1 kg of the original substance is displaced by 1 kg of the new substance.
-- **displacing by units** means units-based displacement. This always displaces by equipment units, converting from kg if necessary using initial charge values. Maintains equipment population. In other words, this assumes 1 unit of equipment using the original substance is displaced by 1 unit of equipment using the new substance.
-
-**Examples:**
-```qubectalk
-# Equivalent displacement (default behavior)
-cap sales to 80% displacing "R-600a" during years 2027 to 2030
-
-# Volume-based displacement (always use kg, regardless of spec)
-cap sales to 50 units displacing by volume "R-600a" during years 2027 to 2030
-# Result: 50 units × 1 kg/unit = 50 kg displaced to R-600a
-
-# Units-based displacement (always use units, regardless of spec)
-cap sales to 100 kg displacing by units "R-600a" during years 2027 to 2030
-# Result: 100 kg ÷ 1 kg/unit = 100 units displaced to R-600a
-```
-
[…]
