### 2026-04-01 #10104 @sampottinger
Ugh sorry trying to generate PDF again… Per [status page](https://www.githubstatus.com/), GitHub was having issues yesterday but it looks like it might be again today looking at the error message. Will try again in a little bit.
