### 2026-04-01 #10104 @sampottinger
Howdy everyone! Regenerated the paper pdf ([post above](https://github.com/openjournals/joss-reviews/issues/10104#issuecomment-4171821864)) and I think I have now responded to all open comments (the earlier error message seemed to have been caused by a GitHub service outage, see [status page](https://www.githubstatus.com/)).

 - Thanks so much @lionel42 and @willgearty for everything!
 - Please let me know if I can help at all in your review @fredshone. I appreciate your comments in the pre-review and for your time here!
