### 2026-04-14 #10104 @fredshone
huge apologies, had a baby, less free time than normal. @sampottinger, congrats on your project. I've had a somewhat good look and have it running locally. However, web interfaces seem broken - firefox & chrome, with tutorial, readme and examples all showing "simulation error" and failing to render chart. Is this expected or is it me?
