### 2026-04-14 #10104 @sampottinger
Hello @fredshone! First, congratulations and no worries at all on my side. With regards to the web interface, my apologies. Though we weren't impacted by the event, we adjusted how we interact with third party libraries to address future threats similar to CanisterWorm. However, this seems to have introduced a regression earlier this month in CI / CD. We are resolving today and apologize for the inconvenience.
