### 2026-04-15 SchmidtDSE/montreal-protocol#796 @sampottinger
Posted after discussion with @GondekNP at https://github.com/SchmidtDSE/retrospectives/pull/3. Thanks!
