### 2026-04-15 #10104 @sampottinger
Hello again @fredshone! I am very sorry again for this poorly timed disruption. We have resolved the issue and added a full post-mortem / retrospective at https://github.com/SchmidtDSE/kigali-sim/issues/796. Though there was no security risk to our users, this was a problem inadvertently introduced within a broader release earlier this month. This was done during proactive security hardening against supply chain attacks observed recently in the NPM community. You should be able to continue your review.

My sincere apologies but please note that you may need to hard refresh (cmd + shift + r / ctrl + shift + r). We regret the inconvenience. Also, thank you @GondekNP for assistance.

EDIT: [Post-mortem](https://github.com/SchmidtDSE/kigali-sim/issues/796) confirms specific timeline, my apologies again.
