### 2026-04-15 #10104 @fredshone
@willgearty 👍🏻

Some notes:

On first scan I was not a big fan of this project. I originally saw heavy AI use with resulting bloat, complexity and so on. But you've won me over. I think the aspirations and application are great.

I'm not too concerned about how proactive your /agents are being, I believe you've met all the updated "human" JOSS requirements. In fact, I hope you did have some help on all those commits and docs.

There are some decisions that bug me. JS to java to web being one, the performance explanation is unsatisfying. Foremost though, is the absolute genius/madness of trying to replace spreadsheet modelling with a DSL (with a 33 page user pdf). I am generally happy to see the back of excel, but my gut feeling is that your DSL scope is going to creep and creep until your web UI starts to look like excel. I see evidence already in your tests and example configs of trying to manage this complexity. Perhaps try to reign in the scope and focus on driving users with the UI design (rather than docs and docs and docs).

I would prefer your paper explained some of these architectural decisions in boring terms. Rather than, at times, try to sell them.  Specifically, talk of "bidirectional graphs", "democratizing advanced computational tools", "HyperTalk-inspired DSL", and "Kishōtenketsu" reads like LLM filler (please don't be offended if it is not - it's good filler).

I would have been happier, for example, and the review would have been easier, with "we built an accessible open source tool for Montreal Protocol policy analysis, co-designed with a dozen governments, with a readable DSL and dual UI/code interface, and here's evidence it's being used..."

Anyway, these are just grumbles, maybe some of them useful in future. Congratulations @sampottinger and good luck with this project!
