### 2026-04-15 #10104 @sampottinger
Hello again @fredshone! Thank you for your kind words and I’m glad to have had your review.

With regards to your final comments… First, for the DSL, thank you for that advice and I think this is a good perspective for us to hold as we continue onwards in collaboration with our users. It has been interesting to see different parts of the software’s population move between modalities (code vs UI, esp the same users over time). However, your point about creep is well received and will be on our minds in the future. My hope is our continued relationship with our users will help it be more genius than madness in the fullness of time. We appreciate this advice as we navigate that. With regards to the paper, that was me 😅. I’ll take that on board but I’m thankful for the deliberation across authors and reviewers that got us to this form of the manuscript and, as part of that, I am thankful for your read on it. My hope is that it is holding a lot of valued ideas within the word limit. Even so, I’ll take those thoughts into my next project.
