### 2026-04-17 #10104 @willgearty
Thank you @fredshone and @lionel42 for your thorough and thoughtful reviews! I'm glad to hear the constructive feedback has resulted in a better software product (and that there are now many ideas for future development).

I'll now move on to my final checks before I can recommend the paper for acceptance.
