### 2026-04-17 #10104 @willgearty
@sampottinger could you please address all of the items under "Additional Author Tasks" above? In the meantime, I'll start my editorial checks.
