### 2026-04-17 #10104 @willgearty
https://doi.org/10.18356/80909941-en-fr indeed looks like the DOI for the Montreal Protocol. Please update @sampottinger
