### 2026-04-18 #10104 @sampottinger
I am sorry about that @willgearty.

The citation check picked up on where we are citing Articles 3 and 5 of the treaty. However, both articles should be cited after the Kigali Amendment if possible. Unfortunately, I believe a DOI is only available for the original treaty and not the amended version given how the treaty series work.

Using the UNEP library link for both with article name appears to have suppressed the warning from editorialbot. I think this may be the clearer citation but please confirm if you still prefer the original Montreal Protocol DOI and I will adjust.

Thanks and apologies for the confusion!
