### 2026-04-21 #10104 @willgearty
Ah, thanks for fixing that up @sampottinger. Please address the remaining author tasks here: https://github.com/openjournals/joss-reviews/issues/10104#issuecomment-4271183222.
