### 2026-04-23 #10104 @sampottinger
Thank you @willgearty! I believe I have completed the author tasks back over at https://github.com/openjournals/joss-reviews/issues/10104#issuecomment-4271183222. However, please let me know if further modifications are required:

- [x] **Double check authors and affiliations (including ORCIDs)**: All names and affiliations confirmed with each of the co-authors
- [x] **Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper**: We have released [0.1.8](https://github.com/SchmidtDSE/kigali-sim/releases/tag/v0.1.8).
- [x] **Archive the release on Zenodo/figshare/etc and post the DOI here**: We have minted [10.5281/zenodo.19710731](https://doi.org/10.5281/zenodo.19710731).
- [x] **Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper**: Confirmed as indicated below.
- [x] **Make sure that the license listed for the archive is the same as the software license:** Confirmed both use the [BSD-3-Clause license](https://github.com/SchmidtDSE/kigali-sim/blob/main/LICENSE.md).

Maybe not necessarily super helpful (😅) but, nevertheless, just confirming the following:

| Author | Affiliation | ORCID | Paper matches Zenodo |
|---|---|---|---|
| Pottinger, A Samuel | Schmidt Center for Data Science and Environment, University of California, Berkeley, California, United States of America | [0000-0002-0458-4985](https://orcid.org/0000-0002-0458-4985) | Yes |
| Natarajan, Balaji | Secretariat of the Multilateral Fund for the Implementation of the Montreal Protocol, United Nations Environment Programme | No preferred ORCiD | Yes |
| de Bruyn, Magali | Schmidt Center for Data Science and Environment, University of California, Berkeley, California, United States of America | [0009-0000-1609-6244](https://orcid.org/0009-0000-1609-6244) | Yes |
| Martinez, Ciera C. | Schmidt Center for Data Science and Environment, University of California, Berkeley, California, United States of America | [0000-0003-4296-998X](https://orcid.org/0000-0003-4296-998X) | Yes |

EDIT: Linked ORCiDs. Note that @iamciera asked for her addition of her middle initial. This is reflected now in our paper and on Zenodo.
