### 2026-04-23 #10104 @sampottinger
Sorry @willgearty, just confirming we completed the additional author items. See https://github.com/openjournals/joss-reviews/issues/10104#issuecomment-4306594397 for details. Just got a fresh copy for you here from editorial bot if helpful (see above).
