### 2026-04-23 #10104 @sampottinger
Ah sorry about that @willgearty! This is fixed at https://zenodo.org/records/19710732 (DOI 10.5281/zenodo.19710732 version specific, 10.5281/zenodo.19710731 all versions).
